R6 Class Representing a Ticker

Base class for getting all data related to indices from Yahoo Finance API

Format

An R6 class object

Public fields

index

Index for which data is retrieved

Active bindings

summary_detail

Contains information available via the Summary tab in Yahoo Finance

Methods

Public methods

Index$new()

Index$set_index()

Index$get_history()

Index$clone()

Method new()

Create a new Index object

Usage

Index$new(index = NA)

Arguments

index

Index

Returns

A new `Index` object

Examples

nifty_50 <- Index$new('^NSEI')

Method set_index()

Set a new index.

Usage

Index$set_index(index)

Arguments

index

New index

Examples

indice <- Index$new('^NSEI')
indice$set_index('^NDX')

Method get_history()

Retrieves historical data

Usage

Index$get_history(period = "ytd", interval = "1d", start = NULL, end = NULL)

Arguments

period

Length of time

interval

Time between data points

start

Specific starting date

end

Specific ending date

Examples

nifty <- Index$new('^NSEI')
nifty$get_history()
nifty$get_history(start = '2022-07-01', interval = '1d')
nifty$get_history(start = '2022-07-01', end = '2022-07-14', interval = '1d')
nifty$get_history(period = '1mo', interval = '1d')

Method clone()

The objects of this class are cloneable with this method.

Usage

Index$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `Index$new`
## ------------------------------------------------

nifty_50 <- Index$new('^NSEI')

## ------------------------------------------------
## Method `Index$set_index`
## ------------------------------------------------

indice <- Index$new('^NSEI')
indice$set_index('^NDX')

## ------------------------------------------------
## Method `Index$get_history`
## ------------------------------------------------

nifty <- Index$new('^NSEI')
nifty$get_history()
#>                    date volume     high      low     open    close adj_close
#> 1   2022-01-03 03:45:00 200500 17646.65 17383.30 17387.15 17625.70  17625.70
#> 2   2022-01-04 03:45:00 247400 17827.60 17593.55 17681.40 17805.25  17805.25
#> 3   2022-01-05 03:45:00 251500 17944.70 17748.85 17820.10 17925.25  17925.25
#> 4   2022-01-06 03:45:00 236500 17797.95 17655.55 17768.50 17745.90  17745.90
#> 5   2022-01-07 03:45:00 239300 17905.00 17704.55 17797.60 17812.70  17812.70
#> 6   2022-01-10 03:45:00 232400 18017.45 17879.15 17913.30 18003.30  18003.30
#> 7   2022-01-11 03:45:00 220200 18081.25 17964.40 17997.75 18055.75  18055.75
#> 8   2022-01-12 03:45:00 245000 18227.95 18128.80 18170.40 18212.35  18212.35
#> 9   2022-01-13 03:45:00 303700 18272.25 18163.80 18257.00 18257.80  18257.80
#> 10  2022-01-14 03:45:00 229500 18286.95 18119.65 18185.00 18255.75  18255.75
#> 11  2022-01-17 03:45:00 266700 18321.55 18228.75 18235.65 18308.10  18308.10
#> 12  2022-01-18 03:45:00 227500 18350.95 18085.90 18337.20 18113.05  18113.05
#> 13  2022-01-19 03:45:00 276700 18129.20 17884.90 18129.20 17938.40  17938.40
#> 14  2022-01-20 03:45:00 258100 17943.70 17648.45 17921.00 17757.00  17757.00
#> 15  2022-01-21 03:45:00 277600 17707.60 17485.85 17613.70 17617.15  17617.15
#> 16  2022-01-24 03:45:00 323800 17599.40 16997.85 17575.15 17149.10  17149.10
#> 17  2022-01-25 03:45:00 326500 17309.15 16836.80 17001.55 17277.95  17277.95
#> 18  2022-01-27 03:45:00 395600 17182.50 16866.75 17062.00 17110.15  17110.15
#> 19  2022-01-28 03:45:00 355300 17373.50 17077.10 17208.30 17101.95  17101.95
#> 20  2022-01-31 03:45:00 321700 17410.00 17264.15 17301.05 17339.85  17339.85
#> 21  2022-02-01 03:45:00 386400 17622.40 17244.55 17529.45 17576.85  17576.85
#> 22  2022-02-02 03:45:00 271200 17794.60 17674.80 17706.20 17780.00  17780.00
#> 23  2022-02-03 03:45:00 226600 17781.15 17511.15 17767.75 17560.20  17560.20
#> 24  2022-02-04 03:45:00 261400 17617.80 17462.55 17590.20 17516.30  17516.30
#> 25  2022-02-07 03:45:00 265000 17536.75 17119.40 17456.30 17213.60  17213.60
#> 26  2022-02-08 03:45:00 268500 17306.45 17043.65 17279.85 17266.75  17266.75
#> 27  2022-02-09 03:45:00 236400 17477.15 17339.00 17370.10 17463.80  17463.80
#> 28  2022-02-10 03:45:00 273600 17639.45 17427.15 17554.10 17605.85  17605.85
#> 29  2022-02-11 03:45:00 253100 17454.75 17303.00 17451.00 17374.75  17374.75
#> 30  2022-02-14 03:45:00 305500 17099.50 16809.65 17076.15 16842.80  16842.80
#> 31  2022-02-15 03:45:00 298700 17375.00 16839.25 16933.25 17352.45  17352.45
#> 32  2022-02-16 03:45:00 244500 17490.60 17257.70 17408.45 17322.20  17322.20
#> 33  2022-02-17 03:45:00 232100 17442.90 17235.85 17396.55 17304.60  17304.60
#> 34  2022-02-18 03:45:00 189600 17380.80 17219.20 17236.05 17276.30  17276.30
#> 35  2022-02-21 03:45:00 215200 17351.05 17070.70 17192.25 17206.65  17206.65
#> 36  2022-02-22 03:45:00 300100 17148.55 16843.80 16847.95 17092.20  17092.20
#> 37  2022-02-23 03:45:00 200200 17220.70 17027.85 17194.50 17063.25  17063.25
#> 38  2022-02-24 03:45:00 458000 16705.25 16203.25 16548.90 16247.95  16247.95
#> 39  2022-02-25 03:45:00 329800 16748.80 16478.30 16515.65 16658.40  16658.40
#> 40  2022-02-28 03:45:00 404200 16815.90 16356.30 16481.60 16793.90  16793.90
#> 41  2022-03-02 03:45:00 517700 16678.50 16478.65 16593.10 16605.95  16605.95
#> 42  2022-03-03 03:45:00 442100 16768.95 16442.95 16723.20 16498.05  16498.05
#> 43  2022-03-04 03:45:00 456100 16456.00 16133.80 16339.45 16245.35  16245.35
#> 44  2022-03-07 03:45:00 585400 15944.60 15711.45 15867.95 15863.15  15863.15
#> 45  2022-03-08 03:45:00 543600 16028.75 15671.45 15747.75 16013.45  16013.45
#> 46  2022-03-09 03:45:00 462200 16418.05 15990.00 16078.00 16345.35  16345.35
#> 47  2022-03-10 03:45:00 486400 16757.30 16447.90 16757.10 16594.90  16594.90
#> 48  2022-03-11 03:45:00 343700 16694.40 16470.90 16528.80 16630.45  16630.45
#> 49  2022-03-14 03:45:00 314700 16887.95 16606.50 16633.70 16871.30  16871.30
#> 50  2022-03-15 03:45:00 381600 16927.75 16555.00 16900.65 16663.00  16663.00
#> 51  2022-03-16 03:45:00 260000 16987.90 16837.85 16876.65 16975.35  16975.35
#> 52  2022-03-17 03:45:00 448300 17344.60 17175.75 17202.90 17287.05  17287.05
#> 53  2022-03-21 03:45:00 285500 17353.35 17096.40 17329.50 17117.60  17117.60
#> 54  2022-03-22 03:45:00 338300 17334.40 17006.30 17120.40 17315.50  17315.50
#> 55  2022-03-23 03:45:00 292100 17442.40 17199.60 17405.05 17245.65  17245.65
#> 56  2022-03-24 03:45:00 290300 17291.75 17091.15 17094.95 17222.75  17222.75
#> 57  2022-03-25 03:45:00 237100 17294.90 17076.55 17289.00 17153.00  17153.00
#> 58  2022-03-28 03:45:00 253200 17235.10 17003.90 17181.85 17222.00  17222.00
#> 59  2022-03-29 03:45:00 235100 17343.65 17235.70 17297.20 17325.30  17325.30
#> 60  2022-03-30 03:45:00 505000 17522.50 17387.20 17468.15 17498.25  17498.25
#> 61  2022-03-31 03:45:00 281100 17559.80 17435.20 17519.20 17464.75  17464.75
#> 62  2022-04-01 03:45:00 291800 17703.70 17422.70 17436.90 17670.45  17670.45
#> 63  2022-04-04 03:45:00 345500 18114.65 17791.40 17809.10 18053.40  18053.40
#> 64  2022-04-05 03:45:00 283500 18095.45 17921.55 18080.60 17957.40  17957.40
#> 65  2022-04-06 03:45:00 328800 17901.00 17779.85 17842.75 17807.65  17807.65
#> 66  2022-04-07 03:45:00 308800 17787.50 17623.70 17723.30 17639.55  17639.55
#> 67  2022-04-08 03:45:00 274400 17842.75 17600.55 17698.15 17784.35  17784.35
#> 68  2022-04-11 03:45:00 251700 17779.05 17650.95 17740.90 17674.95  17674.95
#> 69  2022-04-12 03:45:00 266000 17595.30 17442.35 17584.85 17530.30  17530.30
#> 70  2022-04-13 03:45:00 245100 17663.65 17457.40 17599.90 17475.65  17475.65
#> 71  2022-04-18 03:45:00 376100 17237.75 17067.85 17183.45 17173.65  17173.65
#> 72  2022-04-19 03:45:00 401400 17275.65 16824.70 17258.95 16958.65  16958.65
#> 73  2022-04-20 03:45:00 286100 17186.90 16978.95 17045.25 17136.55  17136.55
#> 74  2022-04-21 03:45:00 285200 17414.70 17215.50 17234.60 17392.60  17392.60
#> 75  2022-04-22 03:45:00 262700 17315.30 17149.20 17242.75 17171.95  17171.95
#> 76  2022-04-25 03:45:00 275700 17054.30 16888.70 17009.05 16953.95  16953.95
#> 77  2022-04-26 03:45:00 261100 17223.85 17064.45 17121.30 17200.80  17200.80
#> 78  2022-04-27 03:45:00 265100 17110.70 16958.45 17073.35 17038.40  17038.40
#> 79  2022-04-28 03:45:00 312900 17322.50 17071.05 17189.50 17245.05  17245.05
#> 80  2022-04-29 03:45:00 336200 17377.65 17053.25 17329.25 17102.55  17102.55
#> 81  2022-05-02 03:45:00 278200 17092.25 16917.25 16924.45 17069.10  17069.10
#> 82  2022-05-04 03:45:00 310600 17132.85 16623.95 17096.60 16677.60  16677.60
#> 83  2022-05-05 03:45:00 265800 16945.70 16651.85 16854.75 16682.65  16682.65
#> 84  2022-05-06 03:45:00 300500 16484.20 16340.90 16415.55 16411.25  16411.25
#> 85  2022-05-09 03:45:00 288400 16403.70 16142.10 16227.70 16301.85  16301.85
#> 86  2022-05-10 03:45:00 283100 16404.55 16197.30 16248.90 16240.05  16240.05
#> 87  2022-05-11 03:45:00 284300 16318.75 15992.60 16270.05 16167.10  16167.10
#> 88  2022-05-12 03:45:00 314900 16041.95 15735.75 16021.10 15808.00  15808.00
#> 89  2022-05-13 03:45:00 369100 16083.60 15740.85 15977.00 15782.15  15782.15
#> 90  2022-05-16 03:45:00 217600 15977.95 15739.65 15845.10 15842.30  15842.30
#> 91  2022-05-17 03:45:00 295700 16284.25 15900.80 15912.60 16259.30  16259.30
#> 92  2022-05-18 03:45:00 290400 16399.80 16211.20 16318.15 16240.30  16240.30
#> 93  2022-05-19 03:45:00 313900 15984.75 15775.20 15917.40 15809.40  15809.40
#> 94  2022-05-20 03:45:00 252400 16283.05 16003.85 16043.80 16266.15  16266.15
#> 95  2022-05-23 03:45:00 293800 16414.70 16185.75 16290.95 16214.70  16214.70
#> 96  2022-05-24 03:45:00 249800 16262.80 16078.60 16225.55 16125.15  16125.15
#> 97  2022-05-25 03:45:00 243300 16223.35 16006.95 16196.35 16025.80  16025.80
#> 98  2022-05-26 03:45:00 314300 16204.45 15903.70 16105.00 16170.15  16170.15
#> 99  2022-05-27 03:45:00 274100 16370.60 16221.95 16296.60 16352.45  16352.45
#> 100 2022-05-30 03:45:00 251400 16695.50 16506.15 16527.90 16661.40  16661.40
#> 101 2022-05-31 03:45:00 651600 16690.75 16521.90 16578.45 16584.55  16584.55
#> 102 2022-06-01 03:45:00 249600 16649.20 16438.85 16594.40 16522.75  16522.75
#> 103 2022-06-02 03:45:00 236000 16646.40 16443.05 16481.65 16628.00  16628.00
#> 104 2022-06-03 03:45:00 245500 16793.85 16567.90 16761.65 16584.30  16584.30
#> 105 2022-06-06 03:45:00 233600 16610.95 16444.55 16530.70 16569.55  16569.55
#> 106 2022-06-07 03:45:00 233800 16487.25 16347.10 16469.60 16416.35  16416.35
#> 107 2022-06-08 03:45:00 243500 16514.30 16293.35 16474.95 16356.25  16356.25
#> 108 2022-06-09 03:45:00 205000 16492.80 16243.85 16263.85 16478.10  16478.10
#> 109 2022-06-10 03:45:00 189700 16324.70 16172.60 16283.95 16201.80  16201.80
#> 110 2022-06-13 03:45:00 225500 15886.15 15684.00 15877.55 15774.40  15774.40
#> 111 2022-06-14 03:45:00 225400 15858.00 15659.45 15674.25 15732.10  15732.10
#> 112 2022-06-15 03:45:00 183000 15783.65 15678.90 15729.25 15692.15  15692.15
#> 113 2022-06-16 03:45:00 264700 15863.15 15335.10 15832.25 15360.60  15360.60
#> 114 2022-06-17 03:45:00 342600 15400.40 15183.40 15272.65 15293.50  15293.50
#> 115 2022-06-20 03:45:00 260000 15382.50 15191.10 15334.50 15350.15  15350.15
#> 116 2022-06-21 03:45:00 262800 15707.25 15419.85 15455.95 15638.80  15638.80
#> 117 2022-06-22 03:45:00 220900 15565.40 15385.95 15545.65 15413.30  15413.30
#> 118 2022-06-23 03:45:00 259200 15628.45 15367.50 15451.55 15556.65  15556.65
#> 119 2022-06-24 03:45:00 219600 15749.25 15619.45 15657.40 15699.25  15699.25
#> 120 2022-06-27 03:45:00 210900 15927.45 15815.50 15926.20 15832.05  15832.05
#> 121 2022-06-28 03:45:00 251900 15892.10 15710.15 15757.45 15850.20  15850.20
#> 122 2022-06-29 03:45:00 444900 15861.60 15687.80 15701.70 15799.10  15799.10
#> 123 2022-06-30 03:45:00 306000 15890.00 15728.85 15774.50 15780.25  15780.25
#> 124 2022-07-01 03:45:00 364100 15793.95 15511.05 15703.70 15752.05  15752.05
#> 125 2022-07-04 03:45:00 304300 15852.35 15661.80 15710.50 15835.35  15835.35
#> 126 2022-07-05 03:45:00 254200 16025.75 15785.45 15909.15 15810.85  15810.85
#> 127 2022-07-06 03:45:00 288400 16011.35 15800.90 15818.20 15989.80  15989.80
#> 128 2022-07-07 03:45:00 264600 16150.50 16045.95 16113.75 16132.90  16132.90
#> 129 2022-07-08 03:45:00 281100 16275.50 16157.90 16273.65 16220.60  16220.60
#> 130 2022-07-11 03:45:00 255900 16248.55 16115.50 16136.15 16216.00  16216.00
#> 131 2022-07-12 03:45:00 208600 16158.75 16031.15 16126.20 16058.30  16058.30
#> 132 2022-07-13 03:45:00 233300 16140.00 15950.15 16128.20 15966.65  15966.65
#> 133 2022-07-14 03:45:00 228600 16070.85 15858.20 16018.85 15938.65  15938.65
#> 134 2022-07-15 03:45:00 205800 16066.95 15927.30 16010.80 16049.20  16049.20
#> 135 2022-07-18 03:45:00 227700 16287.95 16142.20 16151.40 16278.50  16278.50
#> 136 2022-07-19 03:45:00 222600 16359.50 16187.05 16187.05 16340.55  16340.55
#> 137 2022-07-20 03:45:00 284400 16588.00 16490.95 16562.80 16520.85  16520.85
#> 138 2022-07-21 03:45:00 243600 16626.95 16483.90 16523.55 16605.25  16605.25
#> 139 2022-07-22 03:45:00 190900 16752.25 16610.90 16661.25 16719.45  16719.45
#> 140 2022-07-25 03:45:00 211400 16706.05 16564.25 16662.55 16631.00  16631.00
#> 141 2022-07-26 03:45:00 208200 16636.10 16463.30 16632.90 16483.85  16483.85
#> 142 2022-07-27 03:45:00 185200 16653.45 16438.75 16475.35 16641.80  16641.80
#> 143 2022-07-28 03:45:00 376300 16947.65 16746.25 16774.85 16929.60  16929.60
#> 144 2022-07-29 03:45:00 436100 17172.80 17018.15 17079.50 17158.25  17158.25
#> 145 2022-08-01 03:45:00 381500 17356.25 17154.80 17243.20 17340.05  17340.05
#> 146 2022-08-02 03:45:00 333600 17390.15 17215.85 17310.15 17345.45  17345.45
#> 147 2022-08-03 03:45:00 285400 17407.50 17225.85 17349.25 17388.15  17388.15
#> 148 2022-08-04 03:45:00 325000 17490.70 17161.25 17463.10 17382.00  17382.00
#> 149 2022-08-05 03:45:00 355300 17474.40 17348.75 17423.65 17397.50  17397.50
#> 150 2022-08-08 03:45:00 257000 17548.80 17359.75 17401.50 17525.10  17525.10
#> 151 2022-08-10 03:45:00 312700 17566.10 17442.80 17566.10 17534.75  17534.75
#> 152 2022-08-11 03:45:00 311200 17719.30 17631.95 17711.65 17659.00  17659.00
#> 153 2022-08-12 03:45:00 303900 17724.65 17597.85 17659.65 17698.15  17698.15
#> 154 2022-08-16 03:45:00      0 17839.10 17764.05 17797.20 17825.25  17825.25
#> 155 2022-08-17 03:45:00 262800 17965.95 17833.35 17868.15 17944.25  17944.25
#> 156 2022-08-18 03:45:00 263900 17968.45 17852.05 17898.65 17956.50  17956.50
#> 157 2022-08-19 03:45:00 295600 17992.20 17710.75 17966.55 17758.45  17758.45
#> 158 2022-08-22 03:45:00 287600 17690.05 17467.35 17682.90 17490.70  17490.70
#> 159 2022-08-23 03:45:00 285600 17625.55 17345.20 17357.35 17577.50  17577.50
#> 160 2022-08-24 10:01:44      0 17623.65 17499.25 17525.45 17604.95  17604.95
nifty$get_history(start = '2022-07-01', interval = '1d')
#>                   date volume     high      low     open    close adj_close
#> 1  2022-07-01 03:45:00 364100 15793.95 15511.05 15703.70 15752.05  15752.05
#> 2  2022-07-04 03:45:00 304300 15852.35 15661.80 15710.50 15835.35  15835.35
#> 3  2022-07-05 03:45:00 254200 16025.75 15785.45 15909.15 15810.85  15810.85
#> 4  2022-07-06 03:45:00 288400 16011.35 15800.90 15818.20 15989.80  15989.80
#> 5  2022-07-07 03:45:00 264600 16150.50 16045.95 16113.75 16132.90  16132.90
#> 6  2022-07-08 03:45:00 281100 16275.50 16157.90 16273.65 16220.60  16220.60
#> 7  2022-07-11 03:45:00 255900 16248.55 16115.50 16136.15 16216.00  16216.00
#> 8  2022-07-12 03:45:00 208600 16158.75 16031.15 16126.20 16058.30  16058.30
#> 9  2022-07-13 03:45:00 233300 16140.00 15950.15 16128.20 15966.65  15966.65
#> 10 2022-07-14 03:45:00 228600 16070.85 15858.20 16018.85 15938.65  15938.65
#> 11 2022-07-15 03:45:00 205800 16066.95 15927.30 16010.80 16049.20  16049.20
#> 12 2022-07-18 03:45:00 227700 16287.95 16142.20 16151.40 16278.50  16278.50
#> 13 2022-07-19 03:45:00 222600 16359.50 16187.05 16187.05 16340.55  16340.55
#> 14 2022-07-20 03:45:00 284400 16588.00 16490.95 16562.80 16520.85  16520.85
#> 15 2022-07-21 03:45:00 243600 16626.95 16483.90 16523.55 16605.25  16605.25
#> 16 2022-07-22 03:45:00 190900 16752.25 16610.90 16661.25 16719.45  16719.45
#> 17 2022-07-25 03:45:00 211400 16706.05 16564.25 16662.55 16631.00  16631.00
#> 18 2022-07-26 03:45:00 208200 16636.10 16463.30 16632.90 16483.85  16483.85
#> 19 2022-07-27 03:45:00 185200 16653.45 16438.75 16475.35 16641.80  16641.80
#> 20 2022-07-28 03:45:00 376300 16947.65 16746.25 16774.85 16929.60  16929.60
#> 21 2022-07-29 03:45:00 436100 17172.80 17018.15 17079.50 17158.25  17158.25
#> 22 2022-08-01 03:45:00 381500 17356.25 17154.80 17243.20 17340.05  17340.05
#> 23 2022-08-02 03:45:00 333600 17390.15 17215.85 17310.15 17345.45  17345.45
#> 24 2022-08-03 03:45:00 285400 17407.50 17225.85 17349.25 17388.15  17388.15
#> 25 2022-08-04 03:45:00 325000 17490.70 17161.25 17463.10 17382.00  17382.00
#> 26 2022-08-05 03:45:00 355300 17474.40 17348.75 17423.65 17397.50  17397.50
#> 27 2022-08-08 03:45:00 257000 17548.80 17359.75 17401.50 17525.10  17525.10
#> 28 2022-08-10 03:45:00 312700 17566.10 17442.80 17566.10 17534.75  17534.75
#> 29 2022-08-11 03:45:00 311200 17719.30 17631.95 17711.65 17659.00  17659.00
#> 30 2022-08-12 03:45:00 303900 17724.65 17597.85 17659.65 17698.15  17698.15
#> 31 2022-08-16 03:45:00      0 17839.10 17764.05 17797.20 17825.25  17825.25
#> 32 2022-08-17 03:45:00 262800 17965.95 17833.35 17868.15 17944.25  17944.25
#> 33 2022-08-18 03:45:00 263900 17968.45 17852.05 17898.65 17956.50  17956.50
#> 34 2022-08-19 03:45:00 295600 17992.20 17710.75 17966.55 17758.45  17758.45
#> 35 2022-08-22 03:45:00 287600 17690.05 17467.35 17682.90 17490.70  17490.70
#> 36 2022-08-23 03:45:00 285600 17625.55 17345.20 17357.35 17577.50  17577.50
#> 37 2022-08-24 10:01:44      0 17623.65 17499.25 17525.45 17604.95  17604.95
nifty$get_history(start = '2022-07-01', end = '2022-07-14', interval = '1d')
#>                  date volume     high      low     open    close adj_close
#> 1 2022-07-01 03:45:00 364100 15793.95 15511.05 15703.70 15752.05  15752.05
#> 2 2022-07-04 03:45:00 304300 15852.35 15661.80 15710.50 15835.35  15835.35
#> 3 2022-07-05 03:45:00 254200 16025.75 15785.45 15909.15 15810.85  15810.85
#> 4 2022-07-06 03:45:00 288400 16011.35 15800.90 15818.20 15989.80  15989.80
#> 5 2022-07-07 03:45:00 264600 16150.50 16045.95 16113.75 16132.90  16132.90
#> 6 2022-07-08 03:45:00 281100 16275.50 16157.90 16273.65 16220.60  16220.60
#> 7 2022-07-11 03:45:00 255900 16248.55 16115.50 16136.15 16216.00  16216.00
#> 8 2022-07-12 03:45:00 208600 16158.75 16031.15 16126.20 16058.30  16058.30
#> 9 2022-07-13 03:45:00 233300 16140.00 15950.15 16128.20 15966.65  15966.65
nifty$get_history(period = '1mo', interval = '1d')
#>                   date volume     high      low     open    close adj_close
#> 1  2022-07-25 03:45:00 211400 16706.05 16564.25 16662.55 16631.00  16631.00
#> 2  2022-07-26 03:45:00 208200 16636.10 16463.30 16632.90 16483.85  16483.85
#> 3  2022-07-27 03:45:00 185200 16653.45 16438.75 16475.35 16641.80  16641.80
#> 4  2022-07-28 03:45:00 376300 16947.65 16746.25 16774.85 16929.60  16929.60
#> 5  2022-07-29 03:45:00 436100 17172.80 17018.15 17079.50 17158.25  17158.25
#> 6  2022-08-01 03:45:00 381500 17356.25 17154.80 17243.20 17340.05  17340.05
#> 7  2022-08-02 03:45:00 333600 17390.15 17215.85 17310.15 17345.45  17345.45
#> 8  2022-08-03 03:45:00 285400 17407.50 17225.85 17349.25 17388.15  17388.15
#> 9  2022-08-04 03:45:00 325000 17490.70 17161.25 17463.10 17382.00  17382.00
#> 10 2022-08-05 03:45:00 355300 17474.40 17348.75 17423.65 17397.50  17397.50
#> 11 2022-08-08 03:45:00 257000 17548.80 17359.75 17401.50 17525.10  17525.10
#> 12 2022-08-10 03:45:00 312700 17566.10 17442.80 17566.10 17534.75  17534.75
#> 13 2022-08-11 03:45:00 311200 17719.30 17631.95 17711.65 17659.00  17659.00
#> 14 2022-08-12 03:45:00 303900 17724.65 17597.85 17659.65 17698.15  17698.15
#> 15 2022-08-16 03:45:00      0 17839.10 17764.05 17797.20 17825.25  17825.25
#> 16 2022-08-17 03:45:00 262800 17965.95 17833.35 17868.15 17944.25  17944.25
#> 17 2022-08-18 03:45:00 263900 17968.45 17852.05 17898.65 17956.50  17956.50
#> 18 2022-08-19 03:45:00 295600 17992.20 17710.75 17966.55 17758.45  17758.45
#> 19 2022-08-22 03:45:00 287600 17690.05 17467.35 17682.90 17490.70  17490.70
#> 20 2022-08-23 03:45:00 285600 17625.55 17345.20 17357.35 17577.50  17577.50
#> 21 2022-08-24 10:01:44      0 17623.65 17499.25 17525.45 17604.95  17604.95