R6 Class Representing a Ticker

Base class for getting all data related to indices from Yahoo Finance API.

Format

An R6 class object

Public fields

index

Index for which data is retrieved

Methods

Public methods

Index$new()

Index$set_index()

Index$get_history()

Index$clone()

Method new()

Create a new Index object

Usage

Index$new(index = NA)

Arguments

index

Index

Returns

A new `Index` object

Examples

nifty_50 <- Index$new('^NSEI')

Method set_index()

Set a new index.

Usage

Index$set_index(index)

Arguments

index

New index

Examples

indice <- Index$new('^NSEI')
indice$set_index('^NDX')

Method get_history()

Retrieves historical data

Usage

Index$get_history(period = "ytd", interval = "1d", start = NULL, end = NULL)

Arguments

period

Length of time. Defaults to 'ytd'. Valid values are:

'1d'

'5d'

'1mo'

'3mo'

'6mo'

'1y'

'2y'

'5y'

'10y'

'ytd'

'max'

interval

Time between data points. Defaults to '1d'. Valid values are:

'1m'

'2m'

'5m'

'15m'

'30m'

'60m'

'90m'

'1h'

'1d'

'5d'

'1wk'

'1mo'

'3mo'

start

Specific starting date. String or date object in yyyy-mm-dd format.

end

Specific ending date. String or date object in yyyy-mm-dd format.

Returns

A data.frame.

Examples

\donttest{
nifty <- Index$new('^NSEI')
nifty$get_history(start = '2022-07-01', interval = '1d')
nifty$get_history(start = '2022-07-01', end = '2022-07-14', interval = '1d')
nifty$get_history(period = '1mo', interval = '1d')
}

Method clone()

The objects of this class are cloneable with this method.

Usage

Index$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `Index$new`
## ------------------------------------------------

nifty_50 <- Index$new('^NSEI')

## ------------------------------------------------
## Method `Index$set_index`
## ------------------------------------------------

indice <- Index$new('^NSEI')
indice$set_index('^NDX')

## ------------------------------------------------
## Method `Index$get_history`
## ------------------------------------------------

# \donttest{
nifty <- Index$new('^NSEI')
nifty$get_history(start = '2022-07-01', interval = '1d')
#>                    date volume     high      low     open    close adj_close
#> 1   2022-07-01 03:45:00 364100 15793.95 15511.05 15703.70 15752.05  15752.05
#> 2   2022-07-04 03:45:00 304300 15852.35 15661.80 15710.50 15835.35  15835.35
#> 3   2022-07-05 03:45:00 254200 16025.75 15785.45 15909.15 15810.85  15810.85
#> 4   2022-07-06 03:45:00 288400 16011.35 15800.90 15818.20 15989.80  15989.80
#> 5   2022-07-07 03:45:00 264600 16150.50 16045.95 16113.75 16132.90  16132.90
#> 6   2022-07-08 03:45:00 281100 16275.50 16157.90 16273.65 16220.60  16220.60
#> 7   2022-07-11 03:45:00 255900 16248.55 16115.50 16136.15 16216.00  16216.00
#> 8   2022-07-12 03:45:00 208600 16158.75 16031.15 16126.20 16058.30  16058.30
#> 9   2022-07-13 03:45:00 233300 16140.00 15950.15 16128.20 15966.65  15966.65
#> 10  2022-07-14 03:45:00 228600 16070.85 15858.20 16018.85 15938.65  15938.65
#> 11  2022-07-15 03:45:00 205800 16066.95 15927.30 16010.80 16049.20  16049.20
#> 12  2022-07-18 03:45:00 227700 16287.95 16142.20 16151.40 16278.50  16278.50
#> 13  2022-07-19 03:45:00 222600 16359.50 16187.05 16187.05 16340.55  16340.55
#> 14  2022-07-20 03:45:00 284400 16588.00 16490.95 16562.80 16520.85  16520.85
#> 15  2022-07-21 03:45:00 243600 16626.95 16483.90 16523.55 16605.25  16605.25
#> 16  2022-07-22 03:45:00 190900 16752.25 16610.90 16661.25 16719.45  16719.45
#> 17  2022-07-25 03:45:00 211400 16706.05 16564.25 16662.55 16631.00  16631.00
#> 18  2022-07-26 03:45:00 208200 16636.10 16463.30 16632.90 16483.85  16483.85
#> 19  2022-07-27 03:45:00 185200 16653.45 16438.75 16475.35 16641.80  16641.80
#> 20  2022-07-28 03:45:00 376300 16947.65 16746.25 16774.85 16929.60  16929.60
#> 21  2022-07-29 03:45:00 436100 17172.80 17018.15 17079.50 17158.25  17158.25
#> 22  2022-08-01 03:45:00 381500 17356.25 17154.80 17243.20 17340.05  17340.05
#> 23  2022-08-02 03:45:00 333600 17390.15 17215.85 17310.15 17345.45  17345.45
#> 24  2022-08-03 03:45:00 285400 17407.50 17225.85 17349.25 17388.15  17388.15
#> 25  2022-08-04 03:45:00 325000 17490.70 17161.25 17463.10 17382.00  17382.00
#> 26  2022-08-05 03:45:00 355300 17474.40 17348.75 17423.65 17397.50  17397.50
#> 27  2022-08-08 03:45:00 257000 17548.80 17359.75 17401.50 17525.10  17525.10
#> 28  2022-08-10 03:45:00 312700 17566.10 17442.80 17566.10 17534.75  17534.75
#> 29  2022-08-11 03:45:00 311200 17719.30 17631.95 17711.65 17659.00  17659.00
#> 30  2022-08-12 03:45:00 303900 17724.65 17597.85 17659.65 17698.15  17698.15
#> 31  2022-08-16 03:45:00      0 17839.10 17764.05 17797.20 17825.25  17825.25
#> 32  2022-08-17 03:45:00 262800 17965.95 17833.35 17868.15 17944.25  17944.25
#> 33  2022-08-18 03:45:00 263900 17968.45 17852.05 17898.65 17956.50  17956.50
#> 34  2022-08-19 03:45:00 295600 17992.20 17710.75 17966.55 17758.45  17758.45
#> 35  2022-08-22 03:45:00 287600 17690.05 17467.35 17682.90 17490.70  17490.70
#> 36  2022-08-23 03:45:00 285600 17625.55 17345.20 17357.35 17577.50  17577.50
#> 37  2022-08-24 03:45:00 262000 17623.65 17499.25 17525.45 17604.95  17604.95
#> 38  2022-08-25 03:45:00 230200 17726.50 17487.45 17679.00 17522.45  17522.45
#> 39  2022-08-26 03:45:00 266600 17685.85 17519.35 17619.30 17558.90  17558.90
#> 40  2022-08-29 03:45:00 244900 17380.15 17166.20 17188.65 17312.90  17312.90
#> 41  2022-08-30 03:45:00 324700 17777.65 17401.50 17414.95 17759.30  17759.30
#> 42  2022-09-01 03:45:00 308500 17695.60 17468.45 17485.70 17542.80  17542.80
#> 43  2022-09-02 03:45:00 256300 17643.85 17476.45 17598.40 17539.45  17539.45
#> 44  2022-09-05 03:45:00 230300 17683.15 17540.35 17546.45 17665.80  17665.80
#> 45  2022-09-06 03:45:00 251200 17764.65 17587.65 17695.70 17655.60  17655.60
#> 46  2022-09-07 03:45:00 354100 17650.75 17484.30 17519.40 17624.40  17624.40
#> 47  2022-09-08 03:45:00 279800 17807.65 17691.95 17748.15 17798.75  17798.75
#> 48  2022-09-09 03:45:00 270300 17925.95 17786.00 17923.35 17833.35  17833.35
#> 49  2022-09-12 03:45:00 228200 17980.55 17889.15 17890.85 17936.35  17936.35
#> 50  2022-09-13 03:45:00 259900 18088.30 18015.45 18044.45 18070.05  18070.05
#> 51  2022-09-14 03:45:00 365900 18091.55 17771.15 17771.15 18003.75  18003.75
#> 52  2022-09-15 03:45:00 289600 18096.15 17861.50 18046.35 17877.40  17877.40
#> 53  2022-09-16 03:45:00 468500 17820.05 17497.25 17796.80 17530.85  17530.85
#> 54  2022-09-19 03:45:00 258300 17667.20 17429.70 17540.65 17622.25  17622.25
#> 55  2022-09-20 03:45:00 263100 17919.30 17744.40 17770.40 17816.25  17816.25
#> 56  2022-09-21 03:45:00 245500 17838.70 17663.60 17766.35 17718.35  17718.35
#> 57  2022-09-22 03:45:00 284100 17722.75 17532.45 17609.65 17629.80  17629.80
#> 58  2022-09-23 03:45:00 390800 17642.15 17291.65 17593.85 17327.35  17327.35
#> 59  2022-09-26 03:45:00 492000 17196.40 16978.30 17156.30 17016.30  17016.30
#> 60  2022-09-27 03:45:00 359900 17176.45 16942.35 17110.90 17007.40  17007.40
#> 61  2022-09-28 03:45:00 323900 17037.60 16820.40 16870.55 16858.60  16858.60
#> 62  2022-09-29 03:45:00 340000 17026.05 16788.60 16993.60 16818.10  16818.10
#> 63  2022-09-30 03:45:00 376100 17187.10 16747.70 16798.05 17094.35  17094.35
#> 64  2022-10-03 03:45:00 278400 17114.65 16855.55 17102.10 16887.35  16887.35
#> 65  2022-10-04 03:45:00 226000 17287.30 17117.30 17147.45 17274.30  17274.30
#> 66  2022-10-06 03:45:00 265500 17428.80 17315.65 17379.25 17331.80  17331.80
#> 67  2022-10-07 03:45:00 216300 17337.35 17216.95 17287.20 17314.65  17314.65
#> 68  2022-10-10 03:45:00 234000 17280.15 17064.70 17094.35 17241.00  17241.00
#> 69  2022-10-11 03:45:00 282600 17261.80 16950.30 17256.05 16983.55  16983.55
#> 70  2022-10-12 03:45:00 256000 17142.35 16960.05 17025.55 17123.60  17123.60
#> 71  2022-10-13 03:45:00 266400 17112.35 16956.95 17087.35 17014.35  17014.35
#> 72  2022-10-14 03:45:00 227000 17348.55 17169.75 17322.30 17185.70  17185.70
#> 73  2022-10-17 03:45:00 212200 17328.55 17098.55 17144.80 17311.80  17311.80
#> 74  2022-10-18 03:45:00 239500 17527.80 17434.05 17438.75 17486.95  17486.95
#> 75  2022-10-19 03:45:00 210500 17607.60 17472.85 17568.15 17512.25  17512.25
#> 76  2022-10-20 03:45:00 249600 17584.15 17421.00 17423.10 17563.95  17563.95
#> 77  2022-10-21 03:45:00 277700 17670.15 17520.75 17622.85 17576.30  17576.30
#> 78  2022-10-24 12:45:00  45000 17777.55 17707.40 17736.35 17730.75  17730.75
#> 79  2022-10-25 03:45:00 251400 17811.50 17637.00 17808.30 17656.35  17656.35
#> 80  2022-10-27 03:45:00 324600 17783.90 17654.50 17771.40 17736.95  17736.95
#> 81  2022-10-28 03:45:00 250000 17838.90 17723.70 17756.40 17786.80  17786.80
#> 82  2022-10-31 03:45:00 227200 18022.80 17899.90 17910.20 18012.20  18012.20
#> 83  2022-11-01 03:45:00 349900 18175.80 18060.15 18130.70 18145.40  18145.40
#> 84  2022-11-02 03:45:00 270900 18178.75 18048.65 18177.90 18082.85  18082.85
#> 85  2022-11-03 03:45:00 213000 18106.30 17959.20 17968.35 18052.70  18052.70
#> 86  2022-11-04 03:45:00 267900 18135.10 18017.15 18053.40 18117.15  18117.15
#> 87  2022-11-07 03:45:00 314800 18255.50 18064.75 18211.75 18202.80  18202.80
#> 88  2022-11-09 03:45:00 307200 18296.40 18117.50 18288.25 18157.00  18157.00
#> 89  2022-11-10 03:45:00 256500 18103.10 17969.40 18044.35 18028.20  18028.20
#> 90  2022-11-11 03:45:00 378500 18362.30 18259.35 18272.35 18349.70  18349.70
#> 91  2022-11-14 03:45:00 301400 18399.45 18311.40 18376.40 18329.15  18329.15
#> 92  2022-11-15 03:45:00 250900 18427.95 18282.00 18362.75 18403.40  18403.40
#> 93  2022-11-16 03:45:00 219300 18442.15 18344.15 18398.25 18409.65  18409.65
#> 94  2022-11-17 03:45:00 200500 18417.60 18312.95 18358.70 18343.90  18343.90
#> 95  2022-11-18 03:45:00 198800 18394.60 18209.80 18382.95 18307.65  18307.65
#> 96  2022-11-21 03:45:00 213800 18262.30 18133.35 18246.40 18159.95  18159.95
#> 97  2022-11-22 03:45:00 177700 18261.85 18137.70 18179.15 18244.20  18244.20
#> 98  2022-11-23 03:45:00 178800 18325.40 18246.00 18325.20 18267.25  18267.25
#> 99  2022-11-24 03:45:00 204000 18529.70 18294.25 18326.10 18484.10  18484.10
#> 100 2022-11-25 03:45:00 205900 18534.90 18445.10 18528.45 18512.75  18512.75
#> 101 2022-11-28 03:45:00 206500 18614.25 18365.60 18430.55 18562.75  18562.75
#> 102 2022-11-29 03:45:00 195300 18678.10 18552.15 18552.45 18618.05  18618.05
#> 103 2022-11-30 03:45:00 345600 18816.05 18616.55 18625.70 18758.35  18758.35
#> 104 2022-12-01 03:45:00 325000 18887.60 18778.20 18871.95 18812.50  18812.50
#> 105 2022-12-02 03:45:00 254400 18781.95 18639.20 18752.40 18696.10  18696.10
#> 106 2022-12-05 03:45:00 288400 18728.60 18591.35 18719.55 18701.05  18701.05
#> 107 2022-12-06 03:45:00 217800 18654.90 18577.90 18600.65 18642.75  18642.75
#> 108 2022-12-07 03:45:00 200500 18668.30 18528.40 18638.85 18560.50  18560.50
#> 109 2022-12-08 03:45:00 202800 18625.00 18536.95 18570.85 18609.35  18609.35
#> 110 2022-12-09 03:45:00 215800 18664.70 18410.10 18662.40 18496.60  18496.60
#> 111 2022-12-12 03:45:00 193900 18521.55 18345.70 18402.15 18497.15  18497.15
#> 112 2022-12-13 03:45:00 187600 18617.25 18490.20 18524.40 18608.00  18608.00
#> 113 2022-12-14 03:45:00 208000 18696.10 18632.90 18671.25 18660.30  18660.30
#> 114 2022-12-15 03:45:00 183100 18652.90 18387.70 18614.40 18414.90  18414.90
#> 115 2022-12-16 03:45:00 277700 18440.95 18255.15 18319.10 18269.00  18269.00
#> 116 2022-12-19 03:45:00 154800 18431.65 18244.55 18288.10 18420.45  18420.45
#> 117 2022-12-20 03:45:00 169000 18404.90 18202.65 18340.30 18385.30  18385.30
#> 118 2022-12-21 03:45:00 187600 18473.35 18162.75 18435.15 18199.10  18199.10
#> 119 2022-12-22 03:45:00 195700 18318.75 18068.60 18288.80 18127.35  18127.35
#> 120 2022-12-23 03:45:00 221500 18050.45 17779.50 17977.65 17806.80  17806.80
#> 121 2022-12-26 03:45:00 176700 18084.10 17774.25 17830.40 18014.60  18014.60
#> 122 2022-12-27 03:45:00 214300 18149.25 17967.45 18089.80 18132.30  18132.30
#> 123 2022-12-28 03:45:00 193900 18173.10 18068.35 18084.75 18122.50  18122.50
#> 124 2022-12-29 03:45:00 281100 18229.70 17992.80 18045.70 18191.00  18191.00
#> 125 2022-12-30 03:45:00 192000 18265.25 18080.30 18259.10 18105.30  18105.30
#> 126 2023-01-02 03:45:00 256100 18215.15 18086.50 18131.70 18197.45  18197.45
#> 127 2023-01-03 03:45:00 208700 18251.95 18149.80 18163.20 18232.55  18232.55
#> 128 2023-01-04 03:45:00 235200 18243.00 18020.60 18230.65 18042.95  18042.95
#> 129 2023-01-05 03:45:00 269900 18120.30 17892.60 18101.95 17992.15  17992.15
#> 130 2023-01-06 03:45:00 238200 18047.40 17795.55 18008.05 17859.45  17859.45
#> 131 2023-01-09 03:45:00 257200 18141.40 17936.15 17952.55 18101.20  18101.20
#> 132 2023-01-10 03:45:00 283300 18127.60 17856.00 18121.30 17914.15  17914.15
#> 133 2023-01-11 03:45:00 259900 17976.35 17824.35 17924.25 17895.70  17895.70
#> 134 2023-01-12 03:45:00 227800 17945.80 17761.65 17920.85 17858.20  17858.20
#> 135 2023-01-13 03:45:00 256700 17999.35 17774.25 17867.50 17956.60  17956.60
#> 136 2023-01-16 03:45:00 206200 18049.65 17853.65 18033.15 17894.85  17894.85
#> 137 2023-01-17 03:45:00 219100 18072.05 17886.95 17922.80 18053.30  18053.30
#> 138 2023-01-18 03:45:00 255800 18183.75 18032.45 18074.30 18165.35  18165.35
#> 139 2023-01-19 03:45:00 237800 18155.20 18063.75 18119.80 18107.85  18107.85
#> 140 2023-01-20 03:45:00 237200 18145.45 18016.20 18115.60 18027.65  18027.65
#> 141 2023-01-23 03:45:00 202500 18162.60 18063.45 18118.45 18118.55  18118.55
#> 142 2023-01-24 03:45:00 216900 18201.25 18078.65 18183.95 18118.30  18118.30
#> 143 2023-01-25 03:45:00 257200 18100.60 17846.15 18093.35 17891.95  17891.95
#> 144 2023-01-27 03:45:00 476300 17884.75 17493.55 17877.20 17604.35  17604.35
#> 145 2023-01-30 03:45:00 432400 17709.15 17405.55 17541.95 17648.95  17648.95
#> 146 2023-01-31 03:45:00 398300 17735.70 17537.55 17731.45 17662.15  17662.15
#> 147 2023-02-01 03:45:00 512900 17972.20 17353.40 17811.60 17616.30  17616.30
#> 148 2023-02-02 03:45:00 490100 17653.90 17445.95 17517.10 17610.40  17610.40
#> 149 2023-02-03 03:45:00 424100 17870.30 17584.20 17721.75 17854.05  17854.05
#> 150 2023-02-06 03:45:00 282500 17823.70 17698.35 17818.55 17764.60  17764.60
#> 151 2023-02-07 03:45:00 354400 17811.15 17652.55 17790.10 17721.50  17721.50
#> 152 2023-02-08 03:45:00 291000 17898.70 17744.15 17750.30 17871.70  17871.70
#> 153 2023-02-09 03:45:00 260900 17916.90 17779.80 17885.50 17893.45  17893.45
#> 154 2023-02-10 03:45:00 232000 17876.95 17801.00 17847.55 17856.50  17856.50
#> 155 2023-02-13 03:45:00 231300 17880.70 17719.75 17859.10 17770.90  17770.90
#> 156 2023-02-14 03:45:00 244500 17954.55 17800.05 17840.35 17929.85  17929.85
#> 157 2023-02-15 03:45:00 229300 18034.10 17853.80 17896.60 18015.85  18015.85
#> 158 2023-02-16 03:45:00 230200 18134.75 18000.65 18094.75 18035.85  18035.85
#> 159 2023-02-17 03:45:00 208100 18034.25 17884.60 17974.85 17944.20  17944.20
#> 160 2023-02-20 03:45:00 174600 18004.35 17818.40 17965.55 17844.60  17844.60
#> 161 2023-02-21 03:45:00 198200 17924.90 17800.30 17905.80 17826.70  17826.70
#> 162 2023-02-22 03:45:00 204000 17772.50 17529.45 17755.35 17554.30  17554.30
#> 163 2023-02-23 03:45:00 240800 17620.05 17455.40 17574.65 17511.25  17511.25
#> 164 2023-02-24 03:45:00 209400 17599.75 17421.80 17591.35 17465.80  17465.80
#> 165 2023-02-27 03:45:00 246400 17451.60 17299.00 17428.60 17392.70  17392.70
#> 166 2023-02-28 03:45:00 420900 17440.45 17255.20 17383.25 17303.95  17303.95
#> 167 2023-03-01 03:45:00 272900 17467.75 17345.25 17360.10 17450.90  17450.90
#> 168 2023-03-02 03:45:00 310700 17445.80 17306.00 17421.50 17321.90  17321.90
#> 169 2023-03-03 03:45:00 356200 17644.75 17427.70 17451.25 17594.35  17594.35
#> 170 2023-03-06 03:45:00 362800 17799.95 17671.95 17680.35 17711.45  17711.45
#> 171 2023-03-08 03:45:00 267000 17766.50 17602.25 17665.75 17754.40  17754.40
#> 172 2023-03-09 03:45:00 262400 17772.35 17573.60 17772.05 17589.60  17589.60
#> 173 2023-03-10 03:45:00 235900 17451.50 17324.35 17443.80 17412.90  17412.90
#> 174 2023-03-13 03:45:00 254900 17529.90 17113.45 17421.90 17154.30  17154.30
#> 175 2023-03-14 03:45:00 290600 17224.65 16987.10 17160.55 17043.30  17043.30
#> 176 2023-03-15 03:45:00 248200 17211.35 16938.90 17166.45 16972.15  16972.15
#> 177 2023-03-16 03:45:00 349800 17062.45 16850.15 16994.65 16985.60  16985.60
#> 178 2023-03-17 03:45:00 408100 17145.80 16958.15 17111.80 17100.05  17100.05
#> 179 2023-03-20 03:45:00 241800 17066.60 16828.35 17066.60 16988.40  16988.40
#> 180 2023-03-21 03:45:00 246700 17127.70 17016.00 17060.40 17107.50  17107.50
#> 181 2023-03-22 03:45:00      0 17207.25 17107.85 17177.45 17151.90  17151.90
#> 182 2023-03-23 03:45:00 219200 17205.40 17045.30 17097.40 17076.90  17076.90
#> 183 2023-03-24 03:45:00 228000 17109.45 16917.35 17076.20 16945.05  16945.05
#> 184 2023-03-27 03:45:00 218400 17091.00 16918.55 16984.30 16985.70  16985.70
#> 185 2023-03-28 03:45:00 238800 17061.75 16913.75 17031.75 16951.70  16951.70
#> 186 2023-03-29 03:45:00 345900 17126.15 16940.60 16977.30 17080.70  17080.70
#> 187 2023-03-31 03:45:00 263900 17381.60 17204.65 17210.35 17359.75  17359.75
#> 188 2023-04-03 03:45:00 230200 17428.05 17312.75 17427.95 17398.05  17398.05
#> 189 2023-04-05 03:45:00 248300 17570.55 17402.70 17422.30 17557.05  17557.05
#> 190 2023-04-06 03:45:00 242700 17638.70 17502.85 17533.85 17599.15  17599.15
#> 191 2023-04-10 03:45:00 254800 17694.10 17597.95 17634.90 17624.05  17624.05
#> 192 2023-04-11 03:45:00 304300 17748.75 17655.15 17704.80 17722.30  17722.30
#> 193 2023-04-12 03:45:00 232800 17825.75 17717.25 17759.55 17812.40  17812.40
#> 194 2023-04-13 03:45:00 285800 17842.15 17729.65 17807.30 17828.00  17828.00
#> 195 2023-04-17 03:45:00 357400 17863.00 17574.05 17863.00 17706.85  17706.85
#> 196 2023-04-18 03:45:00 258900 17766.60 17610.20 17766.60 17660.15  17660.15
#> 197 2023-04-19 03:45:00 281500 17666.15 17579.85 17653.35 17618.75  17618.75
#> 198 2023-04-20 03:45:00 215100 17684.45 17584.35 17638.60 17624.45  17624.45
#> 199 2023-04-21 03:45:00 230200 17663.20 17553.95 17639.75 17624.05  17624.05
#> 200 2023-04-24 03:45:00 255700 17754.50 17612.50 17707.55 17743.40  17743.40
#> 201 2023-04-25 03:45:00 297400 17807.45 17716.85 17761.55 17769.25  17769.25
#> 202 2023-04-26 03:45:00 233000 17827.75 17711.20 17767.30 17813.60  17813.60
#> 203 2023-04-27 03:45:00 241400 17931.60 17797.90 17813.10 17915.05  17915.05
#> 204 2023-04-28 03:45:00 290300 18089.15 17885.30 17950.40 18065.00  18065.00
#> 205 2023-05-02 03:45:00 265200 18180.25 18101.75 18124.80 18147.65  18147.65
#> 206 2023-05-03 03:45:00 226200 18116.35 18042.40 18113.80 18089.85  18089.85
#> 207 2023-05-04 03:45:00 226600 18267.45 18066.70 18081.00 18255.80  18255.80
#> 208 2023-05-05 03:45:00 239700 18216.95 18055.45 18117.30 18069.00  18069.00
#> 209 2023-05-08 03:45:00 228400 18286.95 18100.30 18120.60 18264.40  18264.40
#> 210 2023-05-09 03:45:00 245400 18344.20 18229.65 18303.40 18265.95  18265.95
#> 211 2023-05-10 03:45:00 220500 18326.75 18211.95 18313.60 18315.10  18315.10
#> 212 2023-05-11 03:45:00 248900 18389.70 18270.40 18357.80 18297.00  18297.00
#> 213 2023-05-12 03:45:00 241700 18342.75 18194.55 18273.75 18314.80  18314.80
#> 214 2023-05-15 03:45:00 216100 18458.90 18287.90 18339.30 18398.85  18398.85
#> 215 2023-05-16 03:45:00 219500 18432.35 18264.35 18432.35 18286.50  18286.50
#> 216 2023-05-17 03:45:00 229900 18309.00 18115.35 18300.45 18181.75  18181.75
#> 217 2023-05-18 03:45:00 272100 18297.20 18104.85 18287.50 18129.95  18129.95
#> 218 2023-05-19 03:45:00 260900 18218.10 18060.40 18186.15 18203.40  18203.40
#> 219 2023-05-22 03:45:00 262600 18335.25 18178.85 18201.10 18314.40  18314.40
#> 220 2023-05-23 03:45:00 270400 18419.75 18324.20 18362.90 18348.00  18348.00
#> 221 2023-05-24 03:45:00 232600 18392.60 18262.95 18294.80 18285.40  18285.40
#> 222 2023-05-25 03:45:00 235900 18338.10 18202.40 18268.90 18321.15  18321.15
#> 223 2023-05-26 03:45:00 198700 18508.55 18333.15 18368.35 18499.35  18499.35
#> 224 2023-05-29 03:45:00 265300 18641.20 18581.25 18619.15 18598.65  18598.65
#> 225 2023-05-30 03:45:00 234400 18662.45 18575.50 18606.65 18633.85  18633.85
#> 226 2023-05-31 03:45:00 696500 18603.90 18483.85 18594.20 18534.40  18534.40
#> 227 2023-06-01 03:45:00 265600 18580.30 18464.55 18579.40 18487.75  18487.75
#> 228 2023-06-02 03:45:00 270900 18573.70 18478.40 18550.85 18534.10  18534.10
#> 229 2023-06-05 03:45:00 256100 18640.15 18582.80 18612.00 18593.85  18593.85
#> 230 2023-06-06 03:45:00 220800 18622.75 18531.60 18600.80 18599.00  18599.00
#> 231 2023-06-07 03:45:00 260100 18738.95 18636.00 18665.60 18726.40  18726.40
#> 232 2023-06-08 03:45:00 286500 18777.90 18615.60 18725.35 18634.55  18634.55
#> 233 2023-06-09 03:45:00 221800 18676.65 18555.40 18655.90 18563.40  18563.40
#> 234 2023-06-12 03:45:00 179500 18633.60 18559.75 18595.05 18601.50  18601.50
#> 235 2023-06-13 03:45:00 233200 18728.90 18631.80 18631.80 18716.15  18716.15
#> 236 2023-06-14 03:45:00 261400 18769.70 18690.00 18744.60 18755.90  18755.90
#> 237 2023-06-15 03:45:00 263000 18794.10 18669.05 18774.45 18688.10  18688.10
#> 238 2023-06-16 03:45:00 272800 18864.70 18710.50 18723.30 18826.00  18826.00
#> 239 2023-06-19 03:45:00 219800 18881.45 18719.15 18873.30 18755.45  18755.45
#> 240 2023-06-20 03:45:00 211600 18839.70 18660.65 18752.35 18816.70  18816.70
#> 241 2023-06-21 03:45:00 217500 18875.90 18794.85 18849.40 18856.85  18856.85
#> 242 2023-06-22 03:45:00 252700 18886.60 18759.50 18853.60 18771.25  18771.25
#> 243 2023-06-23 03:45:00 210600 18756.40 18647.10 18741.85 18665.50  18665.50
#> 244 2023-06-26 03:45:00 171300 18722.05 18646.70 18682.35 18691.20  18691.20
#> 245 2023-06-27 03:45:00 222500 18829.25 18714.25 18748.55 18817.40  18817.40
#> 246 2023-06-28 03:45:00 398800 19011.25 18861.35 18908.15 18972.10  18972.10
#> 247 2023-06-30 03:45:00 247900 19201.70 19024.60 19076.85 19189.05  19189.05
#> 248 2023-07-03 03:45:00 226100 19345.10 19234.40 19246.50 19322.55  19322.55
#> 249 2023-07-04 03:45:00 226900 19434.15 19300.00 19406.60 19389.00  19389.00
#> 250 2023-07-05 03:45:00 224400 19421.60 19339.60 19405.95 19398.50  19398.50
#> 251 2023-07-06 03:45:00 268300 19512.20 19373.00 19385.70 19497.30  19497.30
#> 252 2023-07-07 03:45:00 237100 19523.60 19303.60 19422.80 19331.80  19331.80
#> 253 2023-07-10 03:45:00 268200 19435.85 19327.10 19400.35 19355.90  19355.90
#> 254 2023-07-11 03:45:00 251300 19515.10 19406.45 19427.10 19439.40  19439.40
#> 255 2023-07-12 03:45:00 327900 19507.70 19361.75 19497.45 19384.30  19384.30
#> 256 2023-07-13 03:45:00 310400 19567.00 19385.80 19495.20 19413.75  19413.75
#> 257 2023-07-14 03:45:00 291200 19595.35 19433.50 19493.45 19564.50  19564.50
#> 258 2023-07-17 03:45:00 268700 19731.85 19562.95 19612.15 19711.45  19711.45
#> 259 2023-07-18 03:45:00 286400 19819.45 19690.20 19787.50 19749.25  19749.25
#> 260 2023-07-19 03:45:00 259700 19851.70 19727.45 19802.95 19833.15  19833.15
#> 261 2023-07-20 03:45:00 274700 19991.85 19758.40 19831.70 19979.15  19979.15
#> 262 2023-07-21 03:45:00 312500 19887.40 19700.00 19800.45 19745.00  19745.00
#> 263 2023-07-24 03:45:00 265600 19782.75 19658.30 19748.45 19672.35  19672.35
#> 264 2023-07-25 03:45:00 369700 19729.35 19615.95 19729.35 19680.60  19680.60
#> 265 2023-07-26 03:45:00 244700 19825.60 19716.70 19733.35 19778.30  19778.30
#> 266 2023-07-27 03:45:00 334900 19867.55 19603.55 19850.90 19659.90  19659.90
#> 267 2023-07-28 03:45:00 258700 19695.90 19563.10 19659.75 19646.05  19646.05
#> 268 2023-07-31 03:45:00 295100 19772.75 19597.60 19666.35 19753.80  19753.80
#> 269 2023-08-01 03:45:00 298000 19795.60 19704.60 19784.00 19733.55  19733.55
#> 270 2023-08-02 03:45:00 290700 19678.25 19423.55 19655.40 19526.55  19526.55
#> 271 2023-08-03 03:45:00 315700 19537.65 19296.45 19463.75 19381.65  19381.65
#> 272 2023-08-04 03:45:00 280800 19538.85 19436.45 19462.80 19517.00  19517.00
#> 273 2023-08-07 03:45:00 216600 19620.45 19524.80 19576.85 19597.30  19597.30
#> 274 2023-08-08 03:45:00 260300 19634.40 19533.10 19627.20 19570.85  19570.85
#> 275 2023-08-09 03:45:00 266500 19645.50 19467.50 19578.80 19632.55  19632.55
#> 276 2023-08-10 03:45:00 312200 19623.60 19495.40 19605.55 19543.10  19543.10
#> 277 2023-08-11 03:45:00 235500 19557.75 19412.75 19554.25 19428.30  19428.30
nifty$get_history(start = '2022-07-01', end = '2022-07-14', interval = '1d')
#>                  date volume     high      low     open    close adj_close
#> 1 2022-07-01 03:45:00 364100 15793.95 15511.05 15703.70 15752.05  15752.05
#> 2 2022-07-04 03:45:00 304300 15852.35 15661.80 15710.50 15835.35  15835.35
#> 3 2022-07-05 03:45:00 254200 16025.75 15785.45 15909.15 15810.85  15810.85
#> 4 2022-07-06 03:45:00 288400 16011.35 15800.90 15818.20 15989.80  15989.80
#> 5 2022-07-07 03:45:00 264600 16150.50 16045.95 16113.75 16132.90  16132.90
#> 6 2022-07-08 03:45:00 281100 16275.50 16157.90 16273.65 16220.60  16220.60
#> 7 2022-07-11 03:45:00 255900 16248.55 16115.50 16136.15 16216.00  16216.00
#> 8 2022-07-12 03:45:00 208600 16158.75 16031.15 16126.20 16058.30  16058.30
#> 9 2022-07-13 03:45:00 233300 16140.00 15950.15 16128.20 15966.65  15966.65
nifty$get_history(period = '1mo', interval = '1d')
#>                   date volume     high      low     open    close adj_close
#> 1  2023-07-11 03:45:00 251300 19515.10 19406.45 19427.10 19439.40  19439.40
#> 2  2023-07-12 03:45:00 327900 19507.70 19361.75 19497.45 19384.30  19384.30
#> 3  2023-07-13 03:45:00 310400 19567.00 19385.80 19495.20 19413.75  19413.75
#> 4  2023-07-14 03:45:00 291200 19595.35 19433.50 19493.45 19564.50  19564.50
#> 5  2023-07-17 03:45:00 268700 19731.85 19562.95 19612.15 19711.45  19711.45
#> 6  2023-07-18 03:45:00 286400 19819.45 19690.20 19787.50 19749.25  19749.25
#> 7  2023-07-19 03:45:00 259700 19851.70 19727.45 19802.95 19833.15  19833.15
#> 8  2023-07-20 03:45:00 274700 19991.85 19758.40 19831.70 19979.15  19979.15
#> 9  2023-07-21 03:45:00 312500 19887.40 19700.00 19800.45 19745.00  19745.00
#> 10 2023-07-24 03:45:00 265600 19782.75 19658.30 19748.45 19672.35  19672.35
#> 11 2023-07-25 03:45:00 369700 19729.35 19615.95 19729.35 19680.60  19680.60
#> 12 2023-07-26 03:45:00 244700 19825.60 19716.70 19733.35 19778.30  19778.30
#> 13 2023-07-27 03:45:00 334900 19867.55 19603.55 19850.90 19659.90  19659.90
#> 14 2023-07-28 03:45:00 258700 19695.90 19563.10 19659.75 19646.05  19646.05
#> 15 2023-07-31 03:45:00 295100 19772.75 19597.60 19666.35 19753.80  19753.80
#> 16 2023-08-01 03:45:00 298000 19795.60 19704.60 19784.00 19733.55  19733.55
#> 17 2023-08-02 03:45:00 290700 19678.25 19423.55 19655.40 19526.55  19526.55
#> 18 2023-08-03 03:45:00 315700 19537.65 19296.45 19463.75 19381.65  19381.65
#> 19 2023-08-04 03:45:00 280800 19538.85 19436.45 19462.80 19517.00  19517.00
#> 20 2023-08-07 03:45:00 216600 19620.45 19524.80 19576.85 19597.30  19597.30
#> 21 2023-08-08 03:45:00 260300 19634.40 19533.10 19627.20 19570.85  19570.85
#> 22 2023-08-09 03:45:00 266500 19645.50 19467.50 19578.80 19632.55  19632.55
#> 23 2023-08-10 03:45:00 312200 19623.60 19495.40 19605.55 19543.10  19543.10
#> 24 2023-08-11 03:45:00 235500 19557.75 19412.75 19554.25 19428.30  19428.30
# }