R6 Class Representing a Ticker

Base class for getting all data related to indices from Yahoo Finance API.

Format

An R6 class object

Public fields

index

Index for which data is retrieved

Methods

Public methods

Index$new()

Index$set_index()

Index$get_history()

Index$clone()

Method new()

Create a new Index object

Usage

Index$new(index = NA)

Arguments

index

Index

Returns

A new `Index` object

Examples

nifty_50 <- Index$new('^NSEI')

Method set_index()

Set a new index.

Usage

Index$set_index(index)

Arguments

index

New index

Examples

indice <- Index$new('^NSEI')
indice$set_index('^NDX')

Method get_history()

Retrieves historical data

Usage

Index$get_history(period = "ytd", interval = "1d", start = NULL, end = NULL)

Arguments

period

Length of time. Defaults to 'ytd'. Valid values are:

'1d'

'5d'

'1mo'

'3mo'

'6mo'

'1y'

'2y'

'5y'

'10y'

'ytd'

'max'

interval

Time between data points. Defaults to '1d'. Valid values are:

'1m'

'2m'

'5m'

'15m'

'30m'

'60m'

'90m'

'1h'

'1d'

'5d'

'1wk'

'1mo'

'3mo'

start

Specific starting date. String or date object in yyyy-mm-dd format.

end

Specific ending date. String or date object in yyyy-mm-dd format.

Returns

A data.frame.

Examples

\donttest{
nifty <- Index$new('^NSEI')
nifty$get_history(start = '2022-07-01', interval = '1d')
nifty$get_history(start = '2022-07-01', end = '2022-07-14', interval = '1d')
nifty$get_history(period = '1mo', interval = '1d')
}

Method clone()

The objects of this class are cloneable with this method.

Usage

Index$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `Index$new`
## ------------------------------------------------

nifty_50 <- Index$new('^NSEI')

## ------------------------------------------------
## Method `Index$set_index`
## ------------------------------------------------

indice <- Index$new('^NSEI')
indice$set_index('^NDX')

## ------------------------------------------------
## Method `Index$get_history`
## ------------------------------------------------

# \donttest{
nifty <- Index$new('^NSEI')
nifty$get_history(start = '2022-07-01', interval = '1d')
#>                    date  volume     high      low     open    close adj_close
#> 1   2022-07-01 03:45:00  364100 15793.95 15511.05 15703.70 15752.05  15752.05
#> 2   2022-07-04 03:45:00  304300 15852.35 15661.80 15710.50 15835.35  15835.35
#> 3   2022-07-05 03:45:00  254200 16025.75 15785.45 15909.15 15810.85  15810.85
#> 4   2022-07-06 03:45:00  288400 16011.35 15800.90 15818.20 15989.80   15989.8
#> 5   2022-07-07 03:45:00  264600 16150.50 16045.95 16113.75 16132.90   16132.9
#> 6   2022-07-08 03:45:00  281100 16275.50 16157.90 16273.65 16220.60   16220.6
#> 7   2022-07-11 03:45:00  255900 16248.55 16115.50 16136.15 16216.00     16216
#> 8   2022-07-12 03:45:00  208600 16158.75 16031.15 16126.20 16058.30   16058.3
#> 9   2022-07-13 03:45:00  233300 16140.00 15950.15 16128.20 15966.65  15966.65
#> 10  2022-07-14 03:45:00  228600 16070.85 15858.20 16018.85 15938.65  15938.65
#> 11  2022-07-15 03:45:00  205800 16066.95 15927.30 16010.80 16049.20   16049.2
#> 12  2022-07-18 03:45:00  227700 16287.95 16142.20 16151.40 16278.50   16278.5
#> 13  2022-07-19 03:45:00  222600 16359.50 16187.05 16187.05 16340.55  16340.55
#> 14  2022-07-20 03:45:00  284400 16588.00 16490.95 16562.80 16520.85  16520.85
#> 15  2022-07-21 03:45:00  243600 16626.95 16483.90 16523.55 16605.25  16605.25
#> 16  2022-07-22 03:45:00  190900 16752.25 16610.90 16661.25 16719.45  16719.45
#> 17  2022-07-25 03:45:00  211400 16706.05 16564.25 16662.55 16631.00     16631
#> 18  2022-07-26 03:45:00  208200 16636.10 16463.30 16632.90 16483.85  16483.85
#> 19  2022-07-27 03:45:00  185200 16653.45 16438.75 16475.35 16641.80   16641.8
#> 20  2022-07-28 03:45:00  376300 16947.65 16746.25 16774.85 16929.60   16929.6
#> 21  2022-07-29 03:45:00  436100 17172.80 17018.15 17079.50 17158.25  17158.25
#> 22  2022-08-01 03:45:00  381500 17356.25 17154.80 17243.20 17340.05  17340.05
#> 23  2022-08-02 03:45:00  333600 17390.15 17215.85 17310.15 17345.45  17345.45
#> 24  2022-08-03 03:45:00  285400 17407.50 17225.85 17349.25 17388.15  17388.15
#> 25  2022-08-04 03:45:00  325000 17490.70 17161.25 17463.10 17382.00     17382
#> 26  2022-08-05 03:45:00  355300 17474.40 17348.75 17423.65 17397.50   17397.5
#> 27  2022-08-08 03:45:00  257000 17548.80 17359.75 17401.50 17525.10   17525.1
#> 28  2022-08-10 03:45:00  312700 17566.10 17442.80 17566.10 17534.75  17534.75
#> 29  2022-08-11 03:45:00  311200 17719.30 17631.95 17711.65 17659.00     17659
#> 30  2022-08-12 03:45:00  303900 17724.65 17597.85 17659.65 17698.15  17698.15
#> 31  2022-08-16 03:45:00       0 17839.10 17764.05 17797.20 17825.25  17825.25
#> 32  2022-08-17 03:45:00  262800 17965.95 17833.35 17868.15 17944.25  17944.25
#> 33  2022-08-18 03:45:00  263900 17968.45 17852.05 17898.65 17956.50   17956.5
#> 34  2022-08-19 03:45:00  295600 17992.20 17710.75 17966.55 17758.45  17758.45
#> 35  2022-08-22 03:45:00  287600 17690.05 17467.35 17682.90 17490.70   17490.7
#> 36  2022-08-23 03:45:00  285600 17625.55 17345.20 17357.35 17577.50   17577.5
#> 37  2022-08-24 03:45:00  262000 17623.65 17499.25 17525.45 17604.95  17604.95
#> 38  2022-08-25 03:45:00  230200 17726.50 17487.45 17679.00 17522.45  17522.45
#> 39  2022-08-26 03:45:00  266600 17685.85 17519.35 17619.30 17558.90   17558.9
#> 40  2022-08-29 03:45:00  244900 17380.15 17166.20 17188.65 17312.90   17312.9
#> 41  2022-08-30 03:45:00  324700 17777.65 17401.50 17414.95 17759.30   17759.3
#> 42  2022-09-01 03:45:00  308500 17695.60 17468.45 17485.70 17542.80   17542.8
#> 43  2022-09-02 03:45:00  256300 17643.85 17476.45 17598.40 17539.45  17539.45
#> 44  2022-09-05 03:45:00  230300 17683.15 17540.35 17546.45 17665.80   17665.8
#> 45  2022-09-06 03:45:00  251200 17764.65 17587.65 17695.70 17655.60   17655.6
#> 46  2022-09-07 03:45:00  354100 17650.75 17484.30 17519.40 17624.40   17624.4
#> 47  2022-09-08 03:45:00  279800 17807.65 17691.95 17748.15 17798.75  17798.75
#> 48  2022-09-09 03:45:00  270300 17925.95 17786.00 17923.35 17833.35  17833.35
#> 49  2022-09-12 03:45:00  228200 17980.55 17889.15 17890.85 17936.35  17936.35
#> 50  2022-09-13 03:45:00  259900 18088.30 18015.45 18044.45 18070.05  18070.05
#> 51  2022-09-14 03:45:00  365900 18091.55 17771.15 17771.15 18003.75  18003.75
#> 52  2022-09-15 03:45:00  289600 18096.15 17861.50 18046.35 17877.40   17877.4
#> 53  2022-09-16 03:45:00  468500 17820.05 17497.25 17796.80 17530.85  17530.85
#> 54  2022-09-19 03:45:00  258300 17667.20 17429.70 17540.65 17622.25  17622.25
#> 55  2022-09-20 03:45:00  263100 17919.30 17744.40 17770.40 17816.25  17816.25
#> 56  2022-09-21 03:45:00  245500 17838.70 17663.60 17766.35 17718.35  17718.35
#> 57  2022-09-22 03:45:00  284100 17722.75 17532.45 17609.65 17629.80   17629.8
#> 58  2022-09-23 03:45:00  390800 17642.15 17291.65 17593.85 17327.35  17327.35
#> 59  2022-09-26 03:45:00  492000 17196.40 16978.30 17156.30 17016.30   17016.3
#> 60  2022-09-27 03:45:00  359900 17176.45 16942.35 17110.90 17007.40   17007.4
#> 61  2022-09-28 03:45:00  323900 17037.60 16820.40 16870.55 16858.60   16858.6
#> 62  2022-09-29 03:45:00  340000 17026.05 16788.60 16993.60 16818.10   16818.1
#> 63  2022-09-30 03:45:00  376100 17187.10 16747.70 16798.05 17094.35  17094.35
#> 64  2022-10-03 03:45:00  278400 17114.65 16855.55 17102.10 16887.35  16887.35
#> 65  2022-10-04 03:45:00  226000 17287.30 17117.30 17147.45 17274.30   17274.3
#> 66  2022-10-06 03:45:00  265500 17428.80 17315.65 17379.25 17331.80   17331.8
#> 67  2022-10-07 03:45:00  216300 17337.35 17216.95 17287.20 17314.65  17314.65
#> 68  2022-10-10 03:45:00  234000 17280.15 17064.70 17094.35 17241.00     17241
#> 69  2022-10-11 03:45:00  282600 17261.80 16950.30 17256.05 16983.55  16983.55
#> 70  2022-10-12 03:45:00  256000 17142.35 16960.05 17025.55 17123.60   17123.6
#> 71  2022-10-13 03:45:00  266400 17112.35 16956.95 17087.35 17014.35  17014.35
#> 72  2022-10-14 03:45:00  227000 17348.55 17169.75 17322.30 17185.70   17185.7
#> 73  2022-10-17 03:45:00  212200 17328.55 17098.55 17144.80 17311.80   17311.8
#> 74  2022-10-18 03:45:00  239500 17527.80 17434.05 17438.75 17486.95  17486.95
#> 75  2022-10-19 03:45:00  210500 17607.60 17472.85 17568.15 17512.25  17512.25
#> 76  2022-10-20 03:45:00  249600 17584.15 17421.00 17423.10 17563.95  17563.95
#> 77  2022-10-21 03:45:00  277700 17670.15 17520.75 17622.85 17576.30   17576.3
#> 78  2022-10-24 12:45:00   45000 17777.55 17707.40 17736.35 17730.75  17730.75
#> 79  2022-10-25 03:45:00  251400 17811.50 17637.00 17808.30 17656.35  17656.35
#> 80  2022-10-27 03:45:00  324600 17783.90 17654.50 17771.40 17736.95  17736.95
#> 81  2022-10-28 03:45:00  250000 17838.90 17723.70 17756.40 17786.80   17786.8
#> 82  2022-10-31 03:45:00  227200 18022.80 17899.90 17910.20 18012.20   18012.2
#> 83  2022-11-01 03:45:00  349900 18175.80 18060.15 18130.70 18145.40   18145.4
#> 84  2022-11-02 03:45:00  270900 18178.75 18048.65 18177.90 18082.85  18082.85
#> 85  2022-11-03 03:45:00  213000 18106.30 17959.20 17968.35 18052.70   18052.7
#> 86  2022-11-04 03:45:00  267900 18135.10 18017.15 18053.40 18117.15  18117.15
#> 87  2022-11-07 03:45:00  314800 18255.50 18064.75 18211.75 18202.80   18202.8
#> 88  2022-11-09 03:45:00  307200 18296.40 18117.50 18288.25 18157.00     18157
#> 89  2022-11-10 03:45:00  256500 18103.10 17969.40 18044.35 18028.20   18028.2
#> 90  2022-11-11 03:45:00  378500 18362.30 18259.35 18272.35 18349.70   18349.7
#> 91  2022-11-14 03:45:00  301400 18399.45 18311.40 18376.40 18329.15  18329.15
#> 92  2022-11-15 03:45:00  250900 18427.95 18282.00 18362.75 18403.40   18403.4
#> 93  2022-11-16 03:45:00  219300 18442.15 18344.15 18398.25 18409.65  18409.65
#> 94  2022-11-17 03:45:00  200500 18417.60 18312.95 18358.70 18343.90   18343.9
#> 95  2022-11-18 03:45:00  198800 18394.60 18209.80 18382.95 18307.65  18307.65
#> 96  2022-11-21 03:45:00  213800 18262.30 18133.35 18246.40 18159.95  18159.95
#> 97  2022-11-22 03:45:00  177700 18261.85 18137.70 18179.15 18244.20   18244.2
#> 98  2022-11-23 03:45:00  178800 18325.40 18246.00 18325.20 18267.25  18267.25
#> 99  2022-11-24 03:45:00  204000 18529.70 18294.25 18326.10 18484.10   18484.1
#> 100 2022-11-25 03:45:00  205900 18534.90 18445.10 18528.45 18512.75  18512.75
#> 101 2022-11-28 03:45:00  206500 18614.25 18365.60 18430.55 18562.75  18562.75
#> 102 2022-11-29 03:45:00  195300 18678.10 18552.15 18552.45 18618.05  18618.05
#> 103 2022-11-30 03:45:00  345600 18816.05 18616.55 18625.70 18758.35  18758.35
#> 104 2022-12-01 03:45:00  325000 18887.60 18778.20 18871.95 18812.50   18812.5
#> 105 2022-12-02 03:45:00  254400 18781.95 18639.20 18752.40 18696.10   18696.1
#> 106 2022-12-05 03:45:00  288400 18728.60 18591.35 18719.55 18701.05  18701.05
#> 107 2022-12-06 03:45:00  217800 18654.90 18577.90 18600.65 18642.75  18642.75
#> 108 2022-12-07 03:45:00  200500 18668.30 18528.40 18638.85 18560.50   18560.5
#> 109 2022-12-08 03:45:00  202800 18625.00 18536.95 18570.85 18609.35  18609.35
#> 110 2022-12-09 03:45:00  215800 18664.70 18410.10 18662.40 18496.60   18496.6
#> 111 2022-12-12 03:45:00  193900 18521.55 18345.70 18402.15 18497.15  18497.15
#> 112 2022-12-13 03:45:00  187600 18617.25 18490.20 18524.40 18608.00     18608
#> 113 2022-12-14 03:45:00  208000 18696.10 18632.90 18671.25 18660.30   18660.3
#> 114 2022-12-15 03:45:00  183100 18652.90 18387.70 18614.40 18414.90   18414.9
#> 115 2022-12-16 03:45:00  277700 18440.95 18255.15 18319.10 18269.00     18269
#> 116 2022-12-19 03:45:00  154800 18431.65 18244.55 18288.10 18420.45  18420.45
#> 117 2022-12-20 03:45:00  169000 18404.90 18202.65 18340.30 18385.30   18385.3
#> 118 2022-12-21 03:45:00  187600 18473.35 18162.75 18435.15 18199.10   18199.1
#> 119 2022-12-22 03:45:00  195700 18318.75 18068.60 18288.80 18127.35  18127.35
#> 120 2022-12-23 03:45:00  221500 18050.45 17779.50 17977.65 17806.80   17806.8
#> 121 2022-12-26 03:45:00  176700 18084.10 17774.25 17830.40 18014.60   18014.6
#> 122 2022-12-27 03:45:00  214300 18149.25 17967.45 18089.80 18132.30   18132.3
#> 123 2022-12-28 03:45:00  193900 18173.10 18068.35 18084.75 18122.50   18122.5
#> 124 2022-12-29 03:45:00  281100 18229.70 17992.80 18045.70 18191.00     18191
#> 125 2022-12-30 03:45:00  192000 18265.25 18080.30 18259.10 18105.30   18105.3
#> 126 2023-01-02 03:45:00  256100 18215.15 18086.50 18131.70 18197.45  18197.45
#> 127 2023-01-03 03:45:00  208700 18251.95 18149.80 18163.20 18232.55  18232.55
#> 128 2023-01-04 03:45:00  235200 18243.00 18020.60 18230.65 18042.95  18042.95
#> 129 2023-01-05 03:45:00  269900 18120.30 17892.60 18101.95 17992.15  17992.15
#> 130 2023-01-06 03:45:00  238200 18047.40 17795.55 18008.05 17859.45  17859.45
#> 131 2023-01-09 03:45:00  257200 18141.40 17936.15 17952.55 18101.20   18101.2
#> 132 2023-01-10 03:45:00  283300 18127.60 17856.00 18121.30 17914.15  17914.15
#> 133 2023-01-11 03:45:00  259900 17976.35 17824.35 17924.25 17895.70   17895.7
#> 134 2023-01-12 03:45:00  227800 17945.80 17761.65 17920.85 17858.20   17858.2
#> 135 2023-01-13 03:45:00  256700 17999.35 17774.25 17867.50 17956.60   17956.6
#> 136 2023-01-16 03:45:00  206200 18049.65 17853.65 18033.15 17894.85  17894.85
#> 137 2023-01-17 03:45:00  219100 18072.05 17886.95 17922.80 18053.30   18053.3
#> 138 2023-01-18 03:45:00  255800 18183.75 18032.45 18074.30 18165.35  18165.35
#> 139 2023-01-19 03:45:00  237800 18155.20 18063.75 18119.80 18107.85  18107.85
#> 140 2023-01-20 03:45:00  237200 18145.45 18016.20 18115.60 18027.65  18027.65
#> 141 2023-01-23 03:45:00  202500 18162.60 18063.45 18118.45 18118.55  18118.55
#> 142 2023-01-24 03:45:00  216900 18201.25 18078.65 18183.95 18118.30   18118.3
#> 143 2023-01-25 03:45:00  257200 18100.60 17846.15 18093.35 17891.95  17891.95
#> 144 2023-01-27 03:45:00  476300 17884.75 17493.55 17877.20 17604.35  17604.35
#> 145 2023-01-30 03:45:00  432400 17709.15 17405.55 17541.95 17648.95  17648.95
#> 146 2023-01-31 03:45:00  398300 17735.70 17537.55 17731.45 17662.15  17662.15
#> 147 2023-02-01 03:45:00  512900 17972.20 17353.40 17811.60 17616.30   17616.3
#> 148 2023-02-02 03:45:00  490100 17653.90 17445.95 17517.10 17610.40   17610.4
#> 149 2023-02-03 03:45:00  424100 17870.30 17584.20 17721.75 17854.05  17854.05
#> 150 2023-02-06 03:45:00  282500 17823.70 17698.35 17818.55 17764.60   17764.6
#> 151 2023-02-07 03:45:00  354400 17811.15 17652.55 17790.10 17721.50   17721.5
#> 152 2023-02-08 03:45:00  291000 17898.70 17744.15 17750.30 17871.70   17871.7
#> 153 2023-02-09 03:45:00  260900 17916.90 17779.80 17885.50 17893.45  17893.45
#> 154 2023-02-10 03:45:00  232000 17876.95 17801.00 17847.55 17856.50   17856.5
#> 155 2023-02-13 03:45:00  231300 17880.70 17719.75 17859.10 17770.90   17770.9
#> 156 2023-02-14 03:45:00  244500 17954.55 17800.05 17840.35 17929.85  17929.85
#> 157 2023-02-15 03:45:00  229300 18034.10 17853.80 17896.60 18015.85  18015.85
#> 158 2023-02-16 03:45:00  230200 18134.75 18000.65 18094.75 18035.85  18035.85
#> 159 2023-02-17 03:45:00  208100 18034.25 17884.60 17974.85 17944.20   17944.2
#> 160 2023-02-20 03:45:00  174600 18004.35 17818.40 17965.55 17844.60   17844.6
#> 161 2023-02-21 03:45:00  198200 17924.90 17800.30 17905.80 17826.70   17826.7
#> 162 2023-02-22 03:45:00  204000 17772.50 17529.45 17755.35 17554.30   17554.3
#> 163 2023-02-23 03:45:00  240800 17620.05 17455.40 17574.65 17511.25  17511.25
#> 164 2023-02-24 03:45:00  209400 17599.75 17421.80 17591.35 17465.80   17465.8
#> 165 2023-02-27 03:45:00  246400 17451.60 17299.00 17428.60 17392.70   17392.7
#> 166 2023-02-28 03:45:00  420900 17440.45 17255.20 17383.25 17303.95  17303.95
#> 167 2023-03-01 03:45:00  272900 17467.75 17345.25 17360.10 17450.90   17450.9
#> 168 2023-03-02 03:45:00  310700 17445.80 17306.00 17421.50 17321.90   17321.9
#> 169 2023-03-03 03:45:00  356200 17644.75 17427.70 17451.25 17594.35  17594.35
#> 170 2023-03-06 03:45:00  362800 17799.95 17671.95 17680.35 17711.45  17711.45
#> 171 2023-03-08 03:45:00  267000 17766.50 17602.25 17665.75 17754.40   17754.4
#> 172 2023-03-09 03:45:00  262400 17772.35 17573.60 17772.05 17589.60   17589.6
#> 173 2023-03-10 03:45:00  235900 17451.50 17324.35 17443.80 17412.90   17412.9
#> 174 2023-03-13 03:45:00  254900 17529.90 17113.45 17421.90 17154.30   17154.3
#> 175 2023-03-14 03:45:00  290600 17224.65 16987.10 17160.55 17043.30   17043.3
#> 176 2023-03-15 03:45:00  248200 17211.35 16938.90 17166.45 16972.15  16972.15
#> 177 2023-03-16 03:45:00  349800 17062.45 16850.15 16994.65 16985.60   16985.6
#> 178 2023-03-17 03:45:00  408100 17145.80 16958.15 17111.80 17100.05  17100.05
#> 179 2023-03-20 03:45:00  241800 17066.60 16828.35 17066.60 16988.40   16988.4
#> 180 2023-03-21 03:45:00  246700 17127.70 17016.00 17060.40 17107.50   17107.5
#> 181 2023-03-22 03:45:00       0 17207.25 17107.85 17177.45 17151.90   17151.9
#> 182 2023-03-23 03:45:00  219200 17205.40 17045.30 17097.40 17076.90   17076.9
#> 183 2023-03-24 03:45:00  228000 17109.45 16917.35 17076.20 16945.05  16945.05
#> 184 2023-03-27 03:45:00  218400 17091.00 16918.55 16984.30 16985.70   16985.7
#> 185 2023-03-28 03:45:00  238800 17061.75 16913.75 17031.75 16951.70   16951.7
#> 186 2023-03-29 03:45:00  345900 17126.15 16940.60 16977.30 17080.70   17080.7
#> 187 2023-03-31 03:45:00  263900 17381.60 17204.65 17210.35 17359.75  17359.75
#> 188 2023-04-03 03:45:00  230200 17428.05 17312.75 17427.95 17398.05  17398.05
#> 189 2023-04-05 03:45:00  248300 17570.55 17402.70 17422.30 17557.05  17557.05
#> 190 2023-04-06 03:45:00  242700 17638.70 17502.85 17533.85 17599.15  17599.15
#> 191 2023-04-10 03:45:00  254800 17694.10 17597.95 17634.90 17624.05  17624.05
#> 192 2023-04-11 03:45:00  304300 17748.75 17655.15 17704.80 17722.30   17722.3
#> 193 2023-04-12 03:45:00  232800 17825.75 17717.25 17759.55 17812.40   17812.4
#> 194 2023-04-13 03:45:00  285800 17842.15 17729.65 17807.30 17828.00     17828
#> 195 2023-04-17 03:45:00  357400 17863.00 17574.05 17863.00 17706.85  17706.85
#> 196 2023-04-18 03:45:00  258900 17766.60 17610.20 17766.60 17660.15  17660.15
#> 197 2023-04-19 03:45:00  281500 17666.15 17579.85 17653.35 17618.75  17618.75
#> 198 2023-04-20 03:45:00  215100 17684.45 17584.35 17638.60 17624.45  17624.45
#> 199 2023-04-21 03:45:00  230200 17663.20 17553.95 17639.75 17624.05  17624.05
#> 200 2023-04-24 03:45:00  255700 17754.50 17612.50 17707.55 17743.40   17743.4
#> 201 2023-04-25 03:45:00  297400 17807.45 17716.85 17761.55 17769.25  17769.25
#> 202 2023-04-26 03:45:00  233000 17827.75 17711.20 17767.30 17813.60   17813.6
#> 203 2023-04-27 03:45:00  241400 17931.60 17797.90 17813.10 17915.05  17915.05
#> 204 2023-04-28 03:45:00  290300 18089.15 17885.30 17950.40 18065.00     18065
#> 205 2023-05-02 03:45:00  265200 18180.25 18101.75 18124.80 18147.65  18147.65
#> 206 2023-05-03 03:45:00  226200 18116.35 18042.40 18113.80 18089.85  18089.85
#> 207 2023-05-04 03:45:00  226600 18267.45 18066.70 18081.00 18255.80   18255.8
#> 208 2023-05-05 03:45:00  239700 18216.95 18055.45 18117.30 18069.00     18069
#> 209 2023-05-08 03:45:00  228400 18286.95 18100.30 18120.60 18264.40   18264.4
#> 210 2023-05-09 03:45:00  245400 18344.20 18229.65 18303.40 18265.95  18265.95
#> 211 2023-05-10 03:45:00  220500 18326.75 18211.95 18313.60 18315.10   18315.1
#> 212 2023-05-11 03:45:00  248900 18389.70 18270.40 18357.80 18297.00     18297
#> 213 2023-05-12 03:45:00  241700 18342.75 18194.55 18273.75 18314.80   18314.8
#> 214 2023-05-15 03:45:00  216100 18458.90 18287.90 18339.30 18398.85  18398.85
#> 215 2023-05-16 03:45:00  219500 18432.35 18264.35 18432.35 18286.50   18286.5
#> 216 2023-05-17 03:45:00  229900 18309.00 18115.35 18300.45 18181.75  18181.75
#> 217 2023-05-18 03:45:00  272100 18297.20 18104.85 18287.50 18129.95  18129.95
#> 218 2023-05-19 03:45:00  260900 18218.10 18060.40 18186.15 18203.40   18203.4
#> 219 2023-05-22 03:45:00  262600 18335.25 18178.85 18201.10 18314.40   18314.4
#> 220 2023-05-23 03:45:00  270400 18419.75 18324.20 18362.90 18348.00     18348
#> 221 2023-05-24 03:45:00  232600 18392.60 18262.95 18294.80 18285.40   18285.4
#> 222 2023-05-25 03:45:00  235900 18338.10 18202.40 18268.90 18321.15  18321.15
#> 223 2023-05-26 03:45:00  198700 18508.55 18333.15 18368.35 18499.35  18499.35
#> 224 2023-05-29 03:45:00  265300 18641.20 18581.25 18619.15 18598.65  18598.65
#> 225 2023-05-30 03:45:00  234400 18662.45 18575.50 18606.65 18633.85  18633.85
#> 226 2023-05-31 03:45:00  696500 18603.90 18483.85 18594.20 18534.40   18534.4
#> 227 2023-06-01 03:45:00  265600 18580.30 18464.55 18579.40 18487.75  18487.75
#> 228 2023-06-02 03:45:00  270900 18573.70 18478.40 18550.85 18534.10   18534.1
#> 229 2023-06-05 03:45:00  256100 18640.15 18582.80 18612.00 18593.85  18593.85
#> 230 2023-06-06 03:45:00  220800 18622.75 18531.60 18600.80 18599.00     18599
#> 231 2023-06-07 03:45:00  260100 18738.95 18636.00 18665.60 18726.40   18726.4
#> 232 2023-06-08 03:45:00  286500 18777.90 18615.60 18725.35 18634.55  18634.55
#> 233 2023-06-09 03:45:00  221800 18676.65 18555.40 18655.90 18563.40   18563.4
#> 234 2023-06-12 03:45:00  179500 18633.60 18559.75 18595.05 18601.50   18601.5
#> 235 2023-06-13 03:45:00  233200 18728.90 18631.80 18631.80 18716.15  18716.15
#> 236 2023-06-14 03:45:00  261400 18769.70 18690.00 18744.60 18755.90   18755.9
#> 237 2023-06-15 03:45:00  263000 18794.10 18669.05 18774.45 18688.10   18688.1
#> 238 2023-06-16 03:45:00  272800 18864.70 18710.50 18723.30 18826.00     18826
#> 239 2023-06-19 03:45:00  219800 18881.45 18719.15 18873.30 18755.45  18755.45
#> 240 2023-06-20 03:45:00  211600 18839.70 18660.65 18752.35 18816.70   18816.7
#> 241 2023-06-21 03:45:00  217500 18875.90 18794.85 18849.40 18856.85  18856.85
#> 242 2023-06-22 03:45:00  252700 18886.60 18759.50 18853.60 18771.25  18771.25
#> 243 2023-06-23 03:45:00  210600 18756.40 18647.10 18741.85 18665.50   18665.5
#> 244 2023-06-26 03:45:00  171300 18722.05 18646.70 18682.35 18691.20   18691.2
#> 245 2023-06-27 03:45:00  222500 18829.25 18714.25 18748.55 18817.40   18817.4
#> 246 2023-06-28 03:45:00  398800 19011.25 18861.35 18908.15 18972.10   18972.1
#> 247 2023-06-30 03:45:00  247900 19201.70 19024.60 19076.85 19189.05  19189.05
#> 248 2023-07-03 03:45:00  226100 19345.10 19234.40 19246.50 19322.55  19322.55
#> 249 2023-07-04 03:45:00  226900 19434.15 19300.00 19406.60 19389.00     19389
#> 250 2023-07-05 03:45:00  224400 19421.60 19339.60 19405.95 19398.50   19398.5
#> 251 2023-07-06 03:45:00  268300 19512.20 19373.00 19385.70 19497.30   19497.3
#> 252 2023-07-07 03:45:00  237100 19523.60 19303.60 19422.80 19331.80   19331.8
#> 253 2023-07-10 03:45:00  268200 19435.85 19327.10 19400.35 19355.90   19355.9
#> 254 2023-07-11 03:45:00  251300 19515.10 19406.45 19427.10 19439.40   19439.4
#> 255 2023-07-12 03:45:00  327900 19507.70 19361.75 19497.45 19384.30   19384.3
#> 256 2023-07-13 03:45:00  310400 19567.00 19385.80 19495.20 19413.75  19413.75
#> 257 2023-07-14 03:45:00  291200 19595.35 19433.50 19493.45 19564.50   19564.5
#> 258 2023-07-17 03:45:00  268700 19731.85 19562.95 19612.15 19711.45  19711.45
#> 259 2023-07-18 03:45:00  286400 19819.45 19690.20 19787.50 19749.25  19749.25
#> 260 2023-07-19 03:45:00  259700 19851.70 19727.45 19802.95 19833.15  19833.15
#> 261 2023-07-20 03:45:00  274700 19991.85 19758.40 19831.70 19979.15  19979.15
#> 262 2023-07-21 03:45:00  312500 19887.40 19700.00 19800.45 19745.00     19745
#> 263 2023-07-24 03:45:00  265600 19782.75 19658.30 19748.45 19672.35  19672.35
#> 264 2023-07-25 03:45:00  369700 19729.35 19615.95 19729.35 19680.60   19680.6
#> 265 2023-07-26 03:45:00  244700 19825.60 19716.70 19733.35 19778.30   19778.3
#> 266 2023-07-27 03:45:00  334900 19867.55 19603.55 19850.90 19659.90   19659.9
#> 267 2023-07-28 03:45:00  258700 19695.90 19563.10 19659.75 19646.05  19646.05
#> 268 2023-07-31 03:45:00  295100 19772.75 19597.60 19666.35 19753.80   19753.8
#> 269 2023-08-01 03:45:00  298000 19795.60 19704.60 19784.00 19733.55  19733.55
#> 270 2023-08-02 03:45:00  290700 19678.25 19423.55 19655.40 19526.55  19526.55
#> 271 2023-08-03 03:45:00  315700 19537.65 19296.45 19463.75 19381.65  19381.65
#> 272 2023-08-04 03:45:00  280800 19538.85 19436.45 19462.80 19517.00     19517
#> 273 2023-08-07 03:45:00  216600 19620.45 19524.80 19576.85 19597.30   19597.3
#> 274 2023-08-08 03:45:00  260300 19634.40 19533.10 19627.20 19570.85  19570.85
#> 275 2023-08-09 03:45:00  266500 19645.50 19467.50 19578.80 19632.55  19632.55
#> 276 2023-08-10 03:45:00  312200 19623.60 19495.40 19605.55 19543.10   19543.1
#> 277 2023-08-11 03:45:00  235500 19557.75 19412.75 19554.25 19428.30   19428.3
#> 278 2023-08-14 03:45:00  243900 19465.85 19257.90 19383.95 19434.55  19434.55
#> 279 2023-08-16 03:45:00       0 19482.75 19317.20 19369.00 19465.00     19465
#> 280 2023-08-17 03:45:00  268700 19461.55 19326.25 19450.55 19365.25  19365.25
#> 281 2023-08-18 03:45:00  256100 19373.80 19253.60 19301.75 19310.15  19310.15
#> 282 2023-08-21 03:45:00  262600 19425.95 19296.30 19320.65 19393.60   19393.6
#> 283 2023-08-22 03:45:00  208700 19443.50 19381.30 19417.10 19396.45  19396.45
#> 284 2023-08-23 03:45:00  225200 19472.05 19366.60 19439.20 19444.00     19444
#> 285 2023-08-24 03:45:00  268300 19584.45 19369.00 19535.15 19386.70   19386.7
#> 286 2023-08-25 03:45:00  466500 19339.55 19229.70 19297.40 19265.80   19265.8
#> 287 2023-08-28 03:45:00  248200 19366.85 19249.70 19298.35 19306.05  19306.05
#> 288 2023-08-29 03:45:00  307400 19377.90 19309.10 19374.85 19342.65  19342.65
#> 289 2023-08-30 03:45:00  233000 19452.80 19334.75 19433.45 19347.45  19347.45
#> 290 2023-08-31 03:45:00  562600 19388.20 19223.65 19375.55 19253.80   19253.8
#> 291 2023-09-01 03:45:00  333000 19458.55 19255.70 19258.15 19435.30   19435.3
#> 292 2023-09-04 03:45:00  296800 19545.15 19432.85 19525.05 19528.80   19528.8
#> 293 2023-09-05 03:45:00  256800 19587.05 19525.75 19564.65 19574.90   19574.9
#> 294 2023-09-06 03:45:00  287600 19636.45 19491.50 19581.20 19611.05  19611.05
#> 295 2023-09-07 03:45:00  304900 19737.00 19550.05 19598.65 19727.05  19727.05
#> 296 2023-09-08 03:45:00  288100 19867.15 19727.05 19774.80 19819.95  19819.95
#> 297 2023-09-11 03:45:00  248800 20008.15 19865.35 19890.00 19996.35  19996.35
#> 298 2023-09-12 03:45:00  328100 20110.35 19914.65 20110.15 19993.20   19993.2
#> 299 2023-09-13 03:45:00  251400 20096.90 19944.10 19989.50 20070.00     20070
#> 300 2023-09-14 03:45:00  275100 20167.65 20043.45 20127.95 20103.10   20103.1
#> 301 2023-09-15 03:45:00  438200 20222.45 20129.70 20156.45 20192.35  20192.35
#> 302 2023-09-18 03:45:00  264900 20195.35 20115.70 20155.95 20133.30   20133.3
#> 303 2023-09-20 03:45:00  378100 20050.65 19878.85 19980.75 19901.40   19901.4
#> 304 2023-09-21 03:45:00  275100 19848.75 19709.95 19840.55 19742.35  19742.35
#> 305 2023-09-22 03:45:00  246400 19798.65 19657.50 19744.85 19674.25  19674.25
#> 306 2023-09-25 03:45:00  188400 19734.15 19601.55 19678.20 19674.55  19674.55
#> 307 2023-09-26 03:45:00  204900 19699.35 19637.45 19682.80 19664.70   19664.7
#> 308 2023-09-27 03:45:00  203600 19730.70 19554.00 19637.05 19716.45  19716.45
#> 309 2023-09-28 03:45:00  352800 19766.65 19492.10 19761.80 19523.55  19523.55
#> 310 2023-09-29 03:45:00  243500 19726.25 19551.05 19581.20 19638.30   19638.3
#> 311 2023-10-03 03:45:00  221700 19623.20 19479.65 19622.40 19528.75  19528.75
#> 312 2023-10-04 03:45:00  277100 19457.80 19333.60 19446.30 19436.10   19436.1
#> 313 2023-10-05 03:45:00  234900 19576.95 19487.30 19521.85 19545.75  19545.75
#> 314 2023-10-06 03:45:00  159100 19675.75 19589.40 19621.20 19653.50   19653.5
#> 315 2023-10-09 03:45:00  165100 19588.95 19480.50 19539.45 19512.35  19512.35
#> 316 2023-10-10 03:45:00  216600 19717.80 19565.45 19565.60 19689.85  19689.85
#> 317 2023-10-11 03:45:00  213700 19839.20 19756.95 19767.00 19811.35  19811.35
#> 318 2023-10-12 03:45:00  217900 19843.30 19772.65 19822.70 19794.00     19794
#> 319 2023-10-13 03:45:00  255000 19805.40 19635.30 19654.55 19751.05  19751.05
#> 320 2023-10-16 03:45:00  181000 19781.30 19691.85 19737.25 19731.75  19731.75
#> 321 2023-10-17 03:45:00  185800 19849.75 19775.65 19843.20 19811.50   19811.5
#> 322 2023-10-18 03:45:00  198900 19840.95 19659.95 19820.45 19671.10   19671.1
#> 323 2023-10-19 03:45:00  230300 19681.80 19512.35 19545.20 19624.70   19624.7
#> 324 2023-10-20 03:45:00  198300 19593.80 19518.70 19542.15 19542.65  19542.65
#> 325 2023-10-23 03:45:00  176000 19556.85 19257.85 19521.60 19281.75  19281.75
#> 326 2023-10-25 03:45:00  225300 19347.30 19074.15 19286.45 19122.15  19122.15
#> 327 2023-10-26 03:45:00  300400 19041.70 18837.85 19027.25 18857.25  18857.25
#> 328 2023-10-27 03:45:00  205200 19076.15 18926.65 18928.75 19047.25  19047.25
#> 329 2023-10-30 03:45:00  180100 19158.50 18940.00 19053.40 19140.90   19140.9
#> 330 2023-10-31 03:45:00  206000 19233.70 19056.45 19232.95 19079.60   19079.6
#> 331 2023-11-01 03:45:00  194100 19096.05 18973.70 19064.05 18989.15  18989.15
#> 332 2023-11-02 03:45:00  245600 19175.25 19064.15 19120.00 19133.25  19133.25
#> 333 2023-11-03 03:45:00  189500 19276.25 19210.90 19241.00 19230.60   19230.6
#> 334 2023-11-06 03:45:00  180700 19423.00 19309.70 19345.85 19411.75  19411.75
#> 335 2023-11-07 03:45:00  198700 19423.50 19329.10 19404.05 19406.70   19406.7
#> 336 2023-11-08 03:45:00  195000 19464.40 19401.50 19449.60 19443.50   19443.5
#> 337 2023-11-09 03:45:00  208400 19463.90 19378.35 19457.40 19395.30   19395.3
#> 338 2023-11-10 03:45:00  152200 19451.30 19329.45 19351.85 19425.35  19425.35
#> 339 2023-11-13 03:45:00  189300 19494.40 19414.75 19486.75 19443.55  19443.55
#> 340 2023-11-15 03:45:00  291500 19693.20 19579.65 19651.40 19675.45  19675.45
#> 341 2023-11-16 03:45:00  282700 19875.25 19627.00 19674.70 19765.20   19765.2
#> 342 2023-11-17 03:45:00  236800 19806.00 19667.45 19674.75 19731.80   19731.8
#> 343 2023-11-20 03:45:00  173800 19756.45 19670.50 19731.15 19694.00     19694
#> 344 2023-11-21 03:45:00  195000 19829.10 19754.05 19770.90 19783.40   19783.4
#> 345 2023-11-22 03:45:00  184500 19825.55 19703.85 19784.00 19811.85  19811.85
#> 346 2023-11-23 03:45:00  170000 19875.15 19786.75 19828.45 19802.00     19802
#> 347 2023-11-24 03:45:00  145900 19832.85 19768.85 19809.60 19794.70   19794.7
#> 348 2023-11-28 03:45:00  229600 19916.85 19800.00 19844.65 19889.70   19889.7
#> 349 2023-11-29 03:45:00  236200 20104.65 19956.30 19976.55 20096.60   20096.6
#> 350 2023-11-30 03:45:00  486600 20158.70 20015.85 20108.50 20133.15  20133.15
#> 351 2023-12-01 03:45:00  265800 20291.55 20183.70 20194.10 20267.90   20267.9
#> 352 2023-12-04 03:45:00  381100 20702.65 20507.75 20601.95 20686.80   20686.8
#> 353 2023-12-05 03:45:00  421000 20864.05 20711.15 20808.90 20855.10   20855.1
#> 354 2023-12-06 03:45:00  359200 20961.95 20852.15 20950.75 20937.70   20937.7
#> 355 2023-12-07 03:45:00  275000 20941.25 20850.80 20932.40 20901.15  20901.15
#> 356 2023-12-08 03:45:00  292400 21006.10 20862.70 20934.10 20969.40   20969.4
#> 357 2023-12-11 03:45:00  212900 21026.10 20923.70 20965.30 20997.10   20997.1
#> 358 2023-12-12 03:45:00  244900 21037.90 20867.15 21018.55 20906.40   20906.4
#> 359 2023-12-13 03:45:00  260300 20950.00 20769.50 20929.75 20926.35  20926.35
#> 360 2023-12-14 03:45:00  334400 21210.90 21074.45 21110.40 21182.70   21182.7
#> 361 2023-12-15 03:45:00  508800 21492.30 21235.30 21287.45 21456.65  21456.65
#> 362 2023-12-18 03:45:00  249700 21482.80 21365.35 21434.80 21418.65  21418.65
#> 363 2023-12-19 03:45:00  249800 21505.05 21337.75 21477.65 21453.10   21453.1
#> 364 2023-12-20 03:45:00  363900 21593.00 21087.35 21543.50 21150.15  21150.15
#> 365 2023-12-21 03:45:00  277500 21288.35 20976.80 21033.95 21255.05  21255.05
#> 366 2023-12-22 03:45:00  284700 21390.50 21232.45 21295.85 21349.40   21349.4
#> 367 2023-12-26 03:45:00  219500 21477.15 21329.45 21365.20 21441.35  21441.35
#> 368 2023-12-27 03:45:00  256500 21675.75 21495.80 21497.65 21654.75  21654.75
#> 369 2023-12-28 03:45:00  393100 21801.45 21678.00 21715.00 21778.70   21778.7
#> 370 2023-12-29 03:45:00  270900 21770.30 21676.90 21737.65 21731.40   21731.4
#> 371 2024-01-01 03:45:00  154000 21834.35 21680.85 21727.75 21741.90   21741.9
#> 372 2024-01-02 03:45:00  263700 21755.60 21555.65 21751.35 21665.80   21665.8
#> 373 2024-01-03 03:45:00  311900 21677.00 21500.35 21661.10 21517.35  21517.35
#> 374 2024-01-04 03:45:00  339200 21685.65 21564.55 21605.80 21658.60   21658.6
#> 375 2024-01-05 03:45:00  309300 21749.60 21629.20 21705.75 21710.80   21710.8
#> 376 2024-01-08 03:45:00  231500 21763.95 21492.90 21747.60 21513.00     21513
#> 377 2024-01-09 03:45:00  228600 21724.45 21517.85 21653.60 21544.85  21544.85
#> 378 2024-01-10 03:45:00  217000 21641.85 21448.65 21529.30 21618.70   21618.7
#> 379 2024-01-11 03:45:00  212500 21726.50 21593.75 21688.00 21647.20   21647.2
#> 380 2024-01-12 03:45:00  294700 21928.25 21715.15 21773.55 21894.55  21894.55
#> 381 2024-01-15 03:45:00  345500 22115.55 21963.55 22053.15 22097.45  22097.45
#> 382 2024-01-16 03:45:00  292400 22124.15 21969.80 22080.50 22032.30   22032.3
#> 383 2024-01-17 03:45:00  456000 21851.50 21550.45 21647.25 21571.95  21571.95
#> 384 2024-01-18 03:45:00  387300 21539.40 21285.55 21414.20 21462.25  21462.25
#> 385 2024-01-19 03:45:00  343100 21670.60 21575.00 21615.20 21622.40   21622.4
#> 386 2024-01-23 03:45:00  449700 21750.25 21192.60 21716.70 21238.80   21238.8
#> 387 2024-01-24 03:45:00  407500 21482.35 21137.20 21185.25 21453.95  21453.95
#> 388 2024-01-25 03:45:00  418100 21459.00 21247.05 21454.60 21352.60   21352.6
#> 389 2024-01-29 03:45:00  376700 21763.25 21429.60 21433.10 21737.60   21737.6
#> 390 2024-01-30 03:45:00  375100 21813.05 21501.80 21775.75 21522.10   21522.1
#> 391 2024-01-31 03:45:00  410600 21741.35 21448.85 21487.25 21725.70   21725.7
#> 392 2024-02-01 03:45:00  332500 21832.95 21658.75 21780.65 21697.45  21697.45
#> 393 2024-02-02 03:45:00  442800 22126.80 21805.55 21812.75 21853.80   21853.8
#> 394 2024-02-05 03:45:00  440800 21964.30 21726.95 21921.05 21771.70   21771.7
#> 395 2024-02-06 03:45:00  371000 21951.40 21737.55 21825.20 21929.40   21929.4
#> 396 2024-02-07 03:45:00  346300 22053.30 21860.15 22045.05 21930.50   21930.5
#> 397 2024-02-08 03:45:00  491100 22011.05 21665.30 22009.65 21717.95  21717.95
#> 398 2024-02-09 03:45:00  349200 21804.45 21629.90 21727.00 21782.50   21782.5
#> 399 2024-02-12 03:45:00  287400 21831.70 21574.75 21800.80 21616.05  21616.05
#> 400 2024-02-13 03:45:00  365800 21766.80 21543.35 21664.30 21743.25  21743.25
#> 401 2024-02-14 03:45:00  359100 21870.85 21530.20 21578.15 21840.05  21840.05
#> 402 2024-02-15 03:45:00  345400 21953.85 21794.80 21906.55 21910.75  21910.75
#> 403 2024-02-16 03:45:00  343900 22068.65 21968.95 22020.30 22040.70   22040.7
#> 404 2024-02-19 03:45:00       0 22186.65 22021.05 22103.45 22122.25  22122.25
#> 405 2024-02-20 03:45:00  295700 22215.60 22045.85 22099.20 22196.95  22196.95
#> 406 2024-02-21 03:45:00  364500 22249.40 21997.95 22248.85 22055.05  22055.05
#> 407 2024-02-22 03:45:00  343500 22252.50 21875.25 22081.55 22217.45  22217.45
#> 408 2024-02-23 03:45:00  226000 22297.50 22186.10 22290.00 22212.70   22212.7
#> 409 2024-02-26 03:45:00  207800 22202.15 22075.15 22169.20 22122.05  22122.05
#> 410 2024-02-27 03:45:00  252200 22218.25 22085.65 22090.20 22198.35  22198.35
#> 411 2024-02-28 03:45:00  203000 22229.15 21915.85 22214.10 21951.15  21951.15
#> 412 2024-02-29 03:45:00  360200 22060.55 21860.65 21935.20 21982.80   21982.8
#> 413 2024-03-01 03:45:00  351500 22353.30 22047.75 22048.30 22338.75  22338.75
#> 414 2024-03-04 03:45:00  298800 22440.90 22358.30 22403.50 22405.60   22405.6
#> 415 2024-03-05 03:45:00  296200 22416.90 22269.15 22371.25 22356.30   22356.3
#> 416 2024-03-06 03:45:00  312300 22497.20 22224.35 22327.50 22474.05  22474.05
#> 417 2024-03-07 03:45:00  379900 22525.65 22430.00 22505.30 22493.55  22493.55
#> 418 2024-03-11 03:45:00  277900 22526.60 22307.25 22517.50 22332.65  22332.65
#> 419 2024-03-12 03:45:00  299200 22452.55 22256.00 22334.45 22335.70   22335.7
#> 420 2024-03-13 03:45:00  493300 22446.75 21905.65 22432.20 21997.70   21997.7
#> 421 2024-03-14 03:45:00  426700 22204.60 21917.50 21982.55 22146.65  22146.65
#> 422 2024-03-15 03:45:00  661500 22120.90 21931.70 22064.85 22023.35  22023.35
#> 423 2024-03-18 03:45:00  356300 22123.70 21916.55 21990.10 22055.70   22055.7
#> 424 2024-03-19 03:45:00  344200 21978.30 21793.10 21946.45 21817.45  21817.45
#> 425 2024-03-20 03:45:00  312400 21930.90 21710.20 21843.90 21839.10   21839.1
#> 426 2024-03-21 03:45:00  353200 22080.95 21941.30 21989.90 22011.95  22011.95
#> 427 2024-03-22 03:45:00  388700 22180.70 21883.30 21932.20 22096.75  22096.75
#> 428 2024-03-26 03:45:00  328400 22073.20 21947.55 21947.90 22004.70   22004.7
#> 429 2024-03-27 03:45:00  409100 22193.60 22052.85 22053.95 22123.65  22123.65
#> 430 2024-03-28 03:45:00  410500 22516.00 22163.60 22163.60 22326.90   22326.9
#> 431 2024-04-01 03:45:00       0 22529.95 22427.75 22455.00 22462.00     22462
#> 432 2024-04-02 03:45:00  289500 22497.60 22388.15 22458.80 22453.30   22453.3
#> 433 2024-04-03 03:45:00  309600 22521.10 22346.50 22385.70 22434.65  22434.65
#> 434 2024-04-04 03:45:00  401000 22619.00 22303.80 22592.10 22514.65  22514.65
#> 435 2024-04-05 03:45:00  242200 22537.60 22427.60 22486.40 22513.70   22513.7
#> 436 2024-04-08 03:45:00  227600 22697.30 22550.35 22578.35 22666.30   22666.3
#> 437 2024-04-09 03:45:00  232400 22768.40 22612.25 22765.10 22642.75  22642.75
#> 438 2024-04-10 03:45:00  276800 22775.70 22673.70 22720.25 22753.80   22753.8
#> 439 2024-04-12 03:45:00  357200 22726.45 22503.75 22677.40 22519.40   22519.4
#> 440 2024-04-15 03:45:00  355500 22427.45 22259.55 22339.05 22272.50   22272.5
#> 441 2024-04-16 03:45:00  317300 22213.75 22079.45 22125.30 22147.90   22147.9
#> 442 2024-04-18 03:45:00  456900 22326.50 21961.70 22212.35 21995.85  21995.85
#> 443 2024-04-19 03:45:00  362500 22179.55 21777.65 21861.50 22147.00     22147
#> 444 2024-04-22 03:45:00  279300 22375.65 22198.15 22336.90 22336.40   22336.4
#> 445 2024-04-23 03:45:00  231500 22447.55 22349.45 22447.05 22368.00     22368
#> 446 2024-04-24 03:45:00  235900 22476.45 22384.00 22421.55 22402.40   22402.4
#> 447 2024-04-25 03:45:00  475000 22625.95 22305.25 22316.90 22570.35  22570.35
#> 448 2024-04-26 03:45:00  329900 22620.40 22385.55 22620.40 22419.95  22419.95
#> 449 2024-04-29 03:45:00  292000 22655.80 22441.90 22475.55 22643.40   22643.4
#> 450 2024-04-30 03:45:00  361900 22783.35 22568.40 22679.65 22604.85  22604.85
#> 451 2024-05-02 03:45:00  445900 22710.50 22567.85 22567.85 22648.20   22648.2
#> 452 2024-05-03 03:45:00  438100 22794.70 22348.05 22766.35 22475.85  22475.85
#> 453 2024-05-06 03:45:00  320300 22588.80 22409.45 22561.60 22442.70   22442.7
#> 454 2024-05-07 03:45:00  297800 22499.05 22232.05 22489.75 22302.50   22302.5
#> 455 2024-05-08 03:45:00  277400 22368.65 22185.20 22231.20 22302.50   22302.5
#> 456 2024-05-09 03:45:00  331300 22307.75 21932.40 22224.80 21957.50   21957.5
#> 457 2024-05-10 03:45:00  265800 22131.30 21950.30 21990.95 22055.20   22055.2
#> 458 2024-05-13 03:45:00  278200 22131.65 21821.05 22027.95 22104.05  22104.05
#> 459 2024-05-14 03:45:00  230200 22270.05 22081.25 22112.90 22217.85  22217.85
#> 460 2024-05-15 03:45:00  231900 22297.55 22151.75 22255.60 22200.55  22200.55
#> 461 2024-05-16 03:45:00  368900 22432.25 22054.55 22319.20 22403.85  22403.85
#> 462 2024-05-17 03:45:00  242700 22502.15 22345.65 22415.25 22466.10   22466.1
#> 463 2024-05-21 03:45:00  347600 22591.10 22404.55 22404.55 22529.05  22529.05
#> 464 2024-05-22 03:45:00  290300 22629.50 22483.15 22576.60 22597.80   22597.8
#> 465 2024-05-23 03:45:00  369800 22993.60 22577.45 22614.10 22967.65  22967.65
#> 466 2024-05-24 03:45:00  261900 23026.40 22908.00 22930.75 22957.10   22957.1
#> 467 2024-05-27 03:45:00  260000 23110.80 22871.20 23038.95 22932.45  22932.45
#> 468 2024-05-28 03:45:00  217900 22998.55 22858.50 22977.15 22888.15  22888.15
#> 469 2024-05-29 03:45:00  269900 22825.50 22685.45 22762.75 22704.70   22704.7
#> 470 2024-05-30 03:45:00  373400 22705.75 22417.00 22617.45 22488.65  22488.65
#> 471 2024-05-31 03:45:00  572100 22653.75 22465.10 22568.10 22530.70   22530.7
#> 472 2024-06-03 03:45:00  569400 23338.70 23062.30 23337.90 23263.90   23263.9
#> 473 2024-06-04 03:45:00 1006100 23179.50 21281.45 23179.50 21884.50   21884.5
#> 474 2024-06-05 03:45:00  638500 22670.40 21791.95 22128.35 22620.35  22620.35
#> 475 2024-06-06 03:45:00  480400 22910.15 22642.60 22798.60 22821.40   22821.4
#> 476 2024-06-07 03:45:00  473600 23320.20 22789.05 22821.85 23290.15  23290.15
#> 477 2024-06-10 03:45:00  304400 23411.90 23227.15 23319.15 23259.20   23259.2
#> 478 2024-06-11 03:45:00  305200 23389.45 23206.65 23283.75 23264.85  23264.85
#> 479 2024-06-12 03:45:00  295600 23441.95 23295.95 23344.45 23322.95  23322.95
#> 480 2024-06-13 03:45:00  268300 23481.05 23353.90 23480.95 23398.90   23398.9
#> 481 2024-06-14 03:45:00  223100 23490.40 23334.25 23464.95 23465.60   23465.6
#> 482 2024-06-18 03:45:00  272200 23579.05 23499.70 23570.80 23557.90   23557.9
#> 483 2024-06-19 03:45:00  328800 23664.00 23412.90 23629.85 23516.00     23516
#> 484 2024-06-20 03:45:00  280300 23624.00 23442.60 23586.15 23567.00     23567
#> 485 2024-06-21 03:45:00  609900 23667.10 23398.20 23661.15 23501.10   23501.1
#> 486 2024-06-24 03:45:00  239400 23558.10 23350.00 23382.30 23537.85  23537.85
#> 487 2024-06-25 03:45:00  298100 23754.15 23562.05 23577.10 23721.30   23721.3
#> 488 2024-06-26 03:45:00  287800 23889.90 23670.45 23723.10 23868.80   23868.8
#> 489 2024-06-27 03:45:00  515200 24087.45 23805.40 23881.55 24044.50   24044.5
#> 490 2024-06-28 03:45:00  354800 24174.00 23985.80 24085.90 24010.60   24010.6
#> 491 2024-07-01 03:45:00       0 24164.00 23992.70 23992.95 24141.95  24141.95
#> 492 2024-07-02 03:45:00       0 24236.35 24056.40 24228.75 24123.85  24123.85
#> 493 2024-07-03 03:45:00       0 24309.15 24207.10 24291.75 24286.50   24286.5
#> 494 2024-07-04 03:45:00  251200 24401.00 24281.00 24369.95 24302.15  24302.15
#> 495 2024-07-05 03:45:00  298400 24363.00 24168.85 24213.35 24323.85  24323.85
#> 496 2024-07-08 03:45:00  266300 24344.60 24240.55 24329.45 24320.55  24320.55
#> 497 2024-07-09 03:45:00  250500 24443.60 24331.90 24351.00 24433.20   24433.2
#> 498 2024-07-10 03:45:00  292300 24461.05 24141.80 24459.85 24324.45  24324.45
#> 499 2024-07-11 03:45:00  306400 24402.65 24193.75 24396.55 24315.95  24315.95
#> 500 2024-07-12 03:45:00  325800 24592.20 24331.15 24387.95 24502.15  24502.15
#> 501 2024-07-15 03:45:00  305400 24635.05 24522.75 24587.60 24586.70   24586.7
#> 502 2024-07-16 03:45:00  283200 24661.25 24587.65 24615.90 24613.00     24613
#> 503 2024-07-18 03:45:00  350900 24837.75 24504.45 24543.80 24800.85  24800.85
#> 504 2024-07-19 03:45:00  343800 24854.80 24508.15 24853.80 24530.90   24530.9
#> 505 2024-07-22 03:45:00  324200 24595.20 24362.30 24445.75 24509.25  24509.25
#> 506 2024-07-23 03:45:00  436400 24582.55 24074.20 24568.90 24479.05  24479.05
#> 507 2024-07-24 03:45:00  366600 24504.25 24307.25 24444.95 24413.50   24413.5
#> 508 2024-07-25 03:45:00  391800 24426.15 24210.80 24230.95 24406.10   24406.1
#> 509 2024-07-26 03:45:00  383800 24861.15 24410.90 24423.35 24834.85  24834.85
#> 510 2024-07-29 03:45:00  355000 24999.75 24774.60 24943.30 24836.10   24836.1
#> 511 2024-07-30 03:45:00  385000 24971.75 24798.65 24839.40 24857.30   24857.3
#> 512 2024-07-31 03:45:00  333600 24984.60 24856.50 24886.70 24951.15  24951.15
#> 513 2024-08-01 03:45:00  431300 25078.30 24956.40 25030.95 25010.90   25010.9
#> 514 2024-08-02 03:45:00  345000 24851.90 24686.85 24789.00 24717.70   24717.7
#> 515 2024-08-05 03:45:00  487000 24350.05 23893.70 24302.85 24055.60   24055.6
#> 516 2024-08-06 03:45:00  312300 24382.60 23960.40 24189.85 23992.55  23992.55
#> 517 2024-08-07 03:45:00  317600 24337.70 24184.90 24289.40 24297.50   24297.5
#> 518 2024-08-08 03:45:00  311900 24340.50 24079.70 24248.55 24117.00     24117
#> 519 2024-08-09 03:45:00  215100 24419.75 24311.20 24386.85 24367.50   24367.5
#> 520 2024-08-12 03:45:00  279900 24472.80 24212.10 24320.05 24347.00     24347
#> 521 2024-08-13 03:45:00  239700 24359.95 24116.50 24342.35 24139.00     24139
#> 522 2024-08-14 03:45:00  303300 24196.50 24099.70 24184.40 24143.75  24143.75
#> 523 2024-08-16 03:45:00  271600 24563.90 24204.50 24334.85 24541.15  24541.15
#> 524 2024-08-19 03:45:00  243600 24638.80 24522.95 24636.35 24572.65  24572.65
#> 525 2024-08-20 03:45:00  238300 24734.30 24607.20 24648.90 24698.85  24698.85
#> 526 2024-08-21 03:45:00  257100 24787.95 24654.50 24680.55 24770.20   24770.2
#> 527 2024-08-22 03:45:00  220300 24867.35 24784.45 24863.40 24811.50   24811.5
#> 528 2024-08-23 03:45:00  206800 24858.40 24771.65 24845.40 24823.15  24823.15
#> 529 2024-08-26 03:45:00  210300 25043.80 24874.70 24906.10 25010.60   25010.6
#> 530 2024-08-27 03:45:00  223300 25073.10 24973.65 25024.80 25017.75  25017.75
#> 531 2024-08-28 03:45:00  220400 25129.60 24964.65 25030.80 25052.35  25052.35
#> 532 2024-08-29 03:45:00  354000 25192.90 24998.50 25035.30 25151.95  25151.95
#> 533 2024-08-30 03:45:00  638200 25268.35 25199.40 25249.70 25235.90   25235.9
#> 534 2024-09-02 03:45:00  222800 25333.65 25235.50 25333.60 25278.70   25278.7
#> 535 2024-09-03 03:45:00  212100 25321.70 25235.80 25313.40 25279.85  25279.85
#> 536 2024-09-04 03:45:00  253800 25216.00 25083.80 25089.95 25198.70   25198.7
#> 537 2024-09-05 03:45:00  222200 25275.45 25127.75 25250.50 25145.10   25145.1
#> 538 2024-09-06 03:45:00  311700 25168.75 24801.30 25093.70 24852.15  24852.15
#> 539 2024-09-09 03:45:00  254400 24957.50 24753.15 24823.40 24936.40   24936.4
#> 540 2024-09-10 03:45:00  251300 25130.50 24896.80 24999.40 25041.10   25041.1
#> 541 2024-09-11 03:45:00  279200 25113.70 24885.15 25034.00 24918.45  24918.45
#> 542 2024-09-12 03:45:00  380100 25433.35 24941.45 25059.65 25388.90   25388.9
#> 543 2024-09-13 03:45:00  250800 25430.50 25292.45 25430.45 25356.50   25356.5
#> 544 2024-09-16 03:45:00  168700 25445.70 25336.20 25406.65 25383.75  25383.75
#> 545 2024-09-17 03:45:00  216000 25441.65 25352.25 25416.90 25418.55  25418.55
#> 546 2024-09-18 03:45:00  215700 25482.20 25285.55 25402.40 25377.55  25377.55
#> 547 2024-09-19 03:45:00  314500 25611.95 25376.05 25487.05 25415.80   25415.8
#> 548 2024-09-20 03:45:00  533100 25849.25 25426.60 25525.95 25790.95  25790.95
#> 549 2024-09-23 03:45:00  209200 25956.00 25847.35 25872.55 25939.05  25939.05
#> 550 2024-09-24 03:45:00  384100 26011.55 25886.85 25921.45 25940.40   25940.4
#> 551 2024-09-25 03:45:00  278500 26032.80 25871.35 25899.45 26004.15  26004.15
#> 552 2024-09-26 03:45:00  370900 26250.90 25998.40 26005.40 26216.05  26216.05
#> 553 2024-09-27 03:45:00  490300 26277.35 26151.40 26248.25 26178.95  26178.95
#> 554 2024-09-30 03:45:00  343100 26134.70 25794.10 26061.30 25810.85  25810.85
#> 555 2024-10-01 03:45:00  247400 25907.60 25739.20 25788.45 25796.90   25796.9
#> 556 2024-10-03 03:45:00  423400 25639.45 25230.30 25452.85 25250.10   25250.1
#> 557 2024-10-04 03:45:00  374900 25485.05 24966.80 25181.90 25014.60   25014.6
#> 558 2024-10-07 03:45:00  374300 25143.00 24694.35 25084.10 24795.75  24795.75
#> 559 2024-10-08 03:45:00  333200 25044.00 24756.80 24832.20 25013.15  25013.15
#> 560 2024-10-09 03:45:00  290600 25234.05 24947.70 25065.80 24981.95  24981.95
#> 561 2024-10-10 03:45:00  261400 25134.05 24979.40 25067.05 24998.45  24998.45
#> 562 2024-10-11 03:45:00  210500 25028.65 24920.05 24985.30 24964.25  24964.25
#> 563 2024-10-14 03:45:00  206400 25159.75 25017.50 25023.45 25127.95  25127.95
#> 564 2024-10-15 03:45:00  257200 25212.05 25008.15 25186.30 25057.35  25057.35
#> 565 2024-10-16 03:45:00  226800 25093.40 24908.45 25008.55 24971.30   24971.3
#> 566 2024-10-17 03:45:00  252800 25029.50 24728.90 25027.40 24749.85  24749.85
#> 567 2024-10-18 03:45:00  290900 24886.20 24567.65 24664.95 24854.05  24854.05
#> 568 2024-10-21 03:45:00  242000 24978.30 24679.60 24956.15 24781.10   24781.1
#> 569 2024-10-22 03:45:00  279500 24882.00 24445.80 24798.65 24472.10   24472.1
#> 570 2024-10-23 03:45:00  284600 24604.25 24378.10 24378.15 24435.50   24435.5
#> 571 2024-10-24 03:45:00  216700 24480.65 24341.20 24412.70 24399.40   24399.4
#> 572 2024-10-25 03:45:00  372600 24440.25 24073.90 24418.05 24180.80   24180.8
#> 573 2024-10-28 03:45:00  278600 24492.60 24134.90 24251.10 24339.15  24339.15
#> 574 2024-10-29 03:45:00  332000 24484.50 24140.85 24328.85 24466.85  24466.85
#> 575 2024-10-30 03:45:00  285200 24498.20 24307.30 24371.45 24340.85  24340.85
#> 576 2024-10-31 03:45:00  287000 24372.45 24172.60 24349.85 24205.35  24205.35
#> 577 2024-11-01 03:45:00   38800 24368.25 24280.20 24302.75 24304.35  24304.35
#> 578 2024-11-04 03:45:00  285500 24316.75 23816.15 24315.75 23995.35  23995.35
#> 579 2024-11-05 03:45:00  289500 24229.05 23842.75 23916.50 24213.30   24213.3
#> 580 2024-11-06 03:45:00  351100 24537.60 24204.05 24308.75 24484.05  24484.05
#> 581 2024-11-07 03:45:00  322200 24503.35 24179.05 24489.60 24199.35  24199.35
#> 582 2024-11-08 03:45:00  298700 24276.15 24066.65 24207.70 24148.20   24148.2
#> 583 2024-11-11 03:45:00  273400 24336.80 24004.60 24087.25 24141.30   24141.3
#> 584 2024-11-12 03:45:00  255800 24242.00 23839.15 24225.80 23883.45  23883.45
#> 585 2024-11-13 03:45:00  304600 23873.60 23509.60 23822.45 23559.05  23559.05
#> 586 2024-11-14 10:01:07       0 23675.90 23484.15 23542.15 23532.70   23532.7
nifty$get_history(start = '2022-07-01', end = '2022-07-14', interval = '1d')
#>                  date volume     high      low     open    close adj_close
#> 1 2022-07-01 03:45:00 364100 15793.95 15511.05 15703.70 15752.05  15752.05
#> 2 2022-07-04 03:45:00 304300 15852.35 15661.80 15710.50 15835.35  15835.35
#> 3 2022-07-05 03:45:00 254200 16025.75 15785.45 15909.15 15810.85  15810.85
#> 4 2022-07-06 03:45:00 288400 16011.35 15800.90 15818.20 15989.80   15989.8
#> 5 2022-07-07 03:45:00 264600 16150.50 16045.95 16113.75 16132.90   16132.9
#> 6 2022-07-08 03:45:00 281100 16275.50 16157.90 16273.65 16220.60   16220.6
#> 7 2022-07-11 03:45:00 255900 16248.55 16115.50 16136.15 16216.00     16216
#> 8 2022-07-12 03:45:00 208600 16158.75 16031.15 16126.20 16058.30   16058.3
#> 9 2022-07-13 03:45:00 233300 16140.00 15950.15 16128.20 15966.65  15966.65
nifty$get_history(period = '1mo', interval = '1d')
#>                   date volume     high      low     open    close adj_close
#> 1  2024-10-14 03:45:00 206400 25159.75 25017.50 25023.45 25127.95  25127.95
#> 2  2024-10-15 03:45:00 257200 25212.05 25008.15 25186.30 25057.35  25057.35
#> 3  2024-10-16 03:45:00 226800 25093.40 24908.45 25008.55 24971.30   24971.3
#> 4  2024-10-17 03:45:00 252800 25029.50 24728.90 25027.40 24749.85  24749.85
#> 5  2024-10-18 03:45:00 290900 24886.20 24567.65 24664.95 24854.05  24854.05
#> 6  2024-10-21 03:45:00 242000 24978.30 24679.60 24956.15 24781.10   24781.1
#> 7  2024-10-22 03:45:00 279500 24882.00 24445.80 24798.65 24472.10   24472.1
#> 8  2024-10-23 03:45:00 284600 24604.25 24378.10 24378.15 24435.50   24435.5
#> 9  2024-10-24 03:45:00 216700 24480.65 24341.20 24412.70 24399.40   24399.4
#> 10 2024-10-25 03:45:00 372600 24440.25 24073.90 24418.05 24180.80   24180.8
#> 11 2024-10-28 03:45:00 278600 24492.60 24134.90 24251.10 24339.15  24339.15
#> 12 2024-10-29 03:45:00 332000 24484.50 24140.85 24328.85 24466.85  24466.85
#> 13 2024-10-30 03:45:00 285200 24498.20 24307.30 24371.45 24340.85  24340.85
#> 14 2024-10-31 03:45:00 287000 24372.45 24172.60 24349.85 24205.35  24205.35
#> 15 2024-11-01 03:45:00  38800 24368.25 24280.20 24302.75 24304.35  24304.35
#> 16 2024-11-04 03:45:00 285500 24316.75 23816.15 24315.75 23995.35  23995.35
#> 17 2024-11-05 03:45:00 289500 24229.05 23842.75 23916.50 24213.30   24213.3
#> 18 2024-11-06 03:45:00 351100 24537.60 24204.05 24308.75 24484.05  24484.05
#> 19 2024-11-07 03:45:00 322200 24503.35 24179.05 24489.60 24199.35  24199.35
#> 20 2024-11-08 03:45:00 298700 24276.15 24066.65 24207.70 24148.20   24148.2
#> 21 2024-11-11 03:45:00 273400 24336.80 24004.60 24087.25 24141.30   24141.3
#> 22 2024-11-12 03:45:00 255800 24242.00 23839.15 24225.80 23883.45  23883.45
#> 23 2024-11-13 03:45:00 304600 23873.60 23509.60 23822.45 23559.05  23559.05
#> 24 2024-11-14 10:01:07      0 23675.90 23484.15 23542.15 23532.70   23532.7
# }