R6 Class Representing a Ticker

Base class for getting all data related to ticker from Yahoo Finance API.

Format

An R6 class object

Public fields

symbol

Symbol for which data is retrieved.

Active bindings

valuation_measures

Retrieves valuation measures for most recent four quarters

option_chain

Option chain data for all expiration dates for a given symbol

option_expiration_dates

Option expiration dates

option_strikes

Option strikes

quote

Get real-time quote information for given symbol

recommendations

Recommended symbols

technical_insights

Technical indicators for given symbol

Methods

Public methods

Ticker$new()

Ticker$set_symbol()

Ticker$get_history()

Ticker$clone()

Method new()

Create a new Ticker object.

Usage

Ticker$new(symbol = NA)

Arguments

symbol

Symbol.

Returns

A new `Ticker` object

Examples

aapl <- Ticker$new('aapl')

Method set_symbol()

Set a new symbol.

Usage

Ticker$set_symbol(symbol)

Arguments

symbol

New symbol

Examples

aapl <- Ticker$new('aapl')
aapl$set_symbol('msft')

Method get_history()

Retrieves historical pricing data.

Usage

Ticker$get_history(period = "ytd", interval = "1d", start = NULL, end = NULL)

Arguments

period

Length of time. Defaults to 'ytd'. Valid values are:

'1d'

'5d'

'1mo'

'3mo'

'6mo'

'1y'

'2y'

'5y'

'10y'

'ytd'

'max'

interval

Time between data points. Defaults to '1d'. Valid values are:

'1m'

'2m'

'5m'

'15m'

'30m'

'60m'

'90m'

'1h'

'1d'

'5d'

'1wk'

'1mo'

'3mo'

start

Specific starting date. String or date object in yyyy-mm-dd format.

end

Specific ending date. String or date object in yyyy-mm-dd format.

Returns

A data.frame.

Examples

\donttest{
aapl <- Ticker$new('aapl')
aapl$get_history(start = '2022-07-01', interval = '1d')
aapl$get_history(start = '2022-07-01', end = '2022-07-14', interval = '1d')
aapl$get_history(period = '1mo', interval = '1d')
}

Method clone()

The objects of this class are cloneable with this method.

Usage

Ticker$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `Ticker$new`
## ------------------------------------------------

aapl <- Ticker$new('aapl')

## ------------------------------------------------
## Method `Ticker$set_symbol`
## ------------------------------------------------

aapl <- Ticker$new('aapl')
aapl$set_symbol('msft')

## ------------------------------------------------
## Method `Ticker$get_history`
## ------------------------------------------------

# \donttest{
aapl <- Ticker$new('aapl')
aapl$get_history(start = '2022-07-01', interval = '1d')
#>                    date    volume   high    low   open   close adj_close
#> 1   2022-07-01 13:30:00  71051600 139.04 135.66 136.04 138.930  138.1053
#> 2   2022-07-05 13:30:00  73353800 141.61 136.93 137.77 141.560  140.7197
#> 3   2022-07-06 13:30:00  74064300 144.12 141.08 141.35 142.920  142.0716
#> 4   2022-07-07 13:30:00  66253700 146.55 143.28 143.29 146.350  145.4813
#> 5   2022-07-08 13:30:00  64547800 147.55 145.00 145.26 147.040  146.1672
#> 6   2022-07-11 13:30:00  63141600 146.64 143.78 145.67 144.870  144.0101
#> 7   2022-07-12 13:30:00  77588800 148.45 145.05 145.76 145.860  144.9942
#> 8   2022-07-13 13:30:00  71185600 146.45 142.12 142.99 145.490  144.6264
#> 9   2022-07-14 13:30:00  78140700 148.95 143.25 144.08 148.470  147.5887
#> 10  2022-07-15 13:30:00  76259900 150.86 148.20 149.78 150.170  149.2786
#> 11  2022-07-18 13:30:00  81420900 151.57 146.70 150.74 147.070  146.1970
#> 12  2022-07-19 13:30:00  82982400 151.23 146.91 147.92 151.000  150.1037
#> 13  2022-07-20 13:30:00  64823400 153.72 150.37 151.12 153.040  152.1316
#> 14  2022-07-21 13:30:00  65086600 155.57 151.94 154.50 155.350  154.4279
#> 15  2022-07-22 13:30:00  66675400 156.28 153.41 155.39 154.090  153.1754
#> 16  2022-07-25 13:30:00  53623900 155.04 152.28 154.01 152.950  152.0421
#> 17  2022-07-26 13:30:00  55138700 153.09 150.80 152.26 151.600  150.7001
#> 18  2022-07-27 13:30:00  78620700 157.33 152.16 152.58 156.790  155.8593
#> 19  2022-07-28 13:30:00  81378700 157.64 154.41 156.98 157.350  156.4160
#> 20  2022-07-29 13:30:00 101786900 163.63 159.50 161.24 162.510  161.5454
#> 21  2022-08-01 13:30:00  67829400 163.59 160.89 161.01 161.510  160.5513
#> 22  2022-08-02 13:30:00  59907000 162.41 159.63 160.10 160.010  159.0602
#> 23  2022-08-03 13:30:00  82507500 166.59 160.75 160.84 166.130  165.1439
#> 24  2022-08-04 13:30:00  55474100 167.19 164.43 166.01 165.810  164.8258
#> 25  2022-08-05 13:30:00  56697000 165.85 163.00 163.21 165.350  164.5968
#> 26  2022-08-08 13:30:00  60276900 167.81 164.20 166.37 164.870  164.1190
#> 27  2022-08-09 13:30:00  63135500 165.82 163.25 164.02 164.920  164.1688
#> 28  2022-08-10 13:30:00  70170500 169.34 166.90 167.68 169.240  168.4691
#> 29  2022-08-11 13:30:00  57149200 170.99 168.19 170.06 168.490  167.7225
#> 30  2022-08-12 13:30:00  68039400 172.17 169.40 169.82 172.100  171.3161
#> 31  2022-08-15 13:30:00  54091700 173.39 171.35 171.52 173.190  172.4011
#> 32  2022-08-16 13:30:00  56377100 173.71 171.66 172.78 173.030  172.2419
#> 33  2022-08-17 13:30:00  79542000 176.15 172.57 172.77 174.550  173.7549
#> 34  2022-08-18 13:30:00  62290100 174.90 173.12 173.75 174.150  173.3567
#> 35  2022-08-19 13:30:00  70346300 173.74 171.31 173.03 171.520  170.7387
#> 36  2022-08-22 13:30:00  69026800 169.86 167.14 169.69 167.570  166.8067
#> 37  2022-08-23 13:30:00  54147100 168.71 166.65 167.08 167.230  166.4683
#> 38  2022-08-24 13:30:00  53841500 168.11 166.25 167.32 167.530  166.7669
#> 39  2022-08-25 13:30:00  51218200 170.14 168.35 168.78 170.030  169.2555
#> 40  2022-08-26 13:30:00  78961000 171.05 163.56 170.57 163.620  162.8747
#> 41  2022-08-29 13:30:00  73314000 162.90 159.82 161.15 161.380  160.6449
#> 42  2022-08-30 13:30:00  77906200 162.56 157.72 162.13 158.910  158.1862
#> 43  2022-08-31 13:30:00  87991100 160.58 157.14 160.31 157.220  156.5039
#> 44  2022-09-01 13:30:00  74229900 158.42 154.67 156.64 157.960  157.2405
#> 45  2022-09-02 13:30:00  76957800 160.36 154.97 159.75 155.810  155.1003
#> 46  2022-09-06 13:30:00  73714800 157.09 153.69 156.47 154.530  153.8261
#> 47  2022-09-07 13:30:00  87449600 156.67 153.61 154.82 155.960  155.2496
#> 48  2022-09-08 13:30:00  84923800 156.36 152.68 154.64 154.460  153.7564
#> 49  2022-09-09 13:30:00  68028800 157.82 154.75 155.47 157.370  156.6532
#> 50  2022-09-12 13:30:00 104956000 164.26 159.30 159.59 163.430  162.6856
#> 51  2022-09-13 13:30:00 122656600 160.54 153.37 159.90 153.840  153.1393
#> 52  2022-09-14 13:30:00  87965400 157.10 153.61 154.79 155.310  154.6026
#> 53  2022-09-15 13:30:00  90481100 155.24 151.38 154.65 152.370  151.6759
#> 54  2022-09-16 13:30:00 162278800 151.35 148.37 151.21 150.700  150.0135
#> 55  2022-09-19 13:30:00  81474200 154.56 149.10 149.31 154.480  153.7763
#> 56  2022-09-20 13:30:00 107689800 158.08 153.08 153.40 156.900  156.1853
#> 57  2022-09-21 13:30:00 101696800 158.74 153.60 157.34 153.720  153.0198
#> 58  2022-09-22 13:30:00  86652500 154.47 150.91 152.38 152.740  152.0443
#> 59  2022-09-23 13:30:00  96029900 151.47 148.56 151.19 150.430  149.7448
#> 60  2022-09-26 13:30:00  93339400 153.77 149.64 149.66 150.770  150.0832
#> 61  2022-09-27 13:30:00  84442700 154.72 149.95 152.74 151.760  151.0687
#> 62  2022-09-28 13:30:00 146691400 150.64 144.84 147.64 149.840  149.1575
#> 63  2022-09-29 13:30:00 128138200 146.72 140.68 146.10 142.480  141.8310
#> 64  2022-09-30 13:30:00 124925300 143.10 138.00 141.28 138.200  137.5705
#> 65  2022-10-03 13:30:00 114311700 143.07 137.69 138.21 142.450  141.8011
#> 66  2022-10-04 13:30:00  87830100 146.22 144.26 145.03 146.100  145.4345
#> 67  2022-10-05 13:30:00  79471000 147.38 143.01 144.07 146.400  145.7331
#> 68  2022-10-06 13:30:00  68402200 147.54 145.22 145.81 145.430  144.7676
#> 69  2022-10-07 13:30:00  85925600 143.10 139.45 142.54 140.090  139.4519
#> 70  2022-10-10 13:30:00  74899000 141.89 138.57 140.42 140.420  139.7804
#> 71  2022-10-11 13:30:00  77033700 141.35 138.22 139.90 138.980  138.3469
#> 72  2022-10-12 13:30:00  70433700 140.36 138.16 139.13 138.340  137.7099
#> 73  2022-10-13 13:30:00 113224000 143.59 134.37 134.99 142.990  142.3387
#> 74  2022-10-14 13:30:00  88598000 144.52 138.19 144.31 138.380  137.7497
#> 75  2022-10-17 13:30:00  85250900 142.90 140.27 141.07 142.410  141.7613
#> 76  2022-10-18 13:30:00  99136600 146.70 140.61 145.49 143.750  143.0952
#> 77  2022-10-19 13:30:00  61758300 144.95 141.50 141.69 143.860  143.2047
#> 78  2022-10-20 13:30:00  64522000 145.89 142.65 143.02 143.390  142.7369
#> 79  2022-10-21 13:30:00  86548600 147.85 142.65 142.87 147.270  146.5992
#> 80  2022-10-24 13:30:00  75981900 150.23 146.00 147.19 149.450  148.7692
#> 81  2022-10-25 13:30:00  74732300 152.49 149.36 150.09 152.340  151.6461
#> 82  2022-10-26 13:30:00  88194300 151.99 148.04 150.96 149.350  148.6697
#> 83  2022-10-27 13:30:00 109180200 149.05 144.13 148.07 144.800  144.1404
#> 84  2022-10-28 13:30:00 164762400 157.50 147.82 148.20 155.740  155.0306
#> 85  2022-10-31 13:30:00  97943200 154.24 151.92 153.16 153.340  152.6415
#> 86  2022-11-01 13:30:00  80379300 155.45 149.13 155.08 150.650  149.9638
#> 87  2022-11-02 13:30:00  93604600 152.17 145.00 148.95 145.030  144.3694
#> 88  2022-11-03 13:30:00  97918500 142.80 138.75 142.06 138.880  138.2474
#> 89  2022-11-04 13:30:00 140814800 142.67 134.38 142.09 138.380  137.9782
#> 90  2022-11-07 14:30:00  83374600 139.15 135.67 137.11 138.920  138.5166
#> 91  2022-11-08 14:30:00  89908500 141.43 137.49 140.41 139.500  139.0949
#> 92  2022-11-09 14:30:00  74917800 138.55 134.59 138.50 134.870  134.4784
#> 93  2022-11-10 14:30:00 118854000 146.87 139.50 141.24 146.870  146.4435
#> 94  2022-11-11 14:30:00  93979700 150.01 144.37 145.82 149.700  149.2653
#> 95  2022-11-14 14:30:00  73374100 150.28 147.43 148.97 148.280  147.8494
#> 96  2022-11-15 14:30:00  89868300 153.59 148.56 152.22 150.040  149.6043
#> 97  2022-11-16 14:30:00  64218300 149.87 147.29 149.13 148.790  148.3580
#> 98  2022-11-17 14:30:00  80389400 151.48 146.15 146.43 150.720  150.2824
#> 99  2022-11-18 14:30:00  74829600 152.70 149.97 152.31 151.290  150.8507
#> 100 2022-11-21 14:30:00  58724100 150.37 147.72 150.16 148.010  147.5802
#> 101 2022-11-22 14:30:00  51804100 150.42 146.93 148.13 150.180  149.7439
#> 102 2022-11-23 14:30:00  58301400 151.83 149.34 149.45 151.070  150.6313
#> 103 2022-11-25 14:30:00  35195900 148.88 147.12 148.31 148.110  147.6799
#> 104 2022-11-28 14:30:00  69246000 146.64 143.38 145.14 144.220  143.8012
#> 105 2022-11-29 14:30:00  83763800 144.81 140.35 144.29 141.170  140.7601
#> 106 2022-11-30 14:30:00 111380900 148.72 140.55 141.40 148.030  147.6002
#> 107 2022-12-01 14:30:00  71250400 149.13 146.61 148.21 148.310  147.8793
#> 108 2022-12-02 14:30:00  65447400 148.00 145.65 145.96 147.810  147.3808
#> 109 2022-12-05 14:30:00  68826400 150.92 145.77 147.77 146.630  146.2042
#> 110 2022-12-06 14:30:00  64727200 147.30 141.92 147.07 142.910  142.4950
#> 111 2022-12-07 14:30:00  69721100 143.37 140.00 142.19 140.940  140.5308
#> 112 2022-12-08 14:30:00  62128300 143.52 141.10 142.36 142.650  142.2358
#> 113 2022-12-09 14:30:00  76097000 145.57 140.90 142.34 142.160  141.7472
#> 114 2022-12-12 14:30:00  70462700 144.50 141.06 142.70 144.490  144.0704
#> 115 2022-12-13 14:30:00  93886200 149.97 144.24 149.50 145.470  145.0476
#> 116 2022-12-14 14:30:00  82291200 146.66 141.16 145.35 143.210  142.7942
#> 117 2022-12-15 14:30:00  98931900 141.80 136.03 141.11 136.500  136.1036
#> 118 2022-12-16 14:30:00 160156900 137.65 133.73 136.69 134.510  134.1194
#> 119 2022-12-19 14:30:00  79592600 135.20 131.32 135.11 132.370  131.9856
#> 120 2022-12-20 14:30:00  77432800 133.25 129.89 131.39 132.300  131.9158
#> 121 2022-12-21 14:30:00  85928000 136.81 132.75 132.98 135.450  135.0567
#> 122 2022-12-22 14:30:00  77852100 134.56 130.30 134.35 132.230  131.8460
#> 123 2022-12-23 14:30:00  63814900 132.42 129.64 130.92 131.860  131.4771
#> 124 2022-12-27 14:30:00  69007800 131.41 128.72 131.38 130.030  129.6524
#> 125 2022-12-28 14:30:00  85438400 131.03 125.87 129.67 126.040  125.6740
#> 126 2022-12-29 14:30:00  75703700 130.48 127.73 127.99 129.610  129.2337
#> 127 2022-12-30 14:30:00  77034200 129.95 127.43 128.41 129.930  129.5527
#> 128 2023-01-03 14:30:00 112117500 130.90 124.17 130.28 125.070  124.7068
#> 129 2023-01-04 14:30:00  89113600 128.66 125.08 126.89 126.360  125.9931
#> 130 2023-01-05 14:30:00  80962700 127.77 124.76 127.13 125.020  124.6570
#> 131 2023-01-06 14:30:00  87754700 130.29 124.89 126.01 129.620  129.2436
#> 132 2023-01-09 14:30:00  70790800 133.41 129.89 130.47 130.150  129.7721
#> 133 2023-01-10 14:30:00  63896200 131.26 128.12 130.26 130.730  130.3504
#> 134 2023-01-11 14:30:00  69458900 133.51 130.46 131.25 133.490  133.1024
#> 135 2023-01-12 14:30:00  71379600 134.26 131.44 133.88 133.410  133.0226
#> 136 2023-01-13 14:30:00  57809700 134.92 131.66 132.03 134.760  134.3687
#> 137 2023-01-17 14:30:00  63646600 137.29 134.13 134.83 135.940  135.5453
#> 138 2023-01-18 14:30:00  69672800 138.61 135.03 136.82 135.210  134.8174
#> 139 2023-01-19 14:30:00  58280400 136.25 133.77 134.08 135.270  134.8772
#> 140 2023-01-20 14:30:00  80223600 138.02 134.22 135.28 137.870  137.4697
#> 141 2023-01-23 14:30:00  81760300 143.32 137.90 138.12 141.110  140.7003
#> 142 2023-01-24 14:30:00  66435100 143.16 140.30 140.31 142.530  142.1161
#> 143 2023-01-25 14:30:00  65799300 142.43 138.81 140.89 141.860  141.4481
#> 144 2023-01-26 14:30:00  54105100 144.25 141.90 143.17 143.960  143.5420
#> 145 2023-01-27 14:30:00  70555800 147.23 143.08 143.16 145.930  145.5063
#> 146 2023-01-30 14:30:00  64015300 145.55 142.85 144.96 143.000  142.5848
#> 147 2023-01-31 14:30:00  65874500 144.34 142.28 142.70 144.290  143.8710
#> 148 2023-02-01 14:30:00  77663600 146.61 141.32 143.97 145.430  145.0077
#> 149 2023-02-02 14:30:00 118339000 151.18 148.17 148.90 150.820  150.3821
#> 150 2023-02-03 14:30:00 154357300 157.38 147.83 148.03 154.500  154.0514
#> 151 2023-02-06 14:30:00  69858300 153.10 150.78 152.57 151.730  151.2894
#> 152 2023-02-07 14:30:00  83322600 155.23 150.64 150.64 154.650  154.2009
#> 153 2023-02-08 14:30:00  64120100 154.58 151.17 153.88 151.920  151.4789
#> 154 2023-02-09 14:30:00  56007100 154.33 150.42 153.78 150.870  150.4319
#> 155 2023-02-10 14:30:00  57450700 151.34 149.22 149.46 151.010  150.8014
#> 156 2023-02-13 14:30:00  62199000 154.26 150.92 150.95 153.850  153.6375
#> 157 2023-02-14 14:30:00  61707600 153.77 150.86 152.12 153.200  152.9884
#> 158 2023-02-15 14:30:00  65573800 155.50 152.88 153.11 155.330  155.1154
#> 159 2023-02-16 14:30:00  68167900 156.33 153.35 153.51 153.710  153.4977
#> 160 2023-02-17 14:30:00  59144100 153.00 150.85 152.35 152.550  152.3393
#> 161 2023-02-21 14:30:00  58867200 151.30 148.41 150.20 148.480  148.2749
#> 162 2023-02-22 14:30:00  51011300 149.95 147.16 148.87 148.910  148.7043
#> 163 2023-02-23 14:30:00  48394200 150.34 147.24 150.09 149.400  149.1936
#> 164 2023-02-24 14:30:00  55469600 147.19 145.72 147.11 146.710  146.5074
#> 165 2023-02-27 14:30:00  44998500 149.17 147.45 147.71 147.920  147.7157
#> 166 2023-02-28 14:30:00  50547000 149.08 146.83 147.05 147.410  147.2064
#> 167 2023-03-01 14:30:00  55479000 147.23 145.01 146.83 145.310  145.1093
#> 168 2023-03-02 14:30:00  52238100 146.71 143.90 144.38 145.910  145.7085
#> 169 2023-03-03 14:30:00  70732300 151.11 147.33 148.04 151.030  150.8214
#> 170 2023-03-06 14:30:00  87558000 156.30 153.46 153.79 153.830  153.6175
#> 171 2023-03-07 14:30:00  56182000 154.03 151.13 153.70 151.600  151.3906
#> 172 2023-03-08 14:30:00  47204800 153.47 151.83 152.81 152.870  152.6588
#> 173 2023-03-09 14:30:00  53833600 154.54 150.23 153.56 150.590  150.3820
#> 174 2023-03-10 14:30:00  68572400 150.94 147.61 150.21 148.500  148.2949
#> 175 2023-03-13 13:30:00  84457100 153.14 147.70 147.81 150.470  150.2622
#> 176 2023-03-14 13:30:00  73695900 153.40 150.10 151.28 152.590  152.3792
#> 177 2023-03-15 13:30:00  77167900 153.25 149.92 151.19 152.990  152.7787
#> 178 2023-03-16 13:30:00  76161100 156.46 151.64 152.16 155.850  155.6347
#> 179 2023-03-17 13:30:00  98944600 156.74 154.28 156.08 155.000  154.7859
#> 180 2023-03-20 13:30:00  73641400 157.82 154.15 155.07 157.400  157.1826
#> 181 2023-03-21 13:30:00  73938300 159.40 156.54 157.32 159.280  159.0600
#> 182 2023-03-22 13:30:00  75701800 162.14 157.81 159.30 157.830  157.6120
#> 183 2023-03-23 13:30:00  67622100 161.55 157.68 158.83 158.930  158.7105
#> 184 2023-03-24 13:30:00  59196500 160.34 157.85 158.86 160.250  160.0287
#> 185 2023-03-27 13:30:00  52390300 160.77 157.87 159.94 158.280  158.0614
#> 186 2023-03-28 13:30:00  45992200 158.49 155.98 157.97 157.650  157.4322
#> 187 2023-03-29 13:30:00  51305700 161.05 159.35 159.37 160.770  160.5479
#> 188 2023-03-30 13:30:00  49501700 162.47 161.27 161.53 162.360  162.1357
#> 189 2023-03-31 13:30:00  68749800 165.00 161.91 162.44 164.900  164.6722
#> 190 2023-04-03 13:30:00  56976200 166.29 164.22 164.27 166.170  165.9405
#> 191 2023-04-04 13:30:00  46278300 166.84 165.11 166.60 165.630  165.4012
#> 192 2023-04-05 13:30:00  51511700 165.05 161.80 164.74 163.760  163.5338
#> 193 2023-04-06 13:30:00  45390100 164.96 162.00 162.43 164.660  164.4326
#> 194 2023-04-10 13:30:00  47716900 162.03 160.08 161.42 162.030  161.8062
#> 195 2023-04-11 13:30:00  47644200 162.36 160.51 162.35 160.800  160.5779
#> 196 2023-04-12 13:30:00  50133100 162.06 159.78 161.22 160.100  159.8789
#> 197 2023-04-13 13:30:00  68445600 165.80 161.42 161.63 165.560  165.3313
#> 198 2023-04-14 13:30:00  49386500 166.32 163.82 164.59 165.210  164.9818
#> 199 2023-04-17 13:30:00  41516200 165.39 164.03 165.09 165.230  165.0018
#> 200 2023-04-18 13:30:00  49923000 167.41 165.65 166.10 166.470  166.2401
#> 201 2023-04-19 13:30:00  47720200 168.16 165.54 165.80 167.630  167.3985
#> 202 2023-04-20 13:30:00  52456400 167.87 165.56 166.09 166.650  166.4198
#> 203 2023-04-21 13:30:00  58337300 166.45 164.49 165.05 165.020  164.7921
#> 204 2023-04-24 13:30:00  41949600 165.60 163.89 165.00 165.330  165.1016
#> 205 2023-04-25 13:30:00  48714100 166.31 163.73 165.19 163.770  163.5438
#> 206 2023-04-26 13:30:00  45498800 165.28 162.80 163.06 163.760  163.5338
#> 207 2023-04-27 13:30:00  64902300 168.56 165.19 165.19 168.410  168.1774
#> 208 2023-04-28 13:30:00  55209200 169.85 167.88 168.49 169.680  169.4456
#> 209 2023-05-01 13:30:00  52472900 170.45 168.64 169.28 169.590  169.3557
#> 210 2023-05-02 13:30:00  48425700 170.35 167.54 170.09 168.540  168.3072
#> 211 2023-05-03 13:30:00  65136000 170.92 167.16 169.50 167.450  167.2187
#> 212 2023-05-04 13:30:00  81235400 167.04 164.31 164.89 165.790  165.5610
#> 213 2023-05-05 13:30:00 113316400 174.30 170.76 170.98 173.570  173.3303
#> 214 2023-05-08 13:30:00  55962800 173.85 172.11 172.48 173.500  173.2603
#> 215 2023-05-09 13:30:00  45326900 173.54 171.60 173.05 171.770  171.5327
#> 216 2023-05-10 13:30:00  53724500 174.03 171.90 173.02 173.560  173.3203
#> 217 2023-05-11 13:30:00  49514700 174.59 172.17 173.85 173.750  173.5100
#> 218 2023-05-12 13:30:00  45497800 174.06 171.00 173.62 172.570  172.5700
#> 219 2023-05-15 13:30:00  37266700 173.21 171.47 173.16 172.070  172.0700
#> 220 2023-05-16 13:30:00  42110300 173.14 171.80 171.99 172.070  172.0700
#> 221 2023-05-17 13:30:00  57951600 172.93 170.42 171.71 172.690  172.6900
#> 222 2023-05-18 13:30:00  65496700 175.24 172.58 173.00 175.050  175.0500
#> 223 2023-05-19 13:30:00  55772400 176.39 174.94 176.39 175.160  175.1600
#> 224 2023-05-22 13:30:00  43570900 174.71 173.45 173.98 174.200  174.2000
#> 225 2023-05-23 13:30:00  50747300 173.38 171.28 173.13 171.560  171.5600
#> 226 2023-05-24 13:30:00  45143500 172.42 170.52 171.09 171.840  171.8400
#> 227 2023-05-25 13:30:00  56058300 173.90 171.69 172.41 172.990  172.9900
#> 228 2023-05-26 13:30:00  54835000 175.77 173.11 173.32 175.430  175.4300
#> 229 2023-05-30 13:30:00  55964400 178.99 176.57 176.96 177.300  177.3000
#> 230 2023-05-31 13:30:00  99625300 179.35 176.76 177.33 177.250  177.2500
#> 231 2023-06-01 13:30:00  68901800 180.12 176.93 177.70 180.090  180.0900
#> 232 2023-06-02 13:30:00  61945900 181.78 179.26 181.03 180.950  180.9500
#> 233 2023-06-05 13:30:00 121946500 184.95 178.04 182.63 179.580  179.5800
#> 234 2023-06-06 13:30:00  64848400 180.12 177.43 179.97 179.210  179.2100
#> 235 2023-06-07 13:30:00  61944600 181.21 177.32 178.44 177.820  177.8200
#> 236 2023-06-08 13:30:00  50214900 180.84 177.46 177.90 180.570  180.5700
#> 237 2023-06-09 13:30:00  48870700 182.23 180.63 181.50 180.960  180.9600
#> 238 2023-06-12 13:30:00  54274900 183.89 180.97 181.27 183.790  183.7900
#> 239 2023-06-13 13:30:00  54929100 184.15 182.44 182.80 183.310  183.3100
#> 240 2023-06-14 13:30:00  57462900 184.39 182.02 183.37 183.950  183.9500
#> 241 2023-06-15 13:30:00  65433200 186.52 183.78 183.96 186.010  186.0100
#> 242 2023-06-16 13:30:00 101235600 186.99 184.27 186.73 184.920  184.9200
#> 243 2023-06-20 13:30:00  49799100 186.10 184.41 184.41 185.010  185.0100
#> 244 2023-06-21 13:30:00  49515700 185.41 182.59 184.90 183.960  183.9600
#> 245 2023-06-22 13:30:00  51245300 187.05 183.67 183.74 187.000  187.0000
#> 246 2023-06-23 13:30:00  53079300 187.56 185.01 185.55 186.680  186.6800
#> 247 2023-06-26 13:30:00  48088700 188.05 185.23 186.83 185.270  185.2700
#> 248 2023-06-27 13:30:00  50730800 188.39 185.67 185.89 188.060  188.0600
#> 249 2023-06-28 13:30:00  51216800 189.90 187.60 187.93 189.250  189.2500
#> 250 2023-06-29 13:30:00  46347300 190.07 188.94 189.08 189.590  189.5900
#> 251 2023-06-30 13:30:00  85069600 194.48 191.26 191.63 193.970  193.9700
#> 252 2023-07-03 13:30:00  31458200 193.88 191.76 193.78 192.460  192.4600
#> 253 2023-07-05 13:30:00  46920300 192.98 190.62 191.57 191.330  191.3300
#> 254 2023-07-06 13:30:00  45094300 192.02 189.20 189.84 191.810  191.8100
#> 255 2023-07-07 13:30:00  46778000 192.67 190.24 191.41 190.680  190.6800
#> 256 2023-07-10 13:30:00  59922200 189.99 187.04 189.26 188.610  188.6100
#> 257 2023-07-11 13:30:00  46638100 189.30 186.60 189.16 188.080  188.0800
#> 258 2023-07-12 13:30:00  60750200 191.70 188.47 189.68 189.770  189.7700
#> 259 2023-07-13 13:30:00  41342300 191.19 189.78 190.50 190.540  190.5400
#> 260 2023-07-14 13:30:00  41573900 191.18 189.63 190.23 190.690  190.6900
#> 261 2023-07-17 13:30:00  50520200 194.32 191.81 191.90 193.990  193.9900
#> 262 2023-07-18 13:30:00  48353800 194.33 192.42 193.35 193.730  193.7300
#> 263 2023-07-19 13:30:00  80507300 198.23 192.65 193.10 195.100  195.1000
#> 264 2023-07-20 13:30:00  59581200 196.47 192.50 195.09 193.130  193.1300
#> 265 2023-07-21 13:30:00  71917800 194.97 191.23 194.10 191.940  191.9400
#> 266 2023-07-24 13:30:00  45377800 194.91 192.25 193.41 192.750  192.7500
#> 267 2023-07-25 13:30:00  37283200 194.44 192.92 193.33 193.620  193.6200
#> 268 2023-07-26 13:30:00  47471900 195.64 193.32 193.67 194.500  194.5000
#> 269 2023-07-27 13:30:00  47460200 197.20 192.55 196.02 193.220  193.2200
#> 270 2023-07-28 13:30:00  48291400 196.63 194.14 194.67 195.830  195.8300
#> 271 2023-07-31 13:30:00  38824100 196.49 195.26 196.06 196.450  196.4500
#> 272 2023-08-01 13:30:00  35175100 196.73 195.28 196.24 195.610  195.6100
#> 273 2023-08-02 13:30:00  50389300 195.18 191.85 195.04 192.580  192.5800
#> 274 2023-08-03 13:30:00  61235200 192.37 190.69 191.57 191.170  191.1700
#> 275 2023-08-04 13:30:00 115799700 187.38 181.92 185.52 181.990  181.9900
#> 276 2023-08-07 13:52:25  20442337 183.13 178.75 182.13 179.075  179.0750
aapl$get_history(start = '2022-07-01', end = '2022-07-14', interval = '1d')
#>                  date   volume   high    low   open  close adj_close
#> 1 2022-07-01 13:30:00 71051600 139.04 135.66 136.04 138.93  138.1053
#> 2 2022-07-05 13:30:00 73353800 141.61 136.93 137.77 141.56  140.7197
#> 3 2022-07-06 13:30:00 74064300 144.12 141.08 141.35 142.92  142.0717
#> 4 2022-07-07 13:30:00 66253700 146.55 143.28 143.29 146.35  145.4813
#> 5 2022-07-08 13:30:00 64547800 147.55 145.00 145.26 147.04  146.1672
#> 6 2022-07-11 13:30:00 63141600 146.64 143.78 145.67 144.87  144.0101
#> 7 2022-07-12 13:30:00 77588800 148.45 145.05 145.76 145.86  144.9942
#> 8 2022-07-13 13:30:00 71185600 146.45 142.12 142.99 145.49  144.6264
aapl$get_history(period = '1mo', interval = '1d')
#>                   date    volume   high    low   open  close adj_close
#> 1  2023-07-07 13:30:00  46778000 192.67 190.24 191.41 190.68    190.68
#> 2  2023-07-10 13:30:00  59922200 189.99 187.04 189.26 188.61    188.61
#> 3  2023-07-11 13:30:00  46638100 189.30 186.60 189.16 188.08    188.08
#> 4  2023-07-12 13:30:00  60750200 191.70 188.47 189.68 189.77    189.77
#> 5  2023-07-13 13:30:00  41342300 191.19 189.78 190.50 190.54    190.54
#> 6  2023-07-14 13:30:00  41573900 191.18 189.63 190.23 190.69    190.69
#> 7  2023-07-17 13:30:00  50520200 194.32 191.81 191.90 193.99    193.99
#> 8  2023-07-18 13:30:00  48353800 194.33 192.42 193.35 193.73    193.73
#> 9  2023-07-19 13:30:00  80507300 198.23 192.65 193.10 195.10    195.10
#> 10 2023-07-20 13:30:00  59581200 196.47 192.50 195.09 193.13    193.13
#> 11 2023-07-21 13:30:00  71917800 194.97 191.23 194.10 191.94    191.94
#> 12 2023-07-24 13:30:00  45377800 194.91 192.25 193.41 192.75    192.75
#> 13 2023-07-25 13:30:00  37283200 194.44 192.92 193.33 193.62    193.62
#> 14 2023-07-26 13:30:00  47471900 195.64 193.32 193.67 194.50    194.50
#> 15 2023-07-27 13:30:00  47460200 197.20 192.55 196.02 193.22    193.22
#> 16 2023-07-28 13:30:00  48291400 196.63 194.14 194.67 195.83    195.83
#> 17 2023-07-31 13:30:00  38824100 196.49 195.26 196.06 196.45    196.45
#> 18 2023-08-01 13:30:00  35175100 196.73 195.28 196.24 195.61    195.61
#> 19 2023-08-02 13:30:00  50389300 195.18 191.85 195.04 192.58    192.58
#> 20 2023-08-03 13:30:00  61235200 192.37 190.69 191.57 191.17    191.17
#> 21 2023-08-04 13:30:00 115799700 187.38 181.92 185.52 181.99    181.99
#> 22 2023-08-07 13:52:26  20462666 183.13 178.75 182.13 179.06    179.06
# }