R6 Class Representing a Ticker

Base class for getting all data related to ticker from Yahoo Finance API.

Format

An R6 class object

Public fields

symbol

Symbol for which data is retrieved.

Active bindings

valuation_measures

Retrieves valuation measures for most recent four quarters

recommendations

Recommended symbols

technical_insights

Technical indicators for given symbol

currency

Currency

exchange_name

Exchange name

full_exchange_name

Full exchange name

first_trade_date

First trade date

regular_market_time

Regular market time

timezone

Time zone

exchange_timezone_name

Exchange timezone name

regular_market_price

Regular market price

fifty_two_week_high

Fifty two week high

fifty_two_week_low

Fifty two week low

regular_market_day_high

Regular market day high

regular_market_day_low

Regular market day low

regular_market_volume

Regular market volume

previous_close

Previous close

Methods

Public methods

Ticker$new()

Ticker$set_symbol()

Ticker$get_history()

Ticker$clone()

Method new()

Create a new Ticker object.

Usage

Ticker$new(symbol = NA)

Arguments

symbol

Symbol.

Returns

A new `Ticker` object

Examples

aapl <- Ticker$new('aapl')

Method set_symbol()

Set a new symbol.

Usage

Ticker$set_symbol(symbol)

Arguments

symbol

New symbol

Examples

aapl <- Ticker$new('aapl')
aapl$set_symbol('msft')

Method get_history()

Retrieves historical pricing data.

Usage

Ticker$get_history(period = "ytd", interval = "1d", start = NULL, end = NULL)

Arguments

period

Length of time. Defaults to 'ytd'. Valid values are:

'1d'

'5d'

'1mo'

'3mo'

'6mo'

'1y'

'2y'

'5y'

'10y'

'ytd'

'max'

interval

Time between data points. Defaults to '1d'. Valid values are:

'1m'

'2m'

'5m'

'15m'

'30m'

'60m'

'90m'

'1h'

'1d'

'5d'

'1wk'

'1mo'

'3mo'

start

Specific starting date. String or date object in yyyy-mm-dd format.

end

Specific ending date. String or date object in yyyy-mm-dd format.

Returns

A data.frame.

Examples

\donttest{
aapl <- Ticker$new('aapl')
aapl$get_history(start = '2022-07-01', interval = '1d')
aapl$get_history(start = '2022-07-01', end = '2022-07-14', interval = '1d')
aapl$get_history(period = '1mo', interval = '1d')
}

Method clone()

The objects of this class are cloneable with this method.

Usage

Ticker$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `Ticker$new`
## ------------------------------------------------

aapl <- Ticker$new('aapl')

## ------------------------------------------------
## Method `Ticker$set_symbol`
## ------------------------------------------------

aapl <- Ticker$new('aapl')
aapl$set_symbol('msft')

## ------------------------------------------------
## Method `Ticker$get_history`
## ------------------------------------------------

# \donttest{
aapl <- Ticker$new('aapl')
aapl$get_history(start = '2022-07-01', interval = '1d')
#>                    date    volume   high    low   open  close adj_close
#> 1   2022-07-01 13:30:00  71051600 139.04 135.66 136.04 138.93  137.0661
#> 2   2022-07-05 13:30:00  73353800 141.61 136.93 137.77 141.56  139.6609
#> 3   2022-07-06 13:30:00  74064300 144.12 141.08 141.35 142.92  141.0026
#> 4   2022-07-07 13:30:00  66253700 146.55 143.28 143.29 146.35  144.3866
#> 5   2022-07-08 13:30:00  64547800 147.55 145.00 145.26 147.04  145.0673
#> 6   2022-07-11 13:30:00  63141600 146.64 143.78 145.67 144.87  142.9265
#> 7   2022-07-12 13:30:00  77588800 148.45 145.05 145.76 145.86  143.9032
#> 8   2022-07-13 13:30:00  71185600 146.45 142.12 142.99 145.49  143.5381
#> 9   2022-07-14 13:30:00  78140700 148.95 143.25 144.08 148.47  146.4781
#> 10  2022-07-15 13:30:00  76259900 150.86 148.20 149.78 150.17  148.1553
#> 11  2022-07-18 13:30:00  81420900 151.57 146.70 150.74 147.07  145.0969
#> 12  2022-07-19 13:30:00  82982400 151.23 146.91 147.92 151.00  148.9742
#> 13  2022-07-20 13:30:00  64823400 153.72 150.37 151.12 153.04  150.9868
#> 14  2022-07-21 13:30:00  65086600 155.57 151.94 154.50 155.35  153.2659
#> 15  2022-07-22 13:30:00  66675400 156.28 153.41 155.39 154.09  152.0228
#> 16  2022-07-25 13:30:00  53623900 155.04 152.28 154.01 152.95  150.8980
#> 17  2022-07-26 13:30:00  55138700 153.09 150.80 152.26 151.60  149.5662
#> 18  2022-07-27 13:30:00  78620700 157.33 152.16 152.58 156.79  154.6865
#> 19  2022-07-28 13:30:00  81378700 157.64 154.41 156.98 157.35  155.2390
#> 20  2022-07-29 13:30:00 101786900 163.63 159.50 161.24 162.51  160.3298
#> 21  2022-08-01 13:30:00  67829400 163.59 160.89 161.01 161.51  159.3432
#> 22  2022-08-02 13:30:00  59907000 162.41 159.63 160.10 160.01  157.8633
#> 23  2022-08-03 13:30:00  82507500 166.59 160.75 160.84 166.13  163.9012
#> 24  2022-08-04 13:30:00  55474100 167.19 164.43 166.01 165.81  163.5855
#> 25  2022-08-05 13:30:00  56697000 165.85 163.00 163.21 165.35  163.3583
#> 26  2022-08-08 13:30:00  60276900 167.81 164.20 166.37 164.87  162.8841
#> 27  2022-08-09 13:30:00  63135500 165.82 163.25 164.02 164.92  162.9335
#> 28  2022-08-10 13:30:00  70170500 169.34 166.90 167.68 169.24  167.2014
#> 29  2022-08-11 13:30:00  57149200 170.99 168.19 170.06 168.49  166.4605
#> 30  2022-08-12 13:30:00  68039400 172.17 169.40 169.82 172.10  170.0270
#> 31  2022-08-15 13:30:00  54091700 173.39 171.35 171.52 173.19  171.1039
#> 32  2022-08-16 13:30:00  56377100 173.71 171.66 172.78 173.03  170.9458
#> 33  2022-08-17 13:30:00  79542000 176.15 172.57 172.77 174.55  172.4475
#> 34  2022-08-18 13:30:00  62290100 174.90 173.12 173.75 174.15  172.0523
#> 35  2022-08-19 13:30:00  70346300 173.74 171.31 173.03 171.52  169.4540
#> 36  2022-08-22 13:30:00  69026800 169.86 167.14 169.69 167.57  165.5516
#> 37  2022-08-23 13:30:00  54147100 168.71 166.65 167.08 167.23  165.2157
#> 38  2022-08-24 13:30:00  53841500 168.11 166.25 167.32 167.53  165.5120
#> 39  2022-08-25 13:30:00  51218200 170.14 168.35 168.78 170.03  167.9819
#> 40  2022-08-26 13:30:00  78961000 171.05 163.56 170.57 163.62  161.6491
#> 41  2022-08-29 13:30:00  73314000 162.90 159.82 161.15 161.38  159.4361
#> 42  2022-08-30 13:30:00  77906200 162.56 157.72 162.13 158.91  156.9959
#> 43  2022-08-31 13:30:00  87991100 160.58 157.14 160.31 157.22  155.3262
#> 44  2022-09-01 13:30:00  74229900 158.42 154.67 156.64 157.96  156.0573
#> 45  2022-09-02 13:30:00  76957800 160.36 154.97 159.75 155.81  153.9332
#> 46  2022-09-06 13:30:00  73714800 157.09 153.69 156.47 154.53  152.6686
#> 47  2022-09-07 13:30:00  87449600 156.67 153.61 154.82 155.96  154.0814
#> 48  2022-09-08 13:30:00  84923800 156.36 152.68 154.64 154.46  152.5995
#> 49  2022-09-09 13:30:00  68028800 157.82 154.75 155.47 157.37  155.4744
#> 50  2022-09-12 13:30:00 104956000 164.26 159.30 159.59 163.43  161.4614
#> 51  2022-09-13 13:30:00 122656600 160.54 153.37 159.90 153.84  151.9870
#> 52  2022-09-14 13:30:00  87965400 157.10 153.61 154.79 155.31  153.4392
#> 53  2022-09-15 13:30:00  90481100 155.24 151.38 154.65 152.37  150.5346
#> 54  2022-09-16 13:30:00 162278800 151.35 148.37 151.21 150.70  148.8848
#> 55  2022-09-19 13:30:00  81474200 154.56 149.10 149.31 154.48  152.6192
#> 56  2022-09-20 13:30:00 107689800 158.08 153.08 153.40 156.90  155.0101
#> 57  2022-09-21 13:30:00 101696800 158.74 153.60 157.34 153.72  151.8684
#> 58  2022-09-22 13:30:00  86652500 154.47 150.91 152.38 152.74  150.9002
#> 59  2022-09-23 13:30:00  96029900 151.47 148.56 151.19 150.43  148.6180
#> 60  2022-09-26 13:30:00  93339400 153.77 149.64 149.66 150.77  148.9539
#> 61  2022-09-27 13:30:00  84442700 154.72 149.95 152.74 151.76  149.9320
#> 62  2022-09-28 13:30:00 146691400 150.64 144.84 147.64 149.84  148.0351
#> 63  2022-09-29 13:30:00 128138200 146.72 140.68 146.10 142.48  140.7638
#> 64  2022-09-30 13:30:00 124925300 143.10 138.00 141.28 138.20  136.5353
#> 65  2022-10-03 13:30:00 114311700 143.07 137.69 138.21 142.45  140.7341
#> 66  2022-10-04 13:30:00  87830100 146.22 144.26 145.03 146.10  144.3402
#> 67  2022-10-05 13:30:00  79471000 147.38 143.01 144.07 146.40  144.6366
#> 68  2022-10-06 13:30:00  68402200 147.54 145.22 145.81 145.43  143.6783
#> 69  2022-10-07 13:30:00  85925600 143.10 139.45 142.54 140.09  138.4026
#> 70  2022-10-10 13:30:00  74899000 141.89 138.57 140.42 140.42  138.7286
#> 71  2022-10-11 13:30:00  77033700 141.35 138.22 139.90 138.98  137.3059
#> 72  2022-10-12 13:30:00  70433700 140.36 138.16 139.13 138.34  136.6736
#> 73  2022-10-13 13:30:00 113224000 143.59 134.37 134.99 142.99  141.2676
#> 74  2022-10-14 13:30:00  88598000 144.52 138.19 144.31 138.38  136.7132
#> 75  2022-10-17 13:30:00  85250900 142.90 140.27 141.07 142.41  140.6946
#> 76  2022-10-18 13:30:00  99136600 146.70 140.61 145.49 143.75  142.0185
#> 77  2022-10-19 13:30:00  61758300 144.95 141.50 141.69 143.86  142.1272
#> 78  2022-10-20 13:30:00  64522000 145.89 142.65 143.02 143.39  141.6628
#> 79  2022-10-21 13:30:00  86548600 147.85 142.65 142.87 147.27  145.4961
#> 80  2022-10-24 13:30:00  75981900 150.23 146.00 147.19 149.45  147.6498
#> 81  2022-10-25 13:30:00  74732300 152.49 149.36 150.09 152.34  150.5050
#> 82  2022-10-26 13:30:00  88194300 151.99 148.04 150.96 149.35  147.5510
#> 83  2022-10-27 13:30:00 109180200 149.05 144.13 148.07 144.80  143.0558
#> 84  2022-10-28 13:30:00 164762400 157.50 147.82 148.20 155.74  153.8640
#> 85  2022-10-31 13:30:00  97943200 154.24 151.92 153.16 153.34  151.4930
#> 86  2022-11-01 13:30:00  80379300 155.45 149.13 155.08 150.65  148.8354
#> 87  2022-11-02 13:30:00  93604600 152.17 145.00 148.95 145.03  143.2831
#> 88  2022-11-03 13:30:00  97918500 142.80 138.75 142.06 138.88  137.2072
#> 89  2022-11-04 13:30:00 140814800 142.67 134.38 142.09 138.38  136.9400
#> 90  2022-11-07 14:30:00  83374600 139.15 135.67 137.11 138.92  137.4743
#> 91  2022-11-08 14:30:00  89908500 141.43 137.49 140.41 139.50  138.0483
#> 92  2022-11-09 14:30:00  74917800 138.55 134.59 138.50 134.87  133.4665
#> 93  2022-11-10 14:30:00 118854000 146.87 139.50 141.24 146.87  145.3416
#> 94  2022-11-11 14:30:00  93979700 150.01 144.37 145.82 149.70  148.1422
#> 95  2022-11-14 14:30:00  73374100 150.28 147.43 148.97 148.28  146.7369
#> 96  2022-11-15 14:30:00  89868300 153.59 148.56 152.22 150.04  148.4786
#> 97  2022-11-16 14:30:00  64218300 149.87 147.29 149.13 148.79  147.2416
#> 98  2022-11-17 14:30:00  80389400 151.48 146.15 146.43 150.72  149.1515
#> 99  2022-11-18 14:30:00  74829600 152.70 149.97 152.31 151.29  149.7156
#> 100 2022-11-21 14:30:00  58724100 150.37 147.72 150.16 148.01  146.4697
#> 101 2022-11-22 14:30:00  51804100 150.42 146.93 148.13 150.18  148.6171
#> 102 2022-11-23 14:30:00  58301400 151.83 149.34 149.45 151.07  149.4979
#> 103 2022-11-25 14:30:00  35195900 148.88 147.12 148.31 148.11  146.5687
#> 104 2022-11-28 14:30:00  69246000 146.64 143.38 145.14 144.22  142.7192
#> 105 2022-11-29 14:30:00  83763800 144.81 140.35 144.29 141.17  139.7009
#> 106 2022-11-30 14:30:00 111380900 148.72 140.55 141.40 148.03  146.4895
#> 107 2022-12-01 14:30:00  71250400 149.13 146.61 148.21 148.31  146.7666
#> 108 2022-12-02 14:30:00  65447400 148.00 145.65 145.96 147.81  146.2718
#> 109 2022-12-05 14:30:00  68826400 150.92 145.77 147.77 146.63  145.1041
#> 110 2022-12-06 14:30:00  64727200 147.30 141.92 147.07 142.91  141.4228
#> 111 2022-12-07 14:30:00  69721100 143.37 140.00 142.19 140.94  139.4733
#> 112 2022-12-08 14:30:00  62128300 143.52 141.10 142.36 142.65  141.1655
#> 113 2022-12-09 14:30:00  76097000 145.57 140.90 142.34 142.16  140.6806
#> 114 2022-12-12 14:30:00  70462700 144.50 141.06 142.70 144.49  142.9864
#> 115 2022-12-13 14:30:00  93886200 149.97 144.24 149.50 145.47  143.9562
#> 116 2022-12-14 14:30:00  82291200 146.66 141.16 145.35 143.21  141.7197
#> 117 2022-12-15 14:30:00  98931900 141.80 136.03 141.11 136.50  135.0795
#> 118 2022-12-16 14:30:00 160156900 137.65 133.73 136.69 134.51  133.1102
#> 119 2022-12-19 14:30:00  79592600 135.20 131.32 135.11 132.37  130.9925
#> 120 2022-12-20 14:30:00  77432800 133.25 129.89 131.39 132.30  130.9232
#> 121 2022-12-21 14:30:00  85928000 136.81 132.75 132.98 135.45  134.0404
#> 122 2022-12-22 14:30:00  77852100 134.56 130.30 134.35 132.23  130.8539
#> 123 2022-12-23 14:30:00  63814900 132.42 129.64 130.92 131.86  130.4878
#> 124 2022-12-27 14:30:00  69007800 131.41 128.72 131.38 130.03  128.6768
#> 125 2022-12-28 14:30:00  85438400 131.03 125.87 129.67 126.04  124.7284
#> 126 2022-12-29 14:30:00  75703700 130.48 127.73 127.99 129.61  128.2612
#> 127 2022-12-30 14:30:00  77034200 129.95 127.43 128.41 129.93  128.5778
#> 128 2023-01-03 14:30:00 112117500 130.90 124.17 130.28 125.07  123.7685
#> 129 2023-01-04 14:30:00  89113600 128.66 125.08 126.89 126.36  125.0450
#> 130 2023-01-05 14:30:00  80962700 127.77 124.76 127.13 125.02  123.7190
#> 131 2023-01-06 14:30:00  87754700 130.29 124.89 126.01 129.62  128.2711
#> 132 2023-01-09 14:30:00  70790800 133.41 129.89 130.47 130.15  128.7956
#> 133 2023-01-10 14:30:00  63896200 131.26 128.12 130.26 130.73  129.3696
#> 134 2023-01-11 14:30:00  69458900 133.51 130.46 131.25 133.49  132.1008
#> 135 2023-01-12 14:30:00  71379600 134.26 131.44 133.88 133.41  132.0217
#> 136 2023-01-13 14:30:00  57809700 134.92 131.66 132.03 134.76  133.3576
#> 137 2023-01-17 14:30:00  63646600 137.29 134.13 134.83 135.94  134.5253
#> 138 2023-01-18 14:30:00  69672800 138.61 135.03 136.82 135.21  133.8029
#> 139 2023-01-19 14:30:00  58280400 136.25 133.77 134.08 135.27  133.8623
#> 140 2023-01-20 14:30:00  80223600 138.02 134.22 135.28 137.87  136.4352
#> 141 2023-01-23 14:30:00  81760300 143.32 137.90 138.12 141.11  139.6415
#> 142 2023-01-24 14:30:00  66435100 143.16 140.30 140.31 142.53  141.0468
#> 143 2023-01-25 14:30:00  65799300 142.43 138.81 140.89 141.86  140.3837
#> 144 2023-01-26 14:30:00  54105100 144.25 141.90 143.17 143.96  142.4619
#> 145 2023-01-27 14:30:00  70555800 147.23 143.08 143.16 145.93  144.4114
#> 146 2023-01-30 14:30:00  64015300 145.55 142.85 144.96 143.00  141.5119
#> 147 2023-01-31 14:30:00  65874500 144.34 142.28 142.70 144.29  142.7884
#> 148 2023-02-01 14:30:00  77663600 146.61 141.32 143.97 145.43  143.9166
#> 149 2023-02-02 14:30:00 118339000 151.18 148.17 148.90 150.82  149.2505
#> 150 2023-02-03 14:30:00 154357300 157.38 147.83 148.03 154.50  152.8922
#> 151 2023-02-06 14:30:00  69858300 153.10 150.78 152.57 151.73  150.1510
#> 152 2023-02-07 14:30:00  83322600 155.23 150.64 150.64 154.65  153.0406
#> 153 2023-02-08 14:30:00  64120100 154.58 151.17 153.88 151.92  150.3391
#> 154 2023-02-09 14:30:00  56007100 154.33 150.42 153.78 150.87  149.3000
#> 155 2023-02-10 14:30:00  57450700 151.34 149.22 149.46 151.01  149.6667
#> 156 2023-02-13 14:30:00  62199000 154.26 150.92 150.95 153.85  152.4814
#> 157 2023-02-14 14:30:00  61707600 153.77 150.86 152.12 153.20  151.8372
#> 158 2023-02-15 14:30:00  65573800 155.50 152.88 153.11 155.33  153.9483
#> 159 2023-02-16 14:30:00  68167900 156.33 153.35 153.51 153.71  152.3427
#> 160 2023-02-17 14:30:00  59144100 153.00 150.85 152.35 152.55  151.1930
#> 161 2023-02-21 14:30:00  58867200 151.30 148.41 150.20 148.48  147.1592
#> 162 2023-02-22 14:30:00  51011300 149.95 147.16 148.87 148.91  147.5854
#> 163 2023-02-23 14:30:00  48394200 150.34 147.24 150.09 149.40  148.0710
#> 164 2023-02-24 14:30:00  55469600 147.19 145.72 147.11 146.71  145.4049
#> 165 2023-02-27 14:30:00  44998500 149.17 147.45 147.71 147.92  146.6042
#> 166 2023-02-28 14:30:00  50547000 149.08 146.83 147.05 147.41  146.0987
#> 167 2023-03-01 14:30:00  55479000 147.23 145.01 146.83 145.31  144.0174
#> 168 2023-03-02 14:30:00  52238100 146.71 143.90 144.38 145.91  144.6120
#> 169 2023-03-03 14:30:00  70732300 151.11 147.33 148.04 151.03  149.6865
#> 170 2023-03-06 14:30:00  87558000 156.30 153.46 153.79 153.83  152.4616
#> 171 2023-03-07 14:30:00  56182000 154.03 151.13 153.70 151.60  150.2514
#> 172 2023-03-08 14:30:00  47204800 153.47 151.83 152.81 152.87  151.5101
#> 173 2023-03-09 14:30:00  53833600 154.54 150.23 153.56 150.59  149.2504
#> 174 2023-03-10 14:30:00  68572400 150.94 147.61 150.21 148.50  147.1790
#> 175 2023-03-13 13:30:00  84457100 153.14 147.70 147.81 150.47  149.1315
#> 176 2023-03-14 13:30:00  73695900 153.40 150.10 151.28 152.59  151.2326
#> 177 2023-03-15 13:30:00  77167900 153.25 149.92 151.19 152.99  151.6291
#> 178 2023-03-16 13:30:00  76161100 156.46 151.64 152.16 155.85  154.4636
#> 179 2023-03-17 13:30:00  98944600 156.74 154.28 156.08 155.00  153.6212
#> 180 2023-03-20 13:30:00  73641400 157.82 154.15 155.07 157.40  155.9998
#> 181 2023-03-21 13:30:00  73938300 159.40 156.54 157.32 159.28  157.8631
#> 182 2023-03-22 13:30:00  75701800 162.14 157.81 159.30 157.83  156.4260
#> 183 2023-03-23 13:30:00  67622100 161.55 157.68 158.83 158.93  157.5162
#> 184 2023-03-24 13:30:00  59196500 160.34 157.85 158.86 160.25  158.8245
#> 185 2023-03-27 13:30:00  52390300 160.77 157.87 159.94 158.28  156.8720
#> 186 2023-03-28 13:30:00  45992200 158.49 155.98 157.97 157.65  156.2476
#> 187 2023-03-29 13:30:00  51305700 161.05 159.35 159.37 160.77  159.3399
#> 188 2023-03-30 13:30:00  49501700 162.47 161.27 161.53 162.36  160.9157
#> 189 2023-03-31 13:30:00  68749800 165.00 161.91 162.44 164.90  163.4331
#> 190 2023-04-03 13:30:00  56976200 166.29 164.22 164.27 166.17  164.6918
#> 191 2023-04-04 13:30:00  46278300 166.84 165.11 166.60 165.63  164.1566
#> 192 2023-04-05 13:30:00  51511700 165.05 161.80 164.74 163.76  162.3033
#> 193 2023-04-06 13:30:00  45390100 164.96 162.00 162.43 164.66  163.1953
#> 194 2023-04-10 13:30:00  47716900 162.03 160.08 161.42 162.03  160.5887
#> 195 2023-04-11 13:30:00  47644200 162.36 160.51 162.35 160.80  159.3696
#> 196 2023-04-12 13:30:00  50133100 162.06 159.78 161.22 160.10  158.6758
#> 197 2023-04-13 13:30:00  68445600 165.80 161.42 161.63 165.56  164.0872
#> 198 2023-04-14 13:30:00  49386500 166.32 163.82 164.59 165.21  163.7404
#> 199 2023-04-17 13:30:00  41516200 165.39 164.03 165.09 165.23  163.7602
#> 200 2023-04-18 13:30:00  49923000 167.41 165.65 166.10 166.47  164.9892
#> 201 2023-04-19 13:30:00  47720200 168.16 165.54 165.80 167.63  166.1389
#> 202 2023-04-20 13:30:00  52456400 167.87 165.56 166.09 166.65  165.1676
#> 203 2023-04-21 13:30:00  58337300 166.45 164.49 165.05 165.02  163.5521
#> 204 2023-04-24 13:30:00  41949600 165.60 163.89 165.00 165.33  163.8593
#> 205 2023-04-25 13:30:00  48714100 166.31 163.73 165.19 163.77  162.3132
#> 206 2023-04-26 13:30:00  45498800 165.28 162.80 163.06 163.76  162.3033
#> 207 2023-04-27 13:30:00  64902300 168.56 165.19 165.19 168.41  166.9119
#> 208 2023-04-28 13:30:00  55209200 169.85 167.88 168.49 169.68  168.1706
#> 209 2023-05-01 13:30:00  52472900 170.45 168.64 169.28 169.59  168.0814
#> 210 2023-05-02 13:30:00  48425700 170.35 167.54 170.09 168.54  167.0407
#> 211 2023-05-03 13:30:00  65136000 170.92 167.16 169.50 167.45  165.9604
#> 212 2023-05-04 13:30:00  81235400 167.04 164.31 164.89 165.79  164.3152
#> 213 2023-05-05 13:30:00 113316400 174.30 170.76 170.98 173.57  172.0260
#> 214 2023-05-08 13:30:00  55962800 173.85 172.11 172.48 173.50  171.9566
#> 215 2023-05-09 13:30:00  45326900 173.54 171.60 173.05 171.77  170.2420
#> 216 2023-05-10 13:30:00  53724500 174.03 171.90 173.02 173.56  172.0161
#> 217 2023-05-11 13:30:00  49514700 174.59 172.17 173.85 173.75  172.2044
#> 218 2023-05-12 13:30:00  45497800 174.06 171.00 173.62 172.57  171.2715
#> 219 2023-05-15 13:30:00  37266700 173.21 171.47 173.16 172.07  170.7752
#> 220 2023-05-16 13:30:00  42110300 173.14 171.80 171.99 172.07  170.7752
#> 221 2023-05-17 13:30:00  57951600 172.93 170.42 171.71 172.69  171.3906
#> 222 2023-05-18 13:30:00  65496700 175.24 172.58 173.00 175.05  173.7328
#> 223 2023-05-19 13:30:00  55772400 176.39 174.94 176.39 175.16  173.8420
#> 224 2023-05-22 13:30:00  43570900 174.71 173.45 173.98 174.20  172.8892
#> 225 2023-05-23 13:30:00  50747300 173.38 171.28 173.13 171.56  170.2691
#> 226 2023-05-24 13:30:00  45143500 172.42 170.52 171.09 171.84  170.5470
#> 227 2023-05-25 13:30:00  56058300 173.90 171.69 172.41 172.99  171.6883
#> 228 2023-05-26 13:30:00  54835000 175.77 173.11 173.32 175.43  174.1099
#> 229 2023-05-30 13:30:00  55964400 178.99 176.57 176.96 177.30  175.9659
#> 230 2023-05-31 13:30:00  99625300 179.35 176.76 177.33 177.25  175.9162
#> 231 2023-06-01 13:30:00  68901800 180.12 176.93 177.70 180.09  178.7349
#> 232 2023-06-02 13:30:00  61945900 181.78 179.26 181.03 180.95  179.5884
#> 233 2023-06-05 13:30:00 121946500 184.95 178.04 182.63 179.58  178.2287
#> 234 2023-06-06 13:30:00  64848400 180.12 177.43 179.97 179.21  177.8615
#> 235 2023-06-07 13:30:00  61944600 181.21 177.32 178.44 177.82  176.4820
#> 236 2023-06-08 13:30:00  50214900 180.84 177.46 177.90 180.57  179.2113
#> 237 2023-06-09 13:30:00  48870700 182.23 180.63 181.50 180.96  179.5983
#> 238 2023-06-12 13:30:00  54274900 183.89 180.97 181.27 183.79  182.4070
#> 239 2023-06-13 13:30:00  54929100 184.15 182.44 182.80 183.31  181.9306
#> 240 2023-06-14 13:30:00  57462900 184.39 182.02 183.37 183.95  182.5658
#> 241 2023-06-15 13:30:00  65433200 186.52 183.78 183.96 186.01  184.6104
#> 242 2023-06-16 13:30:00 101235600 186.99 184.27 186.73 184.92  183.5285
#> 243 2023-06-20 13:30:00  49799100 186.10 184.41 184.41 185.01  183.6178
#> 244 2023-06-21 13:30:00  49515700 185.41 182.59 184.90 183.96  182.5758
#> 245 2023-06-22 13:30:00  51245300 187.05 183.67 183.74 187.00  185.5929
#> 246 2023-06-23 13:30:00  53079300 187.56 185.01 185.55 186.68  185.2753
#> 247 2023-06-26 13:30:00  48088700 188.05 185.23 186.83 185.27  183.8759
#> 248 2023-06-27 13:30:00  50730800 188.39 185.67 185.89 188.06  186.6449
#> 249 2023-06-28 13:30:00  51216800 189.90 187.60 187.93 189.25  187.8260
#> 250 2023-06-29 13:30:00  46347300 190.07 188.94 189.08 189.59  188.1634
#> 251 2023-06-30 13:30:00  85069600 194.48 191.26 191.63 193.97  192.5104
#> 252 2023-07-03 13:30:00  31458200 193.88 191.76 193.78 192.46  191.0118
#> 253 2023-07-05 13:30:00  46920300 192.98 190.62 191.57 191.33  189.8903
#> 254 2023-07-06 13:30:00  45094300 192.02 189.20 189.84 191.81  190.3667
#> 255 2023-07-07 13:30:00  46778000 192.67 190.24 191.41 190.68  189.2452
#> 256 2023-07-10 13:30:00  59922200 189.99 187.04 189.26 188.61  187.1908
#> 257 2023-07-11 13:30:00  46638100 189.30 186.60 189.16 188.08  186.6648
#> 258 2023-07-12 13:30:00  60750200 191.70 188.47 189.68 189.77  188.3420
#> 259 2023-07-13 13:30:00  41342300 191.19 189.78 190.50 190.54  189.1062
#> 260 2023-07-14 13:30:00  41573900 191.18 189.63 190.23 190.69  189.2551
#> 261 2023-07-17 13:30:00  50520200 194.32 191.81 191.90 193.99  192.5303
#> 262 2023-07-18 13:30:00  48353800 194.33 192.42 193.35 193.73  192.2722
#> 263 2023-07-19 13:30:00  80507300 198.23 192.65 193.10 195.10  193.6319
#> 264 2023-07-20 13:30:00  59581200 196.47 192.50 195.09 193.13  191.6768
#> 265 2023-07-21 13:30:00  71917800 194.97 191.23 194.10 191.94  190.4957
#> 266 2023-07-24 13:30:00  45377800 194.91 192.25 193.41 192.75  191.2996
#> 267 2023-07-25 13:30:00  37283200 194.44 192.92 193.33 193.62  192.1631
#> 268 2023-07-26 13:30:00  47471900 195.64 193.32 193.67 194.50  193.0364
#> 269 2023-07-27 13:30:00  47460200 197.20 192.55 196.02 193.22  191.7661
#> 270 2023-07-28 13:30:00  48291400 196.63 194.14 194.67 195.83  194.3565
#> 271 2023-07-31 13:30:00  38824100 196.49 195.26 196.06 196.45  194.9718
#> 272 2023-08-01 13:30:00  35175100 196.73 195.28 196.24 195.61  194.1381
#> 273 2023-08-02 13:30:00  50389300 195.18 191.85 195.04 192.58  191.1309
#> 274 2023-08-03 13:30:00  61235200 192.37 190.69 191.57 191.17  189.7315
#> 275 2023-08-04 13:30:00 115799700 187.38 181.92 185.52 181.99  180.6206
#> 276 2023-08-07 13:30:00  97576100 183.13 177.35 182.13 178.85  177.5042
#> 277 2023-08-08 13:30:00  67823000 180.27 177.58 179.69 179.80  178.4471
#> 278 2023-08-09 13:30:00  60378500 180.93 177.01 180.87 178.19  176.8492
#> 279 2023-08-10 13:30:00  54686900 180.75 177.60 179.48 177.97  176.6308
#> 280 2023-08-11 13:30:00  51988100 178.62 176.55 177.32 177.79  176.6905
#> 281 2023-08-14 13:30:00  43675600 179.69 177.31 177.97 179.46  178.3502
#> 282 2023-08-15 13:30:00  43622600 179.48 177.05 178.88 177.45  176.3526
#> 283 2023-08-16 13:30:00  46964900 178.54 176.50 177.13 176.57  175.4780
#> 284 2023-08-17 13:30:00  66062900 177.51 173.48 177.14 174.00  172.9239
#> 285 2023-08-18 13:30:00  61114200 175.10 171.96 172.30 174.49  173.4109
#> 286 2023-08-21 13:30:00  46311900 176.13 173.74 175.07 175.84  174.7525
#> 287 2023-08-22 13:30:00  42084200 177.68 176.25 177.06 177.23  176.1339
#> 288 2023-08-23 13:30:00  52722800 181.55 178.33 178.52 181.12  179.9999
#> 289 2023-08-24 13:30:00  54945800 181.10 176.01 180.67 176.38  175.2892
#> 290 2023-08-25 13:30:00  51449600 179.15 175.82 177.38 178.61  177.5054
#> 291 2023-08-28 13:30:00  43820700 180.59 178.55 180.09 180.19  179.0756
#> 292 2023-08-29 13:30:00  53003900 184.90 179.50 179.70 184.12  182.9813
#> 293 2023-08-30 13:30:00  60813900 187.85 184.74 184.94 187.65  186.4895
#> 294 2023-08-31 13:30:00  60794500 189.12 187.48 187.84 187.87  186.7081
#> 295 2023-09-01 13:30:00  45732600 189.92 188.28 189.49 189.46  188.2883
#> 296 2023-09-05 13:30:00  45280000 189.98 187.61 188.28 189.70  188.5268
#> 297 2023-09-06 13:30:00  81755800 188.85 181.47 188.40 182.91  181.7788
#> 298 2023-09-07 13:30:00 112488800 178.21 173.54 175.18 177.56  176.4619
#> 299 2023-09-08 13:30:00  65551300 180.24 177.79 178.35 178.18  177.0780
#> 300 2023-09-11 13:30:00  58953100 180.30 177.34 180.07 179.36  178.2508
#> 301 2023-09-12 13:30:00  90370200 180.13 174.82 179.49 176.30  175.2097
#> 302 2023-09-13 13:30:00  84267900 177.30 173.98 176.51 174.21  173.1326
#> 303 2023-09-14 13:30:00  60895800 176.10 173.58 174.00 175.74  174.6531
#> 304 2023-09-15 13:30:00 109205100 176.50 173.82 176.48 175.01  173.9277
#> 305 2023-09-18 13:30:00  67257600 179.38 176.17 176.48 177.97  176.8694
#> 306 2023-09-19 13:30:00  51826900 179.63 177.13 177.52 179.07  177.9626
#> 307 2023-09-20 13:30:00  58436200 179.70 175.40 179.26 175.49  174.4047
#> 308 2023-09-21 13:30:00  63047900 176.30 173.86 174.55 173.93  172.8543
#> 309 2023-09-22 13:30:00  56725400 177.08 174.05 174.67 174.79  173.7090
#> 310 2023-09-25 13:30:00  46172700 176.97 174.15 174.20 176.08  174.9911
#> 311 2023-09-26 13:30:00  64588900 175.20 171.66 174.82 171.96  170.8965
#> 312 2023-09-27 13:30:00  66921800 173.04 169.05 172.62 170.43  169.3760
#> 313 2023-09-28 13:30:00  56294400 172.03 167.62 169.34 170.69  169.6344
#> 314 2023-09-29 13:30:00  51814200 173.07 170.34 172.02 171.21  170.1512
#> 315 2023-10-02 13:30:00  52164500 174.30 170.93 171.22 173.75  172.6754
#> 316 2023-10-03 13:30:00  49594600 173.63 170.82 172.26 172.40  171.3338
#> 317 2023-10-04 13:30:00  53020300 174.21 170.97 171.09 173.66  172.5860
#> 318 2023-10-05 13:30:00  48527900 175.45 172.68 173.79 174.91  173.8283
#> 319 2023-10-06 13:30:00  57224100 177.99 173.18 173.80 177.49  176.3923
#> 320 2023-10-09 13:30:00  42390800 179.05 175.80 176.81 178.99  177.8830
#> 321 2023-10-10 13:30:00  43698000 179.72 177.95 178.10 178.39  177.2868
#> 322 2023-10-11 13:30:00  47551100 179.85 177.60 178.20 179.80  178.6880
#> 323 2023-10-12 13:30:00  56743100 182.34 179.04 180.07 180.71  179.5924
#> 324 2023-10-13 13:30:00  51427100 181.93 178.14 181.42 178.85  177.7439
#> 325 2023-10-16 13:30:00  52517000 179.08 176.51 176.75 178.72  177.6147
#> 326 2023-10-17 13:30:00  57549400 178.42 174.80 176.65 177.15  176.0544
#> 327 2023-10-18 13:30:00  54764400 177.58 175.11 175.58 175.84  174.7525
#> 328 2023-10-19 13:30:00  59302900 177.84 175.19 176.04 175.46  174.3749
#> 329 2023-10-20 13:30:00  64189300 175.42 172.64 175.31 172.88  171.8108
#> 330 2023-10-23 13:30:00  55980100 174.01 169.93 170.91 173.00  171.9301
#> 331 2023-10-24 13:30:00  43816600 173.67 171.45 173.05 173.44  172.3674
#> 332 2023-10-25 13:30:00  57157000 173.06 170.65 171.88 171.10  170.0418
#> 333 2023-10-26 13:30:00  70625300 171.38 165.67 170.37 166.89  165.8579
#> 334 2023-10-27 13:30:00  58499100 168.96 166.83 166.91 168.22  167.1796
#> 335 2023-10-30 13:30:00  51131000 171.17 168.87 169.02 170.29  169.2368
#> 336 2023-10-31 13:30:00  44846000 170.90 167.90 169.35 170.77  169.7139
#> 337 2023-11-01 13:30:00  56934900 174.23 170.12 171.00 173.97  172.8941
#> 338 2023-11-02 13:30:00  77334800 177.78 175.46 175.52 177.57  176.4718
#> 339 2023-11-03 13:30:00  79763700 176.82 173.35 174.24 176.65  175.5575
#> 340 2023-11-06 14:30:00  63841300 179.43 176.21 176.38 179.23  178.1216
#> 341 2023-11-07 14:30:00  70530000 182.44 178.97 179.18 181.82  180.6955
#> 342 2023-11-08 14:30:00  49340300 183.45 181.59 182.35 182.89  181.7589
#> 343 2023-11-09 14:30:00  53763500 184.12 181.81 182.96 182.41  181.2819
#> 344 2023-11-10 14:30:00  66133400 186.57 183.53 183.97 186.40  185.4913
#> 345 2023-11-13 14:30:00  43627500 186.03 184.21 185.82 184.80  183.8991
#> 346 2023-11-14 14:30:00  60108400 188.11 186.30 187.70 187.44  186.5262
#> 347 2023-11-15 14:30:00  53790500 189.50 187.78 187.85 188.01  187.0934
#> 348 2023-11-16 14:30:00  54412900 190.96 188.65 189.57 189.71  188.7851
#> 349 2023-11-17 14:30:00  50922700 190.38 188.57 190.25 189.69  188.7652
#> 350 2023-11-20 14:30:00  46505100 191.91 189.88 189.89 191.45  190.5166
#> 351 2023-11-21 14:30:00  38134500 191.52 189.74 191.41 190.64  189.7106
#> 352 2023-11-22 14:30:00  39617700 192.93 190.83 191.49 191.31  190.3773
#> 353 2023-11-24 14:30:00  24048300 190.90 189.25 190.87 189.97  189.0439
#> 354 2023-11-27 14:30:00  40552600 190.67 188.90 189.92 189.79  188.8647
#> 355 2023-11-28 14:30:00  38415400 191.08 189.40 189.78 190.40  189.4718
#> 356 2023-11-29 14:30:00  43014200 192.09 188.97 190.90 189.37  188.4468
#> 357 2023-11-30 14:30:00  48794400 190.32 188.19 189.84 189.95  189.0240
#> 358 2023-12-01 14:30:00  45679300 191.56 189.23 190.33 191.24  190.3077
#> 359 2023-12-04 14:30:00  43389500 190.05 187.45 189.98 189.43  188.5065
#> 360 2023-12-05 14:30:00  66628400 194.40 190.18 190.21 193.42  192.4771
#> 361 2023-12-06 14:30:00  41089700 194.76 192.11 194.45 192.32  191.3824
#> 362 2023-12-07 14:30:00  47477700 195.00 193.59 193.63 194.27  193.3229
#> 363 2023-12-08 14:30:00  53377300 195.99 193.67 194.20 195.71  194.7559
#> 364 2023-12-11 14:30:00  60943700 193.49 191.42 193.11 193.18  192.2382
#> 365 2023-12-12 14:30:00  52696900 194.72 191.72 193.08 194.71  193.7608
#> 366 2023-12-13 14:30:00  70404200 198.00 194.85 195.09 197.96  196.9949
#> 367 2023-12-14 14:30:00  66831600 199.62 196.16 198.02 198.11  197.1442
#> 368 2023-12-15 14:30:00 128256700 198.40 197.00 197.53 197.57  196.6068
#> 369 2023-12-18 14:30:00  55751900 196.63 194.39 196.09 195.89  194.9350
#> 370 2023-12-19 14:30:00  40714100 196.95 195.89 196.16 196.94  195.9799
#> 371 2023-12-20 14:30:00  52242800 197.68 194.83 196.90 194.83  193.8802
#> 372 2023-12-21 14:30:00  46482500 197.08 193.50 196.10 194.68  193.7309
#> 373 2023-12-22 14:30:00  37122800 195.41 192.97 195.18 193.60  192.6562
#> 374 2023-12-26 14:30:00  28919300 193.89 192.83 193.61 193.05  192.1089
#> 375 2023-12-27 14:30:00  48087700 193.50 191.09 192.49 193.15  192.2084
#> 376 2023-12-28 14:30:00  34049900 194.66 193.17 194.14 193.58  192.6363
#> 377 2023-12-29 14:30:00  42628800 194.40 191.73 193.90 192.53  191.5914
#> 378 2024-01-02 14:30:00  82488700 188.44 183.89 187.15 185.64  184.7350
#> 379 2024-01-03 14:30:00  58414500 185.88 183.43 184.22 184.25  183.3518
#> 380 2024-01-04 14:30:00  71983600 183.09 180.88 182.15 181.91  181.0232
#> 381 2024-01-05 14:30:00  62303300 182.76 180.17 181.99 181.18  180.2967
#> 382 2024-01-08 14:30:00  59144500 185.60 181.50 182.09 185.56  184.6554
#> 383 2024-01-09 14:30:00  42841800 185.15 182.73 183.92 185.14  184.2374
#> 384 2024-01-10 14:30:00  46792900 186.40 183.92 184.35 186.19  185.2823
#> 385 2024-01-11 14:30:00  49128400 187.05 183.62 186.54 185.59  184.6852
#> 386 2024-01-12 14:30:00  40444700 186.74 185.19 186.06 185.92  185.0136
#> 387 2024-01-16 14:30:00  65603000 184.26 180.93 182.16 183.63  182.7348
#> 388 2024-01-17 14:30:00  47317400 182.93 180.30 181.27 182.68  181.7894
#> 389 2024-01-18 14:30:00  78005800 189.14 185.83 186.09 188.63  187.7104
#> 390 2024-01-19 14:30:00  68741000 191.95 188.82 189.33 191.56  190.6261
#> 391 2024-01-22 14:30:00  60133900 195.33 192.26 192.30 193.89  192.9447
#> 392 2024-01-23 14:30:00  42355600 195.75 193.83 195.02 195.18  194.2285
#> 393 2024-01-24 14:30:00  53631300 196.38 194.34 195.42 194.50  193.5518
#> 394 2024-01-25 14:30:00  54822100 196.27 193.11 195.22 194.17  193.2234
#> 395 2024-01-26 14:30:00  44594000 194.76 191.94 194.27 192.42  191.4819
#> 396 2024-01-29 14:30:00  47145600 192.20 189.58 192.01 191.73  190.7953
#> 397 2024-01-30 14:30:00  55859400 191.80 187.47 190.94 188.04  187.1233
#> 398 2024-01-31 14:30:00  55467800 187.10 184.35 187.04 184.40  183.5010
#> 399 2024-02-01 14:30:00  64885400 186.95 183.82 183.99 186.86  185.9490
#> 400 2024-02-02 14:30:00 102518000 187.33 179.25 179.86 185.85  184.9440
#> 401 2024-02-05 14:30:00  69668800 189.25 185.84 188.15 187.68  186.7650
#> 402 2024-02-06 14:30:00  43490800 189.31 186.77 186.86 189.30  188.3771
#> 403 2024-02-07 14:30:00  53439000 191.05 188.61 190.64 189.41  188.4866
#> 404 2024-02-08 14:30:00  40962000 189.54 187.35 189.39 188.32  187.4019
#> 405 2024-02-09 14:30:00  45155200 189.99 188.00 188.65 188.85  188.1691
#> 406 2024-02-12 14:30:00  41781900 188.67 186.79 188.42 187.15  186.4753
#> 407 2024-02-13 14:30:00  56529500 186.21 183.51 185.77 185.04  184.3729
#> 408 2024-02-14 14:30:00  54630500 185.53 182.44 185.32 184.15  183.4861
#> 409 2024-02-15 14:30:00  65434500 184.49 181.35 183.55 183.86  183.1971
#> 410 2024-02-16 14:30:00  49701400 184.85 181.67 183.42 182.31  181.6527
#> 411 2024-02-20 14:30:00  53665600 182.43 180.00 181.79 181.56  180.9054
#> 412 2024-02-21 14:30:00  41529700 182.89 180.66 181.94 182.32  181.6627
#> 413 2024-02-22 14:30:00  52292200 184.96 182.46 183.48 184.37  183.7053
#> 414 2024-02-23 14:30:00  45119700 185.04 182.23 185.01 182.52  181.8620
#> 415 2024-02-26 14:30:00  40867400 182.76 180.65 182.24 181.16  180.5069
#> 416 2024-02-27 14:30:00  54318900 183.92 179.56 181.10 182.63  181.9716
#> 417 2024-02-28 14:30:00  48953900 183.12 180.13 182.51 181.42  180.7659
#> 418 2024-02-29 14:30:00 136682600 182.57 179.53 181.27 180.75  180.0983
#> 419 2024-03-01 14:30:00  73488000 180.53 177.38 179.55 179.66  179.0123
#> 420 2024-03-04 14:30:00  81510100 176.90 173.79 176.15 175.10  174.4687
#> 421 2024-03-05 14:30:00  95132400 172.04 169.62 170.76 170.12  169.5067
#> 422 2024-03-06 14:30:00  68587700 171.24 168.68 171.06 169.12  168.5103
#> 423 2024-03-07 14:30:00  71765100 170.73 168.49 169.15 169.00  168.3907
#> 424 2024-03-08 14:30:00  76114600 173.70 168.94 169.00 170.73  170.1145
#> 425 2024-03-11 13:30:00  60139500 174.38 172.05 172.94 172.75  172.1272
#> 426 2024-03-12 13:30:00  59825400 174.03 171.01 173.15 173.23  172.6054
#> 427 2024-03-13 13:30:00  52488700 173.19 170.76 172.77 171.13  170.5130
#> 428 2024-03-14 13:30:00  72913500 174.31 172.05 172.91 173.00  172.3763
#> 429 2024-03-15 13:30:00 121664700 172.62 170.29 171.17 172.62  171.9977
#> 430 2024-03-18 13:30:00  75604200 177.71 173.52 175.57 173.72  173.0937
#> 431 2024-03-19 13:30:00  55215200 176.61 173.03 174.34 176.08  175.4452
#> 432 2024-03-20 13:30:00  53423100 178.67 175.09 175.72 178.67  178.0258
#> 433 2024-03-21 13:30:00 106181300 177.49 170.84 177.05 171.37  170.7522
#> 434 2024-03-22 13:30:00  71106600 173.05 170.06 171.76 172.28  171.6589
#> 435 2024-03-25 13:30:00  54288300 171.94 169.45 170.57 170.85  170.2340
#> 436 2024-03-26 13:30:00  57388400 171.42 169.58 170.00 169.71  169.0981
#> 437 2024-03-27 13:30:00  60273300 173.60 170.11 170.41 173.31  172.6852
#> 438 2024-03-28 13:30:00  65672700 172.23 170.51 171.75 171.48  170.8618
#> 439 2024-04-01 13:30:00  46240500 171.25 169.48 171.19 170.03  169.4170
#> 440 2024-04-02 13:30:00  49329500 169.34 168.23 169.08 168.84  168.2313
#> 441 2024-04-03 13:30:00  47691700 170.68 168.58 168.79 169.65  169.0383
#> 442 2024-04-04 13:30:00  53704400 171.92 168.82 170.29 168.82  168.2113
#> 443 2024-04-05 13:30:00  42055200 170.39 168.95 169.59 169.58  168.9686
#> 444 2024-04-08 13:30:00  37425500 169.20 168.24 169.03 168.45  167.8427
#> 445 2024-04-09 13:30:00  42451200 170.08 168.35 168.70 169.67  169.0583
#> 446 2024-04-10 13:30:00  49709300 169.09 167.11 168.80 167.78  167.1751
#> 447 2024-04-11 13:30:00  91070300 175.46 168.16 168.34 175.04  174.4089
#> 448 2024-04-12 13:30:00 101593300 178.36 174.21 174.26 176.55  175.9135
#> 449 2024-04-15 13:30:00  73531800 176.63 172.50 175.36 172.69  172.0674
#> 450 2024-04-16 13:30:00  73711200 173.76 168.27 171.75 169.38  168.7693
#> 451 2024-04-17 13:30:00  50901200 170.65 168.00 169.61 168.00  167.3943
#> 452 2024-04-18 13:30:00  43122900 168.64 166.55 168.03 167.04  166.4378
#> 453 2024-04-19 13:30:00  67772100 166.40 164.08 166.21 165.00  164.4051
#> 454 2024-04-22 13:30:00  48116400 167.26 164.77 165.52 165.84  165.2421
#> 455 2024-04-23 13:30:00  49537800 167.05 164.92 165.35 166.90  166.2983
#> 456 2024-04-24 13:30:00  48251800 169.30 166.21 166.54 169.02  168.4106
#> 457 2024-04-25 13:30:00  50558300 170.61 168.15 169.53 169.89  169.2775
#> 458 2024-04-26 13:30:00  44838400 171.34 169.18 169.88 169.30  168.6896
#> 459 2024-04-29 13:30:00  68169400 176.03 173.10 173.37 173.50  172.8745
#> 460 2024-04-30 13:30:00  65934800 174.99 170.00 173.33 170.33  169.7159
#> 461 2024-05-01 13:30:00  50383100 172.71 169.11 169.58 169.30  168.6896
#> 462 2024-05-02 13:30:00  94214900 173.42 170.89 172.51 173.03  172.4062
#> 463 2024-05-03 13:30:00 163224100 187.00 182.66 186.65 183.38  182.7188
#> 464 2024-05-06 13:30:00  78569700 184.20 180.42 182.35 181.71  181.0549
#> 465 2024-05-07 13:30:00  77305800 184.90 181.32 183.45 182.40  181.7424
#> 466 2024-05-08 13:30:00  45057100 183.07 181.45 182.85 182.74  182.0812
#> 467 2024-05-09 13:30:00  48983000 184.66 182.11 182.56 184.57  183.9046
#> 468 2024-05-10 13:30:00  50759500 185.09 182.13 184.90 183.05  182.6374
#> 469 2024-05-13 13:30:00  72044800 187.10 184.62 185.44 186.28  185.8601
#> 470 2024-05-14 13:30:00  52393600 188.30 186.29 187.51 187.43  187.0076
#> 471 2024-05-15 13:30:00  70400000 190.65 187.37 187.91 189.72  189.2924
#> 472 2024-05-16 13:30:00  52845200 191.10 189.66 190.47 189.84  189.4121
#> 473 2024-05-17 13:30:00  41282900 190.81 189.18 189.51 189.87  189.4420
#> 474 2024-05-20 13:30:00  44361300 191.92 189.01 189.33 191.04  190.6094
#> 475 2024-05-21 13:30:00  42309400 192.73 190.92 191.09 192.35  191.9165
#> 476 2024-05-22 13:30:00  34648500 192.82 190.27 192.27 190.90  190.4697
#> 477 2024-05-23 13:30:00  51005900 191.00 186.63 190.98 186.88  186.4588
#> 478 2024-05-24 13:30:00  36294600 190.58 188.04 188.82 189.98  189.5518
#> 479 2024-05-28 13:30:00  52280100 193.00 189.10 191.51 189.99  189.5618
#> 480 2024-05-29 13:30:00  53068000 192.25 189.51 189.61 190.29  189.8611
#> 481 2024-05-30 13:30:00  49947900 192.18 190.63 190.76 191.29  190.8589
#> 482 2024-05-31 13:30:00  75158300 192.57 189.91 191.44 192.25  191.8167
#> 483 2024-06-03 13:30:00  50080500 194.99 192.52 192.90 194.03  193.5927
#> 484 2024-06-04 13:30:00  47471400 195.32 193.03 194.64 194.35  193.9120
#> 485 2024-06-05 13:30:00  54156800 196.90 194.87 195.40 195.87  195.4285
#> 486 2024-06-06 13:30:00  41181800 196.50 194.17 195.69 194.48  194.0417
#> 487 2024-06-07 13:30:00  53103900 196.94 194.14 194.65 196.89  196.4462
#> 488 2024-06-10 13:30:00  97262100 197.30 192.15 196.90 193.12  192.6847
#> 489 2024-06-11 13:30:00 172373300 207.16 193.63 193.65 207.15  206.6831
#> 490 2024-06-12 13:30:00 198134300 220.20 206.90 207.37 213.07  212.5898
#> 491 2024-06-13 13:30:00  97862700 216.75 211.60 214.74 214.24  213.7571
#> 492 2024-06-14 13:30:00  70122700 215.17 211.30 213.85 212.49  212.0111
#> 493 2024-06-17 13:30:00  93728300 218.95 212.72 213.37 216.67  216.1817
#> 494 2024-06-18 13:30:00  79943300 218.63 213.00 217.59 214.29  213.8070
#> 495 2024-06-20 13:30:00  86172500 214.24 208.85 213.93 209.68  209.2074
#> 496 2024-06-21 13:30:00 246421400 211.89 207.11 210.39 207.49  207.0224
#> 497 2024-06-24 13:30:00  80727000 212.70 206.59 207.72 208.14  207.6709
#> 498 2024-06-25 13:30:00  56713900 211.38 208.61 209.15 209.07  208.5988
#> 499 2024-06-26 13:30:00  66213200 214.86 210.64 211.50 213.25  212.7694
#> 500 2024-06-27 13:30:00  49772700 215.74 212.35 214.69 214.10  213.6175
#> 501 2024-06-28 13:30:00  82542700 216.07 210.30 215.77 210.62  210.1453
#> 502 2024-07-01 13:30:00  60402900 217.51 211.92 212.09 216.75  216.2615
#> 503 2024-07-02 13:30:00  58046200 220.38 215.10 216.15 220.27  219.7735
#> 504 2024-07-03 13:30:00  37369800 221.55 219.03 220.00 221.55  221.0507
#> 505 2024-07-05 13:30:00  60412400 226.45 221.65 221.65 226.34  225.8298
#> 506 2024-07-08 13:30:00  59085900 227.85 223.25 227.09 227.82  227.3065
#> 507 2024-07-09 13:30:00  48076100 229.40 226.37 227.93 228.68  228.1646
#> 508 2024-07-10 13:30:00  62627700 233.08 229.25 229.30 232.98  232.4549
#> 509 2024-07-11 13:30:00  64710600 232.39 225.77 231.39 227.57  227.0571
#> 510 2024-07-12 13:30:00  53046500 232.64 228.68 228.92 230.54  230.0204
#> 511 2024-07-15 13:30:00  62631300 237.23 233.09 236.48 234.40  233.8717
#> 512 2024-07-16 13:30:00  43234300 236.27 232.33 235.00 234.82  234.2907
#> 513 2024-07-17 13:30:00  57345900 231.46 226.64 229.45 228.88  228.3641
#> 514 2024-07-18 13:30:00  66034600 230.44 222.27 230.28 224.18  223.6747
#> 515 2024-07-19 13:30:00  49151500 226.80 223.28 224.82 224.31  223.8044
#> 516 2024-07-22 13:30:00  48201800 227.78 223.09 227.01 223.96  223.4552
#> 517 2024-07-23 13:30:00  39960300 226.94 222.68 224.37 225.01  224.5029
#> 518 2024-07-24 13:30:00  61777600 224.80 217.13 224.00 218.54  218.0474
#> 519 2024-07-25 13:30:00  51391200 220.85 214.62 218.93 217.49  216.9998
#> 520 2024-07-26 13:30:00  41601300 219.49 216.01 218.70 217.96  217.4688
#> 521 2024-07-29 13:30:00  36311800 219.30 215.75 216.96 218.24  217.7481
#> 522 2024-07-30 13:30:00  41643800 220.33 216.12 219.19 218.80  218.3069
#> 523 2024-07-31 13:30:00  50036300 223.82 220.63 221.44 222.08  221.5795
#> 524 2024-08-01 13:30:00  62501000 224.48 217.02 224.37 218.36  217.8678
#> 525 2024-08-02 13:30:00 105568600 225.60 217.71 219.15 219.86  219.3645
#> 526 2024-08-05 13:30:00 119548600 213.50 196.00 199.09 209.27  208.7983
#> 527 2024-08-06 13:30:00  69660500 209.99 201.07 205.30 207.23  206.7629
#> 528 2024-08-07 13:30:00  63516400 213.64 206.39 206.90 209.82  209.3471
#> 529 2024-08-08 13:30:00  47161100 214.20 208.83 213.11 213.31  212.8292
#> 530 2024-08-09 13:30:00  42201600 216.78 211.97 212.10 216.24  215.7526
#> 531 2024-08-12 13:30:00  38028100 219.51 215.60 216.07 217.53  217.2909
#> 532 2024-08-13 13:30:00  44155300 221.89 219.01 219.01 221.27  221.0268
#> 533 2024-08-14 13:30:00  41960600 223.03 219.70 220.57 221.72  221.4763
#> 534 2024-08-15 13:30:00  46414000 225.35 222.76 224.60 224.72  224.4730
#> 535 2024-08-16 13:30:00  44340200 226.83 223.65 223.92 226.05  225.8016
#> 536 2024-08-19 13:30:00  40687800 225.99 223.04 225.72 225.89  225.6418
#> 537 2024-08-20 13:30:00  30299000 227.17 225.45 225.77 226.51  226.2611
#> 538 2024-08-21 13:30:00  34765500 227.98 225.05 226.52 226.40  226.1512
#> 539 2024-08-22 13:30:00  43695300 228.34 223.90 227.79 224.53  224.2832
#> 540 2024-08-23 13:30:00  38677300 228.22 224.33 225.66 226.84  226.5907
#> 541 2024-08-26 13:30:00  30602200 227.28 223.89 226.76 227.18  226.9303
#> 542 2024-08-27 13:30:00  35934600 228.85 224.89 226.00 228.03  227.7794
#> 543 2024-08-28 13:30:00  38052200 229.86 225.68 227.92 226.49  226.2411
#> 544 2024-08-29 13:30:00  51906300 232.92 228.88 230.10 229.79  229.5375
#> 545 2024-08-30 13:30:00  52990800 230.40 227.48 230.19 229.00  228.7483
#> 546 2024-09-03 13:30:00  50190600 229.00 221.17 228.55 222.77  222.5252
#> 547 2024-09-04 13:30:00  43840200 221.78 217.48 221.66 220.85  220.6073
#> 548 2024-09-05 13:30:00  36615400 225.48 221.52 221.63 222.38  222.1356
#> 549 2024-09-06 13:30:00  48423000 225.24 219.77 223.95 220.82  220.5773
#> 550 2024-09-09 13:30:00  67180000 221.27 216.71 220.82 220.91  220.6672
#> 551 2024-09-10 13:30:00  51591000 221.48 216.73 218.92 220.11  219.8681
#> 552 2024-09-11 13:30:00  44587100 223.09 217.89 221.46 222.66  222.4153
#> 553 2024-09-12 13:30:00  37498200 223.55 219.82 222.50 222.77  222.5252
#> 554 2024-09-13 13:30:00  36766600 224.04 221.91 223.58 222.50  222.2555
#> 555 2024-09-16 13:30:00  59357400 217.22 213.92 216.54 216.32  216.0823
#> 556 2024-09-17 13:30:00  45519300 216.90 214.50 215.75 216.79  216.5517
#> 557 2024-09-18 13:30:00  59894900 222.71 217.54 217.55 220.69  220.4475
#> 558 2024-09-19 13:30:00  66781300 229.82 224.63 224.99 228.87  228.6185
#> 559 2024-09-20 13:30:00 318679900 233.09 227.62 229.97 228.20  227.9492
#> 560 2024-09-23 13:30:00  54146000 229.45 225.81 227.34 226.47  226.2211
#> 561 2024-09-24 13:30:00  43556100 229.35 225.73 228.65 227.37  227.1201
#> 562 2024-09-25 13:30:00  42308700 227.29 224.02 224.93 226.37  226.1212
#> 563 2024-09-26 13:30:00  36636700 228.50 225.41 227.30 227.52  227.2700
#> 564 2024-09-27 13:30:00  34026000 229.52 227.30 228.46 227.79  227.5397
#> 565 2024-09-30 13:30:00  54541900 233.00 229.65 230.04 233.00  232.7439
#> 566 2024-10-01 13:30:00  63285000 229.65 223.74 229.52 226.21  225.9614
#> 567 2024-10-02 13:30:00  32880600 227.37 223.02 225.89 226.78  226.5308
#> 568 2024-10-03 13:30:00  34044200 226.81 223.32 225.14 225.67  225.4220
#> 569 2024-10-04 13:30:00  37245100 228.00 224.13 227.90 226.80  226.5508
#> 570 2024-10-07 13:30:00  39505400 225.69 221.33 224.50 221.69  221.4464
#> 571 2024-10-08 13:30:00  31855700 225.98 223.25 224.30 225.77  225.5219
#> 572 2024-10-09 13:30:00  33591100 229.75 224.83 225.23 229.54  229.2877
#> 573 2024-10-10 13:30:00  28183500 229.50 227.17 227.78 229.04  228.7883
#> 574 2024-10-11 13:30:00  31759200 229.41 227.34 229.30 227.55  227.2999
#> 575 2024-10-14 13:30:00  39882100 231.73 228.60 228.70 231.30  231.0458
#> 576 2024-10-15 13:30:00  64751400 237.49 232.37 233.61 233.85  233.5930
#> 577 2024-10-16 13:30:00  34082200 232.12 229.84 231.60 231.78  231.5253
#> 578 2024-10-17 13:30:00  32993800 233.85 230.52 233.43 232.15  231.8949
#> 579 2024-10-18 13:30:00  46431500 236.18 234.01 236.18 235.00  234.7417
#> 580 2024-10-21 13:30:00  36254500 236.85 234.45 234.45 236.48  236.2201
#> 581 2024-10-22 13:30:00  38846600 236.22 232.60 233.89 235.86  235.6008
#> 582 2024-10-23 13:30:00  52287000 235.14 227.76 234.08 230.76  230.5064
#> 583 2024-10-24 13:30:00  31109500 230.82 228.41 229.98 230.57  230.3166
#> 584 2024-10-25 13:30:00  38802300 233.22 229.57 229.74 231.41  231.1557
#> 585 2024-10-28 13:30:00  36087100 234.73 232.55 233.32 233.40  233.1435
#> 586 2024-10-29 13:30:00  35417200 234.33 232.32 233.10 233.67  233.4132
#> 587 2024-10-30 13:30:00  47070900 233.47 229.55 232.61 230.10  229.8471
#> 588 2024-10-31 13:30:00  64370100 229.83 225.37 229.34 225.91  225.6617
#> 589 2024-11-01 13:30:00  65276700 225.35 220.27 220.97 222.91  222.6650
#> 590 2024-11-04 14:30:00  44944500 222.79 219.71 220.99 222.01  221.7660
#> 591 2024-11-05 14:30:00  28111300 223.95 221.14 221.80 223.45  223.2044
#> 592 2024-11-06 14:30:00  54561100 226.07 221.19 222.61 222.72  222.4752
#> 593 2024-11-07 14:30:00  42137700 227.88 224.57 224.63 227.48  227.2300
#> 594 2024-11-08 14:30:00  38328800 228.66 226.41 227.17 226.96  226.9600
#> 595 2024-11-11 14:30:00  42005600 225.70 221.50 225.00 224.23  224.2300
#> 596 2024-11-12 14:30:00  40398300 225.59 223.36 224.55 224.23  224.2300
#> 597 2024-11-13 14:30:00  48566200 226.65 222.76 224.01 225.12  225.1200
#> 598 2024-11-14 14:30:00  44923900 228.87 225.00 225.02 228.22  228.2200
#> 599 2024-11-15 14:30:00  47923700 226.92 224.27 226.40 225.00  225.0000
#> 600 2024-11-18 14:30:00  44686000 229.74 225.17 225.25 228.02  228.0200
#> 601 2024-11-19 14:30:00  36211800 230.16 226.66 226.98 228.28  228.2800
#> 602 2024-11-20 14:30:00  35169600 229.93 225.89 228.06 229.00  229.0000
#> 603 2024-11-21 14:30:00  42108300 230.16 225.71 228.88 228.52  228.5200
#> 604 2024-11-22 14:30:00  38168300 230.72 228.06 228.06 229.87  229.8700
#> 605 2024-11-25 14:30:00  90152800 233.25 229.74 231.46 232.87  232.8700
#> 606 2024-11-26 14:30:00  45986200 235.57 233.33 233.33 235.06  235.0600
#> 607 2024-11-27 14:30:00  33498400 235.69 233.81 234.47 234.93  234.9300
#> 608 2024-11-29 14:30:00  28481400 237.81 233.97 234.81 237.33  237.3300
#> 609 2024-12-02 14:30:00  48137100 240.79 237.16 237.27 239.59  239.5900
#> 610 2024-12-03 14:30:00  38861000 242.76 238.90 239.81 242.65  242.6500
#> 611 2024-12-04 14:30:00  44383900 244.11 241.25 242.87 243.01  243.0100
#> 612 2024-12-05 14:30:00  40033900 244.54 242.13 243.99 243.04  243.0400
#> 613 2024-12-06 14:30:00  36870600 244.63 242.08 242.91 242.84  242.8400
#> 614 2024-12-09 14:30:00  44649200 247.24 241.75 241.83 246.75  246.7500
#> 615 2024-12-10 14:30:00  36914800 248.21 245.34 246.89 247.77  247.7700
#> 616 2024-12-11 14:30:00  45205800 250.80 246.26 247.96 246.49  246.4900
#> 617 2024-12-12 14:30:00  32777500 248.74 245.68 246.89 247.96  247.9600
#> 618 2024-12-13 14:30:00  33155300 249.29 246.24 247.82 248.13  248.1300
#> 619 2024-12-16 14:30:00  51694800 251.38 247.65 247.99 251.04  251.0400
#> 620 2024-12-17 14:30:00  51356400 253.83 249.78 250.08 253.48  253.4800
#> 621 2024-12-18 14:30:00  56774100 254.28 247.74 252.16 248.05  248.0500
#> 622 2024-12-19 14:30:00  60882300 252.00 247.09 247.50 249.79  249.7900
#> 623 2024-12-20 14:30:00 147495300 255.00 245.69 248.04 254.49  254.4900
#> 624 2024-12-23 14:30:00  40858800 255.65 253.45 254.77 255.27  255.2700
#> 625 2024-12-24 14:30:00  23234700 258.21 255.29 255.49 258.20  258.2000
#> 626 2024-12-26 14:30:00  27237100 260.10 257.63 258.19 259.02  259.0200
#> 627 2024-12-27 14:30:00  42355300 258.70 253.06 257.83 255.59  255.5900
#> 628 2024-12-30 14:30:00  35557500 253.50 250.75 252.23 252.20  252.2000
#> 629 2024-12-31 14:30:00  39480700 253.28 249.43 252.44 250.42  250.4200
#> 630 2025-01-02 14:30:00  55740700 249.10 241.82 248.93 243.85  243.8500
#> 631 2025-01-03 14:30:00  40244100 244.18 241.89 243.36 243.36  243.3600
#> 632 2025-01-06 14:30:00  45045600 247.33 243.20 244.31 245.00  245.0000
#> 633 2025-01-07 14:30:00  40856000 245.55 241.35 242.98 242.21  242.2100
#> 634 2025-01-08 14:30:00  37628900 243.71 240.05 241.92 242.70  242.7000
#> 635 2025-01-10 14:30:00  61710900 240.16 233.00 240.01 236.85  236.8500
#> 636 2025-01-13 14:30:00  49630700 234.67 229.72 233.53 234.40  234.4000
#> 637 2025-01-14 14:30:00  39435300 236.12 232.47 234.75 233.28  233.2800
#> 638 2025-01-15 14:30:00  39832000 238.96 234.43 234.64 237.87  237.8700
#> 639 2025-01-16 14:30:00  71759100 238.01 228.03 237.35 228.26  228.2600
#> 640 2025-01-17 14:30:00  68488300 232.29 228.48 232.12 229.98  229.9800
#> 641 2025-01-21 14:30:00  98070400 224.42 219.38 224.00 222.64  222.6400
#> 642 2025-01-22 14:30:00  64126500 224.12 219.79 219.79 223.83  223.8300
#> 643 2025-01-23 14:30:00  60234800 227.03 222.30 224.74 223.66  223.6600
#> 644 2025-01-24 14:30:00  54619500 225.63 221.41 224.78 222.78  222.7800
aapl$get_history(start = '2022-07-01', end = '2022-07-14', interval = '1d')
#>                  date   volume   high    low   open  close adj_close
#> 1 2022-07-01 13:30:00 71051600 139.04 135.66 136.04 138.93  137.0661
#> 2 2022-07-05 13:30:00 73353800 141.61 136.93 137.77 141.56  139.6609
#> 3 2022-07-06 13:30:00 74064300 144.12 141.08 141.35 142.92  141.0026
#> 4 2022-07-07 13:30:00 66253700 146.55 143.28 143.29 146.35  144.3866
#> 5 2022-07-08 13:30:00 64547800 147.55 145.00 145.26 147.04  145.0673
#> 6 2022-07-11 13:30:00 63141600 146.64 143.78 145.67 144.87  142.9265
#> 7 2022-07-12 13:30:00 77588800 148.45 145.05 145.76 145.86  143.9032
#> 8 2022-07-13 13:30:00 71185600 146.45 142.12 142.99 145.49  143.5381
aapl$get_history(period = '1mo', interval = '1d')
#>                   date   volume   high    low   open  close adj_close
#> 1  2024-12-26 14:30:00 27237100 260.10 257.63 258.19 259.02    259.02
#> 2  2024-12-27 14:30:00 42355300 258.70 253.06 257.83 255.59    255.59
#> 3  2024-12-30 14:30:00 35557500 253.50 250.75 252.23 252.20    252.20
#> 4  2024-12-31 14:30:00 39480700 253.28 249.43 252.44 250.42    250.42
#> 5  2025-01-02 14:30:00 55740700 249.10 241.82 248.93 243.85    243.85
#> 6  2025-01-03 14:30:00 40244100 244.18 241.89 243.36 243.36    243.36
#> 7  2025-01-06 14:30:00 45045600 247.33 243.20 244.31 245.00    245.00
#> 8  2025-01-07 14:30:00 40856000 245.55 241.35 242.98 242.21    242.21
#> 9  2025-01-08 14:30:00 37628900 243.71 240.05 241.92 242.70    242.70
#> 10 2025-01-10 14:30:00 61710900 240.16 233.00 240.01 236.85    236.85
#> 11 2025-01-13 14:30:00 49630700 234.67 229.72 233.53 234.40    234.40
#> 12 2025-01-14 14:30:00 39435300 236.12 232.47 234.75 233.28    233.28
#> 13 2025-01-15 14:30:00 39832000 238.96 234.43 234.64 237.87    237.87
#> 14 2025-01-16 14:30:00 71759100 238.01 228.03 237.35 228.26    228.26
#> 15 2025-01-17 14:30:00 68488300 232.29 228.48 232.12 229.98    229.98
#> 16 2025-01-21 14:30:00 98070400 224.42 219.38 224.00 222.64    222.64
#> 17 2025-01-22 14:30:00 64126500 224.12 219.79 219.79 223.83    223.83
#> 18 2025-01-23 14:30:00 60234800 227.03 222.30 224.74 223.66    223.66
#> 19 2025-01-24 14:30:00 54619500 225.63 221.41 224.78 222.78    222.78
# }