R6 Class Representing a Ticker

Base class for getting all data related to ticker from Yahoo Finance API

Format

An R6 class object

Public fields

symbol

Symbol for which data is retrieved

Active bindings

asset_profile

Information related to the company's location, operations, and officers.

calendar_events

Earnings and Revenue expectations for upcoming earnings date.

company_officers

Retrieves top executives for given symbol and their total pay package.

earnings_history

Data related to historical earnings (actual vs. estimate)

earnings

Historical earnings data.

earnings_trend

Historical trend data for earnings and revenue estimations

esg_scores

Data related to environmental, social, and governance metrics

financial_data

Financial key performance indicators

fund_bond_holdings

Retrieves aggregated maturity and duration information for a given symbol

fund_bond_ratings

Retrieves aggregated maturity and duration information

fund_equity_holdings

Fund equity holdings

fund_holding_info

Contains information for a funds top holdings, bond ratings, bond holdings, equity holdings, sector weightings, and category breakdown

fund_ownership

Top 10 owners of a given symbol

fund_performance

Historical return data for a given symbol and its specific category

fund_profile

Summary level information for a given symbol

fund_section_weightings

Retrieves aggregated sector weightings for a given symbol

fund_top_holdings

Retrieves Top 10 holdings for a given symbol

fund_holdings

Holding info for the given fund

grading_history

Data related to upgrades / downgrades by companies

index_trend

Trend data related to given symbol's index, specificially PE and PEG ratios

inside_holders

Data related to stock holdings of a given symbol(s) insiders

insider_transactions

Transactions by insiders for a given symbol(s)

institution_ownership

Top 10 owners of a given symbol

key_stats

KPIs for given symbol

major_holders

Data showing breakdown of owners of given symbol(s), insiders, institutions, etc.

page_views

Short, Mid, and Long-term trend data regarding a symbol's page views

price

Detailed pricing data for given symbol, exchange, quote type, currency, market cap, pre / post market data, etc.

quote_type

Stock exchange specific data for given symbol

recommendation_trend

Data related to historical recommendations (buy, hold, sell) for a given symbol

security_filings

Historical SEC filings

share_purchase_activity

High-level buy / sell data

summary_detail

Contains information available via the Summary tab in Yahoo Finance

summary_profile

Return business summary of given symbol

valuation_measures

Retrieves valuation measures for most recent four quarters

option_chain

Option chain data for all expiration dates for a given symbol

option_expiration_dates

Option expiration dates

option_strikes

Option strikes

quote

Get real-time quote information for given symbol

recommendations

Recommended symbols

technical_insights

Technical indicators for given symbol

Methods

Public methods

Ticker$new()

Ticker$set_symbol()

Ticker$get_balance_sheet()

Ticker$get_cash_flow()

Ticker$get_income_statement()

Ticker$get_history()

Ticker$clone()

Method new()

Create a new Ticker object

Usage

Ticker$new(symbol = NA)

Arguments

symbol

Symbol

Returns

A new `Ticker` object

Examples

aapl <- Ticker$new('aapl')

Method set_symbol()

Set a new symbol.

Usage

Ticker$set_symbol(symbol)

Arguments

symbol

New symbol

Examples

aapl <- Ticker$new('aapl')
aapl$set_symbol('msft')

Method get_balance_sheet()

Retrieves balance sheet data for most recent four quarters or most recent four years

Usage

Ticker$get_balance_sheet(
  frequency = c("annual", "quarter"),
  clean_names = TRUE
)

Arguments

frequency

Annual or quarter.

clean_names

Logical; if TRUE, converts column names to snake case.

Examples

aapl <- Ticker$new('aapl')
aapl$get_balance_sheet('annual')
aapl$get_balance_sheet('quarter')

Method get_cash_flow()

Retrieves cash flow data for most recent four quarters or most recent four years

Usage

Ticker$get_cash_flow(frequency = c("annual", "quarter"), clean_names = TRUE)

Arguments

frequency

Annual or quarter.

clean_names

Logical; if TRUE, converts column names to snake case.

Examples

aapl <- Ticker$new('aapl')
aapl$get_cash_flow('annual')
aapl$get_cash_flow('quarter')

Method get_income_statement()

Retrieves income statement data for most recent four quarters or most recent four years

Usage

Ticker$get_income_statement(
  frequency = c("annual", "quarter"),
  clean_names = TRUE
)

Arguments

frequency

Annual or quarter.

clean_names

Logical; if TRUE, converts column names to snake case.

Examples

aapl <- Ticker$new('aapl')
aapl$get_income_statement('annual')
aapl$get_income_statement('quarter')

Method get_history()

Retrieves historical pricing data

Usage

Ticker$get_history(period = "ytd", interval = "1d", start = NULL, end = NULL)

Arguments

period

Length of time

interval

Time between data points

start

Specific starting date

end

Specific ending date

Examples

aapl <- Ticker$new('aapl')
aapl$get_history()
aapl$get_history(start = '2022-07-01', interval = '1d')
aapl$get_history(start = '2022-07-01', end = '2022-07-14', interval = '1d')
aapl$get_history(period = '1mo', interval = '1d')

Method clone()

The objects of this class are cloneable with this method.

Usage

Ticker$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `Ticker$new`
## ------------------------------------------------

aapl <- Ticker$new('aapl')

## ------------------------------------------------
## Method `Ticker$set_symbol`
## ------------------------------------------------

aapl <- Ticker$new('aapl')
aapl$set_symbol('msft')

## ------------------------------------------------
## Method `Ticker$get_balance_sheet`
## ------------------------------------------------

aapl <- Ticker$new('aapl')
aapl$get_balance_sheet('annual')
#> # A tibble: 4 x 24
#>   end_date          cash short~1 net_r~2 inven~3 other~4 total~5 long_~6 prope~7
#>   <date>           <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#> 1 2021-09-25 34940000000 2.77e10 5.15e10  6.58e9 1.41e10 1.35e11 1.28e11 4.95e10
#> 2 2020-09-26 38016000000 5.29e10 3.74e10  4.06e9 1.13e10 1.44e11 1.01e11 4.53e10
#> 3 2019-09-28 48844000000 5.17e10 4.58e10  4.11e9 1.24e10 1.63e11 1.05e11 3.74e10
#> 4 2018-09-29 25913000000 4.04e10 4.90e10  3.96e9 1.21e10 1.31e11 1.71e11 4.13e10
#> # ... with 15 more variables: other_assets <dbl>, total_assets <dbl>,
#> #   accounts_payable <dbl>, short_long_term_debt <dbl>,
#> #   other_current_liab <dbl>, long_term_debt <dbl>, other_liab <dbl>,
#> #   total_current_liabilities <dbl>, total_liab <dbl>, common_stock <dbl>,
#> #   retained_earnings <dbl>, treasury_stock <dbl>,
#> #   other_stockholder_equity <dbl>, total_stockholder_equity <dbl>,
#> #   net_tangible_assets <dbl>, and abbreviated variable names ...
aapl$get_balance_sheet('quarter')
#> # A tibble: 4 x 24
#>   end_date          cash short~1 net_r~2 inven~3 other~4 total~5 long_~6 prope~7
#>   <date>           <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#> 1 2022-06-25 27502000000 2.07e10 4.22e10  5.43e9 1.64e10 1.12e11 1.31e11 4.03e10
#> 2 2022-03-26 28098000000 2.34e10 4.54e10  5.46e9 1.58e10 1.18e11 1.41e11 3.93e10
#> 3 2021-12-25 37119000000 2.68e10 6.53e10  5.88e9 1.81e10 1.53e11 1.39e11 3.92e10
#> 4 2021-09-25 34940000000 2.77e10 5.15e10  6.58e9 1.41e10 1.35e11 1.28e11 4.95e10
#> # ... with 15 more variables: other_assets <dbl>, total_assets <dbl>,
#> #   accounts_payable <dbl>, short_long_term_debt <dbl>,
#> #   other_current_liab <dbl>, long_term_debt <dbl>, other_liab <dbl>,
#> #   total_current_liabilities <dbl>, total_liab <dbl>, common_stock <dbl>,
#> #   retained_earnings <dbl>, treasury_stock <dbl>,
#> #   other_stockholder_equity <dbl>, total_stockholder_equity <dbl>,
#> #   net_tangible_assets <dbl>, and abbreviated variable names ...

## ------------------------------------------------
## Method `Ticker$get_cash_flow`
## ------------------------------------------------

aapl <- Ticker$new('aapl')
aapl$get_cash_flow('annual')
#> # A tibble: 4 x 20
#>   end_date   net_i~1 depre~2 change~3 change~4 change~5 chang~6 change~7 total~8
#>   <date>       <dbl>   <dbl>    <dbl>    <dbl>    <dbl>   <dbl>    <dbl>   <dbl>
#> 1 2021-09-25 9.47e10 1.13e10  2.98e 9 -1.01e10  1.40e10 -2.64e9 -6.15e 9 1.04e11
#> 2 2020-09-26 5.74e10 1.11e10  6.52e 9  6.92e 9 -1.98e 9 -1.27e8  8.81e 8 8.07e10
#> 3 2019-09-28 5.53e10 1.25e10  5.08e 9  2.45e 8 -2.55e 9 -2.89e8 -8.96e 8 6.94e10
#> 4 2018-09-29 5.95e10 1.09e10 -2.77e10 -5.32e 9  9.17e 9  8.28e8  3.00e10 7.74e10
#> # ... with 11 more variables: capital_expenditures <dbl>, investments <dbl>,
#> #   other_cashflows_from_investing_activities <int>,
#> #   total_cashflows_from_investing_activities <dbl>, dividends_paid <dbl>,
#> #   net_borrowings <dbl>, other_cashflows_from_financing_activities <int>,
#> #   total_cash_from_financing_activities <dbl>, change_in_cash <dbl>,
#> #   repurchase_of_stock <dbl>, issuance_of_stock <int>, and abbreviated
#> #   variable names 1: net_income, 2: depreciation, 3: change_to_netincome, ...
aapl$get_cash_flow('quarter')
#> # A tibble: 4 x 20
#>   end_date   net_inc~1 depre~2 chang~3 chang~4 change~5 chang~6 change~7 total~8
#>   <date>         <dbl>   <dbl>   <dbl>   <dbl>    <dbl>   <int>    <dbl>   <dbl>
#> 1 2022-06-25   1.94e10  2.80e9  3.87e9 -9.81e8 -4.72e 9 -1.6 e7  2.50e 9 2.29e10
#> 2 2022-03-26   2.50e10  2.74e9  2.47e9  9.48e9 -2.14e10  3.84e8  9.49e 9 2.82e10
#> 3 2021-12-25   3.46e10  2.70e9  3.11e9 -3.93e9  2.03e10  6.81e8 -1.05e10 4.70e10
#> 4 2021-09-25   2.06e10  2.99e9 -1.55e9 -8.81e9  1.40e10 -1.43e9 -5.60e 9 2.02e10
#> # ... with 11 more variables: capital_expenditures <dbl>, investments <dbl>,
#> #   other_cashflows_from_investing_activities <int>,
#> #   total_cashflows_from_investing_activities <dbl>, dividends_paid <dbl>,
#> #   net_borrowings <dbl>, other_cashflows_from_financing_activities <int>,
#> #   total_cash_from_financing_activities <dbl>, change_in_cash <dbl>,
#> #   repurchase_of_stock <dbl>, issuance_of_stock <int>, and abbreviated
#> #   variable names 1: net_income, 2: depreciation, 3: change_to_netincome, ...

## ------------------------------------------------
## Method `Ticker$get_income_statement`
## ------------------------------------------------

aapl <- Ticker$new('aapl')
aapl$get_income_statement('annual')
#> # A tibble: 4 x 16
#>   end_date   total_rev~1 cost_~2 gross~3 resea~4 selli~5 total~6 opera~7 total~8
#>   <date>           <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <int>
#> 1 2021-09-25     3.66e11 2.13e11 1.53e11 2.19e10 2.20e10 2.57e11 1.09e11  2.58e8
#> 2 2020-09-26     2.75e11 1.70e11 1.05e11 1.88e10 1.99e10 2.08e11 6.63e10  8.03e8
#> 3 2019-09-28     2.60e11 1.62e11 9.84e10 1.62e10 1.82e10 1.96e11 6.39e10  1.81e9
#> 4 2018-09-29     2.66e11 1.64e11 1.02e11 1.42e10 1.67e10 1.95e11 7.09e10  2.00e9
#> # ... with 7 more variables: ebit <dbl>, interest_expense <dbl>,
#> #   income_before_tax <dbl>, income_tax_expense <dbl>,
#> #   net_income_from_continuing_ops <dbl>, net_income <dbl>,
#> #   net_income_applicable_to_common_shares <dbl>, and abbreviated variable
#> #   names 1: total_revenue, 2: cost_of_revenue, 3: gross_profit,
#> #   4: research_development, 5: selling_general_administrative,
#> #   6: total_operating_expenses, 7: operating_income, ...
aapl$get_income_statement('quarter')
#> # A tibble: 4 x 16
#>   end_date   total_rev~1 cost_~2 gross~3 resea~4 selli~5 total~6 opera~7 total~8
#>   <date>           <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <int>
#> 1 2022-06-25     8.30e10 4.71e10 3.59e10  6.80e9  6.01e9 5.99e10 2.31e10 -1   e7
#> 2 2022-03-26     9.73e10 5.47e10 4.26e10  6.39e9  6.19e9 6.73e10 3.00e10  1.6 e8
#> 3 2021-12-25     1.24e11 6.97e10 5.42e10  6.31e9  6.45e9 8.25e10 4.15e10 -2.47e8
#> 4 2021-09-25     8.34e10 4.82e10 3.52e10  5.77e9  5.62e9 5.96e10 2.38e10 -5.38e8
#> # ... with 7 more variables: ebit <dbl>, interest_expense <int>,
#> #   income_before_tax <dbl>, income_tax_expense <dbl>,
#> #   net_income_from_continuing_ops <dbl>, net_income <dbl>,
#> #   net_income_applicable_to_common_shares <dbl>, and abbreviated variable
#> #   names 1: total_revenue, 2: cost_of_revenue, 3: gross_profit,
#> #   4: research_development, 5: selling_general_administrative,
#> #   6: total_operating_expenses, 7: operating_income, ...

## ------------------------------------------------
## Method `Ticker$get_history`
## ------------------------------------------------

aapl <- Ticker$new('aapl')
aapl$get_history()
#>                    date    volume   high    low   open  close adj_close
#> 1   2022-01-03 14:30:00 104487900 182.88 177.71 177.83 182.01  181.2599
#> 2   2022-01-04 14:30:00  99310400 182.94 179.12 182.63 179.70  178.9594
#> 3   2022-01-05 14:30:00  94537600 180.17 174.64 179.61 174.92  174.1992
#> 4   2022-01-06 14:30:00  96904000 175.30 171.64 172.70 172.00  171.2912
#> 5   2022-01-07 14:30:00  86709100 174.14 171.03 172.89 172.17  171.4605
#> 6   2022-01-10 14:30:00 106765600 172.50 168.17 169.08 172.19  171.4804
#> 7   2022-01-11 14:30:00  76138300 175.18 170.82 172.32 175.08  174.3585
#> 8   2022-01-12 14:30:00  74805200 177.18 174.82 176.12 175.53  174.8066
#> 9   2022-01-13 14:30:00  84505800 176.62 171.79 175.78 172.19  171.4804
#> 10  2022-01-14 14:30:00  80440800 173.78 171.09 171.34 173.07  172.3568
#> 11  2022-01-18 14:30:00  90956700 172.54 169.41 171.51 169.80  169.1003
#> 12  2022-01-19 14:30:00  94815000 171.08 165.94 170.00 166.23  165.5450
#> 13  2022-01-20 14:30:00  91420500 169.68 164.18 166.98 164.51  163.8321
#> 14  2022-01-21 14:30:00 122848900 166.33 162.30 164.42 162.41  161.7407
#> 15  2022-01-24 14:30:00 162294600 162.30 154.70 160.02 161.62  160.9539
#> 16  2022-01-25 14:30:00 115798400 162.76 157.02 158.98 159.78  159.1215
#> 17  2022-01-26 14:30:00 108275300 164.39 157.82 163.50 159.69  159.0319
#> 18  2022-01-27 14:30:00 121954600 163.84 158.28 162.45 159.22  158.5638
#> 19  2022-01-28 14:30:00 179935700 170.35 162.80 165.71 170.33  169.6281
#> 20  2022-01-31 14:30:00 115541600 175.00 169.51 170.16 174.78  174.0597
#> 21  2022-02-01 14:30:00  86213900 174.84 172.31 174.01 174.61  173.8904
#> 22  2022-02-02 14:30:00  84914300 175.88 173.33 174.75 175.84  175.1153
#> 23  2022-02-03 14:30:00  89418100 176.24 172.12 174.48 172.90  172.1875
#> 24  2022-02-04 14:30:00  82465400 174.10 170.68 171.68 172.39  171.8983
#> 25  2022-02-07 14:30:00  77251200 173.95 170.95 172.86 171.66  171.1704
#> 26  2022-02-08 14:30:00  74829200 175.35 171.43 171.73 174.83  174.3313
#> 27  2022-02-09 14:30:00  71285000 176.65 174.90 176.05 176.28  175.7772
#> 28  2022-02-10 14:30:00  90865900 175.48 171.55 174.14 172.12  171.6291
#> 29  2022-02-11 14:30:00  98670700 173.08 168.04 172.33 168.64  168.1590
#> 30  2022-02-14 14:30:00  86185500 169.58 166.56 167.37 168.88  168.3983
#> 31  2022-02-15 14:30:00  62527400 172.95 170.25 170.97 172.79  172.2972
#> 32  2022-02-16 14:30:00  61177400 173.34 170.05 171.85 172.55  172.0579
#> 33  2022-02-17 14:30:00  69589300 171.91 168.47 171.03 168.88  168.3983
#> 34  2022-02-18 14:30:00  82772700 170.54 166.19 169.82 167.30  166.8228
#> 35  2022-02-22 14:30:00  91162800 166.69 162.15 164.98 164.32  163.8513
#> 36  2022-02-23 14:30:00  90009200 166.15 159.75 165.54 160.07  159.6135
#> 37  2022-02-24 14:30:00 141147500 162.85 152.00 152.58 162.74  162.2758
#> 38  2022-02-25 14:30:00  91974200 165.12 160.87 163.84 164.85  164.3798
#> 39  2022-02-28 14:30:00  95056600 165.42 162.43 163.06 165.12  164.6490
#> 40  2022-03-01 14:30:00  83474400 166.60 161.97 164.70 163.20  162.7345
#> 41  2022-03-02 14:30:00  79724800 167.36 162.95 164.39 166.56  166.0849
#> 42  2022-03-03 14:30:00  76678400 168.91 165.55 168.47 166.23  165.7559
#> 43  2022-03-04 14:30:00  83737200 165.55 162.10 164.49 163.17  162.7046
#> 44  2022-03-07 14:30:00  96418800 165.02 159.04 163.36 159.30  158.8457
#> 45  2022-03-08 14:30:00 131148300 162.88 155.80 158.82 157.44  156.9910
#> 46  2022-03-09 14:30:00  91454900 163.41 159.41 161.48 162.95  162.4852
#> 47  2022-03-10 14:30:00 105342000 160.39 155.98 160.20 158.52  158.0679
#> 48  2022-03-11 14:30:00  96970100 159.28 154.50 158.93 154.73  154.2887
#> 49  2022-03-14 13:30:00 108732100 154.12 150.10 151.45 150.62  150.1904
#> 50  2022-03-15 13:30:00  92964300 155.57 150.38 150.90 155.09  154.6476
#> 51  2022-03-16 13:30:00 102300200 160.00 154.46 157.05 159.59  159.1348
#> 52  2022-03-17 13:30:00  75615400 161.00 157.63 158.61 160.62  160.1619
#> 53  2022-03-18 13:30:00 123511700 164.48 159.76 160.51 163.98  163.5123
#> 54  2022-03-21 13:30:00  95811400 166.35 163.01 163.51 165.38  164.9083
#> 55  2022-03-22 13:30:00  81532000 169.42 164.91 165.51 168.82  168.3385
#> 56  2022-03-23 13:30:00  98062700 172.64 167.65 167.99 170.21  169.7245
#> 57  2022-03-24 13:30:00  90131400 174.14 170.21 171.06 174.07  173.5735
#> 58  2022-03-25 13:30:00  80546200 175.28 172.75 173.88 174.72  174.2217
#> 59  2022-03-28 13:30:00  90371900 175.73 172.00 172.17 175.60  175.0992
#> 60  2022-03-29 13:30:00 100589400 179.01 176.34 176.69 178.96  178.4496
#> 61  2022-03-30 13:30:00  92633200 179.61 176.70 178.55 177.77  177.2630
#> 62  2022-03-31 13:30:00 103049300 178.03 174.40 177.84 174.61  174.1120
#> 63  2022-04-01 13:30:00  78751300 174.88 171.94 174.03 174.31  173.8128
#> 64  2022-04-04 13:30:00  76468400 178.49 174.44 174.57 178.44  177.9310
#> 65  2022-04-05 13:30:00  73401800 178.30 174.42 177.50 175.06  174.5607
#> 66  2022-04-06 13:30:00  89058800 173.63 170.13 172.36 171.83  171.3399
#> 67  2022-04-07 13:30:00  77594700 173.36 169.85 171.16 172.14  171.6490
#> 68  2022-04-08 13:30:00  76575500 171.78 169.20 171.78 170.09  169.6049
#> 69  2022-04-11 13:30:00  72246700 169.03 165.50 168.71 165.75  165.2772
#> 70  2022-04-12 13:30:00  79265200 169.87 166.64 168.02 167.66  167.1818
#> 71  2022-04-13 13:30:00  70618900 171.04 166.77 167.39 170.40  169.9140
#> 72  2022-04-14 13:30:00  75329400 171.27 165.04 170.62 165.29  164.8186
#> 73  2022-04-18 13:30:00  69023900 166.60 163.57 163.92 165.07  164.5992
#> 74  2022-04-19 13:30:00  67723800 167.82 163.91 165.02 167.40  166.9225
#> 75  2022-04-20 13:30:00  67929800 168.88 166.10 168.76 167.23  166.7530
#> 76  2022-04-21 13:30:00  87227800 171.53 165.91 168.91 166.42  165.9453
#> 77  2022-04-22 13:30:00  84882400 167.87 161.50 166.46 161.79  161.3285
#> 78  2022-04-25 13:30:00  96046400 163.17 158.46 161.12 162.88  162.4154
#> 79  2022-04-26 13:30:00  95623200 162.34 156.72 162.25 156.80  156.3528
#> 80  2022-04-27 13:30:00  88063200 159.79 155.38 155.91 156.57  156.1234
#> 81  2022-04-28 13:30:00 130216800 164.52 158.93 159.25 163.64  163.1733
#> 82  2022-04-29 13:30:00 131747600 166.20 157.25 161.84 157.65  157.2003
#> 83  2022-05-02 13:30:00 123055300 158.23 153.27 156.71 157.96  157.5095
#> 84  2022-05-03 13:30:00  88966500 160.71 156.32 158.15 159.48  159.0251
#> 85  2022-05-04 13:30:00 108256500 166.48 159.26 159.67 166.02  165.5465
#> 86  2022-05-05 13:30:00 130525300 164.08 154.95 163.85 156.77  156.3229
#> 87  2022-05-06 13:30:00 116124600 159.44 154.18 156.01 157.28  157.0618
#> 88  2022-05-09 13:30:00 131577900 155.83 151.49 154.93 152.06  151.8491
#> 89  2022-05-10 13:30:00 115366700 156.74 152.93 155.52 154.51  154.2957
#> 90  2022-05-11 13:30:00 142689800 155.45 145.81 153.50 146.50  146.2968
#> 91  2022-05-12 13:30:00 182602000 146.20 138.80 142.77 142.56  142.3622
#> 92  2022-05-13 13:30:00 113990900 148.10 143.11 144.59 147.11  146.9059
#> 93  2022-05-16 13:30:00  86643800 147.52 144.18 145.55 145.54  145.3381
#> 94  2022-05-17 13:30:00  78336300 149.77 146.68 148.86 149.24  149.0330
#> 95  2022-05-18 13:30:00 109742900 147.36 139.90 146.85 140.82  140.6247
#> 96  2022-05-19 13:30:00 136095600 141.66 136.60 139.88 137.35  137.1595
#> 97  2022-05-20 13:30:00 137426100 140.70 132.61 139.09 137.59  137.3991
#> 98  2022-05-23 13:30:00 117726300 143.26 137.65 137.79 143.11  142.9115
#> 99  2022-05-24 13:30:00 104132700 141.97 137.33 140.81 140.36  140.1653
#> 100 2022-05-25 13:30:00  92482700 141.79 138.34 138.43 140.52  140.3251
#> 101 2022-05-26 13:30:00  90601500 144.34 137.14 137.39 143.78  143.5806
#> 102 2022-05-27 13:30:00  90978500 149.68 145.26 145.39 149.64  149.4324
#> 103 2022-05-31 13:30:00 103718400 150.66 146.84 149.07 148.84  148.6335
#> 104 2022-06-01 13:30:00  74286600 151.74 147.68 149.90 148.71  148.5037
#> 105 2022-06-02 13:30:00  72348100 151.27 146.86 147.83 151.21  151.0003
#> 106 2022-06-03 13:30:00  88570300 147.97 144.46 146.90 145.38  145.1783
#> 107 2022-06-06 13:30:00  71598400 148.57 144.90 147.03 146.14  145.9373
#> 108 2022-06-07 13:30:00  67808200 149.00 144.10 144.35 148.71  148.5037
#> 109 2022-06-08 13:30:00  53950200 149.87 147.46 148.58 147.96  147.7548
#> 110 2022-06-09 13:30:00  69473000 147.95 142.53 147.08 142.64  142.4421
#> 111 2022-06-10 13:30:00  91437900 140.76 137.06 140.28 137.13  136.9398
#> 112 2022-06-13 13:30:00 122207100 135.20 131.44 132.87 131.88  131.6971
#> 113 2022-06-14 13:30:00  84784300 133.89 131.48 133.13 132.76  132.5758
#> 114 2022-06-15 13:30:00  91533000 137.34 132.16 134.29 135.43  135.2421
#> 115 2022-06-16 13:30:00 108123900 132.39 129.04 132.08 130.06  129.8796
#> 116 2022-06-17 13:30:00 134520300 133.08 129.81 130.07 131.56  131.3775
#> 117 2022-06-21 13:30:00  81000500 137.06 133.32 133.42 135.87  135.6815
#> 118 2022-06-22 13:30:00  73409200 137.76 133.91 134.79 135.35  135.1623
#> 119 2022-06-23 13:30:00  72433800 138.59 135.63 136.82 138.27  138.0782
#> 120 2022-06-24 13:30:00  89116800 141.91 139.77 139.90 141.66  141.4635
#> 121 2022-06-27 13:30:00  70207900 143.49 140.97 142.70 141.66  141.4635
#> 122 2022-06-28 13:30:00  67083400 143.42 137.32 142.13 137.44  137.2494
#> 123 2022-06-29 13:30:00  66242400 140.67 136.67 137.46 139.23  139.0369
#> 124 2022-06-30 13:30:00  98964500 138.37 133.77 137.25 136.72  136.5303
#> 125 2022-07-01 13:30:00  71051600 139.04 135.66 136.04 138.93  138.7373
#> 126 2022-07-05 13:30:00  73353800 141.61 136.93 137.77 141.56  141.3636
#> 127 2022-07-06 13:30:00  74064300 144.12 141.08 141.35 142.92  142.7218
#> 128 2022-07-07 13:30:00  66253700 146.55 143.28 143.29 146.35  146.1470
#> 129 2022-07-08 13:30:00  64547800 147.55 145.00 145.26 147.04  146.8360
#> 130 2022-07-11 13:30:00  63141600 146.64 143.78 145.67 144.87  144.6690
#> 131 2022-07-12 13:30:00  77588800 148.45 145.05 145.76 145.86  145.6577
#> 132 2022-07-13 13:30:00  71185600 146.45 142.12 142.99 145.49  145.2882
#> 133 2022-07-14 13:30:00  78140700 148.95 143.25 144.08 148.47  148.2641
#> 134 2022-07-15 13:30:00  76259900 150.86 148.20 149.78 150.17  149.9617
#> 135 2022-07-18 13:30:00  81420900 151.57 146.70 150.74 147.07  146.8660
#> 136 2022-07-19 13:30:00  82982400 151.23 146.91 147.92 151.00  150.7905
#> 137 2022-07-20 13:30:00  64823400 153.72 150.37 151.12 153.04  152.8277
#> 138 2022-07-21 13:30:00  65086600 155.57 151.94 154.50 155.35  155.1345
#> 139 2022-07-22 13:30:00  66675400 156.28 153.41 155.39 154.09  153.8763
#> 140 2022-07-25 13:30:00  53623900 155.04 152.28 154.01 152.95  152.7378
#> 141 2022-07-26 13:30:00  55138700 153.09 150.80 152.26 151.60  151.3897
#> 142 2022-07-27 13:30:00  78620700 157.33 152.16 152.58 156.79  156.5725
#>  [ reached 'max' / getOption("max.print") -- omitted 28 rows ]
aapl$get_history(start = '2022-07-01', interval = '1d')
#>                   date    volume   high    low   open  close adj_close
#> 1  2022-07-01 13:30:00  71051600 139.04 135.66 136.04 138.93  138.7373
#> 2  2022-07-05 13:30:00  73353800 141.61 136.93 137.77 141.56  141.3636
#> 3  2022-07-06 13:30:00  74064300 144.12 141.08 141.35 142.92  142.7218
#> 4  2022-07-07 13:30:00  66253700 146.55 143.28 143.29 146.35  146.1470
#> 5  2022-07-08 13:30:00  64547800 147.55 145.00 145.26 147.04  146.8360
#> 6  2022-07-11 13:30:00  63141600 146.64 143.78 145.67 144.87  144.6690
#> 7  2022-07-12 13:30:00  77588800 148.45 145.05 145.76 145.86  145.6577
#> 8  2022-07-13 13:30:00  71185600 146.45 142.12 142.99 145.49  145.2882
#> 9  2022-07-14 13:30:00  78140700 148.95 143.25 144.08 148.47  148.2641
#> 10 2022-07-15 13:30:00  76259900 150.86 148.20 149.78 150.17  149.9617
#> 11 2022-07-18 13:30:00  81420900 151.57 146.70 150.74 147.07  146.8660
#> 12 2022-07-19 13:30:00  82982400 151.23 146.91 147.92 151.00  150.7905
#> 13 2022-07-20 13:30:00  64823400 153.72 150.37 151.12 153.04  152.8277
#> 14 2022-07-21 13:30:00  65086600 155.57 151.94 154.50 155.35  155.1345
#> 15 2022-07-22 13:30:00  66675400 156.28 153.41 155.39 154.09  153.8763
#> 16 2022-07-25 13:30:00  53623900 155.04 152.28 154.01 152.95  152.7378
#> 17 2022-07-26 13:30:00  55138700 153.09 150.80 152.26 151.60  151.3897
#> 18 2022-07-27 13:30:00  78620700 157.33 152.16 152.58 156.79  156.5725
#> 19 2022-07-28 13:30:00  81378700 157.64 154.41 156.98 157.35  157.1317
#> 20 2022-07-29 13:30:00 101786900 163.63 159.50 161.24 162.51  162.2846
#> 21 2022-08-01 13:30:00  67829400 163.59 160.89 161.01 161.51  161.2860
#> 22 2022-08-02 13:30:00  59907000 162.41 159.63 160.10 160.01  159.7880
#> 23 2022-08-03 13:30:00  82507500 166.59 160.75 160.84 166.13  165.8996
#> 24 2022-08-04 13:30:00  55474100 167.19 164.43 166.01 165.81  165.5800
#> 25 2022-08-05 13:30:00  56697000 165.85 163.00 163.21 165.35  165.3500
#> 26 2022-08-08 13:30:00  60276900 167.81 164.20 166.37 164.87  164.8700
#> 27 2022-08-09 13:30:00  63135500 165.82 163.25 164.02 164.92  164.9200
#> 28 2022-08-10 13:30:00  70170500 169.34 166.90 167.68 169.24  169.2400
#> 29 2022-08-11 13:30:00  57149200 170.99 168.19 170.06 168.49  168.4900
#> 30 2022-08-12 13:30:00  68039400 172.17 169.40 169.82 172.10  172.1000
#> 31 2022-08-15 13:30:00  54091700 173.39 171.35 171.52 173.19  173.1900
#> 32 2022-08-16 13:30:00  56377100 173.71 171.66 172.78 173.03  173.0300
#> 33 2022-08-17 13:30:00  79542000 176.15 172.57 172.77 174.55  174.5500
#> 34 2022-08-18 13:30:00  62290100 174.90 173.12 173.75 174.15  174.1500
#> 35 2022-08-19 13:30:00  70346300 173.74 171.31 173.03 171.52  171.5200
#> 36 2022-08-22 13:30:00  69026800 169.86 167.14 169.69 167.57  167.5700
#> 37 2022-08-23 13:30:00  54147100 168.71 166.65 167.08 167.23  167.2300
#> 38 2022-08-24 13:30:00  53841500 168.11 166.25 167.32 167.53  167.5300
#> 39 2022-08-25 13:30:00  51218200 170.14 168.35 168.78 170.03  170.0300
#> 40 2022-08-26 13:30:00  78961000 171.05 163.56 170.57 163.62  163.6200
#> 41 2022-08-29 13:30:00  73314000 162.90 159.82 161.15 161.38  161.3800
#> 42 2022-08-30 13:30:00  77906200 162.56 157.72 162.13 158.91  158.9100
#> 43 2022-08-31 13:30:00  87991100 160.58 157.14 160.31 157.22  157.2200
#> 44 2022-09-01 13:30:00  74229900 158.42 154.67 156.64 157.96  157.9600
#> 45 2022-09-02 13:30:00  76905200 160.36 154.97 159.75 155.81  155.8100
#> 46 2022-09-06 13:30:00  73604800 157.09 153.69 156.47 154.53  154.5300
aapl$get_history(start = '2022-07-01', end = '2022-07-14', interval = '1d')
#>                  date   volume   high    low   open  close adj_close
#> 1 2022-07-01 13:30:00 71051600 139.04 135.66 136.04 138.93  138.7373
#> 2 2022-07-05 13:30:00 73353800 141.61 136.93 137.77 141.56  141.3636
#> 3 2022-07-06 13:30:00 74064300 144.12 141.08 141.35 142.92  142.7218
#> 4 2022-07-07 13:30:00 66253700 146.55 143.28 143.29 146.35  146.1470
#> 5 2022-07-08 13:30:00 64547800 147.55 145.00 145.26 147.04  146.8360
#> 6 2022-07-11 13:30:00 63141600 146.64 143.78 145.67 144.87  144.6690
#> 7 2022-07-12 13:30:00 77588800 148.45 145.05 145.76 145.86  145.6577
#> 8 2022-07-13 13:30:00 71185600 146.45 142.12 142.99 145.49  145.2882
aapl$get_history(period = '1mo', interval = '1d')
#>                   date   volume   high    low   open  close adj_close
#> 1  2022-08-08 13:30:00 60276900 167.81 164.20 166.37 164.87    164.87
#> 2  2022-08-09 13:30:00 63135500 165.82 163.25 164.02 164.92    164.92
#> 3  2022-08-10 13:30:00 70170500 169.34 166.90 167.68 169.24    169.24
#> 4  2022-08-11 13:30:00 57149200 170.99 168.19 170.06 168.49    168.49
#> 5  2022-08-12 13:30:00 68039400 172.17 169.40 169.82 172.10    172.10
#> 6  2022-08-15 13:30:00 54091700 173.39 171.35 171.52 173.19    173.19
#> 7  2022-08-16 13:30:00 56377100 173.71 171.66 172.78 173.03    173.03
#> 8  2022-08-17 13:30:00 79542000 176.15 172.57 172.77 174.55    174.55
#> 9  2022-08-18 13:30:00 62290100 174.90 173.12 173.75 174.15    174.15
#> 10 2022-08-19 13:30:00 70346300 173.74 171.31 173.03 171.52    171.52
#> 11 2022-08-22 13:30:00 69026800 169.86 167.14 169.69 167.57    167.57
#> 12 2022-08-23 13:30:00 54147100 168.71 166.65 167.08 167.23    167.23
#> 13 2022-08-24 13:30:00 53841500 168.11 166.25 167.32 167.53    167.53
#> 14 2022-08-25 13:30:00 51218200 170.14 168.35 168.78 170.03    170.03
#> 15 2022-08-26 13:30:00 78961000 171.05 163.56 170.57 163.62    163.62
#> 16 2022-08-29 13:30:00 73314000 162.90 159.82 161.15 161.38    161.38
#> 17 2022-08-30 13:30:00 77906200 162.56 157.72 162.13 158.91    158.91
#> 18 2022-08-31 13:30:00 87991100 160.58 157.14 160.31 157.22    157.22
#> 19 2022-09-01 13:30:00 74229900 158.42 154.67 156.64 157.96    157.96
#> 20 2022-09-02 13:30:00 76905200 160.36 154.97 159.75 155.81    155.81
#> 21 2022-09-06 13:30:00 73604800 157.09 153.69 156.47 154.53    154.53